### 2023-11-09 #6043 @editorialbot
[PRE REVIEW]: Cellpy – an open-source library for processing and analysis of battery testing data
**Submitting author:** @jepegit (<a href="http://orcid.org/0000-0001-7662-4707">Jan Petter Maehlen</a>)
**Repository:** https://github.com/jepegit/cellpy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0post4
