+++ dev_environment.yml
@@ -1,14 +1,12 @@
 name: cellpy_dev
 channels:
     - conda-forge
-    - defaults
 dependencies:
-    - python >=3.9
+    - python=3.11
     - pip
     - numpy
     - openpyxl
     - xlrd
-    - matplotlib
     - cryptography
     - click
     - scipy
@@ -19,10 +17,7 @@ dependencies:
     - PyGithub
     - cookiecutter
     - tqdm
-    - ipython
-    - ipywidgets
-    - lmfit
-    - ipykernel
+    - dateparser
     - hdf5
     - pytables
     - python-box
@@ -30,20 +25,20 @@ dependencies:
     - sqlalchemy
     - python-dotenv
     - fabric
+    - lmfit
+    - matplotlib
+    - plotly
+    - ipython
+    - ipywidgets
+    - ipykernel
     - rich
-    - dateparser
     - coverage
     - pytest
     - pytest-timeout
     - pytest-benchmark
-    - dateparser
     - nox
     - black
     - plotly
-    - holoviz
-    - jupyter
-    - jupyterlab
-    - ipywidgets
     - seaborn
     - sphinx
     - nbsphinx
