### 2023-12-05 0660ca6
Update pytest_posix.yml
