### 2023-12-05 0ec1f27
a couple of future warnings from pandas appeared (in cellreader.py) - add TODOs there
