### 2023-12-08 f72e712
add try-except block for ica post-processing step and add if-clause (suggested by Vajee).
