### 2023-12-08 0e352ef
start refactoring concatenator
