### 2023-12-09 c8d5913
fix bug in batch update (nom-cap) and refactor get_cycle_numbers in cellreader (added a method also for getting the rates).
