+++ cellpy/readers/cellreader.py
@@ -373,6 +373,15 @@ class CellpyCell:
     def nom_cap(self, c):
         self.data.nom_cap = self._dump_cellpy_unit(c, "nominal_capacity")
 
+    @property
+    def nominal_capacity(self):
+        """returns the nominal capacity"""
+        return self.data.nom_cap
+
+    @nominal_capacity.setter
+    def nominal_capacity(self, c):
+        self.data.nom_cap = self._dump_cellpy_unit(c, "nominal_capacity")
+
     def _dump_cellpy_unit(self, value, parameter):
         """Parse for unit, update cellpy_units class, and return magnitude."""
         if isinstance(value, numbers.Number):
@@ -4560,32 +4569,76 @@ class CellpyCell:
             no_cycles = np.amax(steptable[self.headers_step_table.cycle])
         return no_cycles
 
+    def get_rates(self, steptable=None, agg="first", direction=None):
+        """
+        Get the rates in the test (only valid for constant current).
+        Args:
+            steptable: provide custom steptable (if None, the steptable from the cellpydata object will be used).
+            agg (str): perform an aggregation if more than one step of charge or discharge is found
+                (e.g. "mean", "first", "max"). For example, if agg='mean', the average rate for each cycle
+                will be returned. Set to None if you want to keep all the rates.
+            direction (str or list of str): only select rates for this direction (e.g. "charge" or "discharge").
+
+        Returns:
+            pandas.DataFrame with cycle, type, and rate_avr (i.e. C-rate) columns.
+        """
+
+        if steptable is None:
+            steptable = self.data.steps
+        rates = steptable[
+            [
+                self.headers_step_table.cycle,
+                self.headers_step_table.type,
+                self.headers_step_table.rate_avr,
+            ]
+        ].dropna()
+
+        if agg:
+            rates = rates.groupby(
+                [self.headers_step_table.cycle, self.headers_step_table.type]
+            ).agg(agg).reset_index()
+
+        if direction is not None:
+            if not isinstance(direction, (list, tuple)):
+                direction = [direction]
+            rates = rates.loc[rates[self.headers_step_table.type].isin(direction), :]
+
+        return rates
+
     def get_cycle_numbers(
         self,
         steptable=None,
         rate=None,
         rate_on=None,
         rate_std=None,
-        rate_column=None,
+        rate_agg="first",
         inverse=False,
     ):
-        """Get a list containing all the cycle numbers in the test.
+        """Get a array containing the cycle numbers in the test.
 
         Parameters:
+            steptable (pandas.DataFrame): the step-table to use (if None, the step-table
+                from the cellpydata object will be used).
             rate (float): the rate to filter on. Remark that it should be given
                 as a float, i.e. you will have to convert from C-rate to
                 the actual numeric value. For example, use rate=0.05 if you want
                 to filter on cycles that has a C/20 rate.
-            rate_on (str): only select cycles if based on the rate of this step-type (e.g. on="charge").
+            rate_on (str): only select cycles if based on the rate of this step-type (e.g. on="discharge").
             rate_std (float): allow for this inaccuracy in C-rate when selecting cycles
-            rate_column (str): column header name of the rate column,
+            rate_agg (str): perform an aggregation on rate if more than one step of charge or discharge is found
+                (e.g. "mean", "first", "max"). For example, if agg='mean', the average rate for each cycle
+                will be returned. Set to None if you want to keep all the rates.
             inverse (bool): select steps that does not have the given C-rate.
 
         Returns:
             numpy.ndarray of cycle numbers.
         """
 
[…]
