### 2023-12-09 6cd94a9
add possibility to filter on C-rate for all collectors
