+++ cellpy/utils/helpers.py
@@ -96,36 +96,16 @@ def _make_average_legacy(
 
 def _make_average(
     frames,
-    # keys=None,
     columns=None,
     skip_st_dev_for_equivalent_cycle_index=True,
-    # key_index_bounds=None,
 ):
-    # if key_index_bounds is None:
-    #     key_index_bounds = [1, -2]
     hdr_norm_cycle = hdr_summary["normalized_cycle_index"]
-    hdr_cum_charge = hdr_summary["cumulated_charge_capacity"]
-    cell_id = ""
     not_a_number = np.NaN
     new_frames = []
 
     if columns is None:
         columns = frames[0].columns
 
-    # if keys is not None:
-    #     if isinstance(keys, (list, tuple)):
-    #         cell_id = list(
-    #             set(
-    #                 [
-    #                     "_".join(
-    #                         k.split("_")[key_index_bounds[0]:key_index_bounds[1]]
-    #                     )
-    #                     for k in keys
-    #                 ]
-    #             )
-    #         )[0]
-    #     elif isinstance(keys, str):
-    #         cell_id = keys
     new_frame = pd.concat(frames, axis=1)
     normalized_cycle_index_frame = pd.DataFrame(index=new_frame.index)
     for col in columns:
@@ -147,9 +127,6 @@ def _make_average(
             if number_of_cols > 1:
                 avg_frame = (
                     new_frame[col].agg(["mean", "std"], axis=1)
-                    # .rename(
-                    #     columns={"mean": "value"}
-                    # )
                 )
             else:
                 avg_frame = pd.DataFrame(
@@ -173,7 +150,6 @@ def _make_average(
             cols.remove(n)
     cols.extend(new_cols)
     final_frame = final_frame.reindex(columns=cols)
-    # return final_frame, cell_id
     return final_frame
 
 
@@ -312,7 +288,7 @@ def _remove_outliers_from_summary(s, filter_vals, freeze_indexes=None):
 def remove_outliers_from_summary_on_window(
     s, window_size=3, cut=0.1, iterations=1, col_name=None, freeze_indexes=None
 ):
-    """Removes outliers based based on neighbours"""
+    """Removes outliers based on neighbours"""
     if col_name is None:
         col = hdr_summary["charge_capacity"]
 
@@ -709,7 +685,6 @@ def concatenate_summaries(
     inverse=False,
     inverted=False,
     key_index_bounds=None,
-    mode="collector",
 ) -> pd.DataFrame:
     """Merge all summaries in a batch into a gigantic summary data frame.
 
@@ -734,16 +709,12 @@ def concatenate_summaries(
         inverted (bool): select cycles that do not have the steps filtered by given C-rate.
         key_index_bounds (list): used when creating a common label for the cells by splitting and combining from
             key_index_bound[0] to key_index_bound[1].
-        mode (str): set to something else than "collector" to get the "old" behaviour of this function.
 
     Returns:
         ``pandas.DataFrame``
     """
 
-    if mode != "collector":
-        warnings.warn(
-            "This helper function will be removed shortly", category=DeprecationWarning
-        )
+    warnings.warn("This helper function is not maintained anymore", category=DeprecationWarning)
 
     if key_index_bounds is None:
         key_index_bounds = [1, -2]
@@ -763,7 +734,6 @@ def concatenate_summaries(
         group_nest.append(b.pages.group.to_list())
 
     default_columns = [hdr_summary["charge_capacity_gravimetric"]]
-    reserved_cell_label_names = ["FC"]
     hdr_norm_cycle = hdr_summary["normalized_cycle_index"]
 
     if columns is None:
@@ -817,7 +787,6 @@ def concatenate_summaries(
[…]
