### 2023-12-11 d92eb9b
fix bug in b.plot for bokeh (version > 3.0.0; new keyword arguments)
