### 2023-12-15 5ca1d16
new batch summary-plotter
