### 2023-12-15 9d6f8ba
create a new convenience function in the batch module that loads journal file if it exists, if not, go through the normal process of making journal and updating (loading files)
