### 2023-12-16 d0f2ea5
fix bug in generate_summary_frame_for_plotting (name of index in pages depends on how it was mad) and update a bit the str representations are for journal and dbreader
