### 2023-12-16 60a476c
batch mark and drop
