### 2023-12-16 9cd8e28
batch mark and drop
