### 2023-12-16 8579ec3
start to test implementing seaborn in collectors (experimental mode)
