+++ cellpy/utils/collectors.py
@@ -13,9 +13,6 @@ from itertools import count
 from multiprocessing import Process
 
 import pandas as pd
-import plotly.express as px
-import plotly.io as pio
-import plotly.graph_objects as go
 import numpy as np
 
 import cellpy
@@ -25,12 +22,31 @@ from cellpy.utils.helpers import concat_summaries
 from cellpy.utils.plotutils import plot_concatenated
 from cellpy.utils import ica
 
-DEFAULT_CYCLES = [1, 10, 20]
+supported_backends = []
 
+try:
+    import plotly.express as px
+    import plotly.io as pio
+    import plotly.graph_objects as go
+    supported_backends.append("plotly")
+except ImportError:
+    print("WARNING: plotly not installed")
+
+try:
+    import seaborn as sns
+    supported_backends.append("seaborn")
+except ImportError:
+    print("WARNING: seaborn not installed")
+
+
+DEFAULT_CYCLES = [1, 10, 20]
 CELLPY_MINIMUM_VERSION = "1.0.0"
 PLOTLY_BASE_TEMPLATE = "seaborn"
 IMAGE_TO_FILE_TIMEOUT = 30
-SUPPORTED_BACKENDS = ["plotly"]
+
+if not supported_backends:
+    print("WARNING: no supported backends found")
+    print("WARNING: install plotly or seaborn to enable plotting")
 
 px_template_all_axis_shown = dict(
     xaxis=dict(
@@ -113,6 +129,7 @@ class BatchCollector:
         elevated_plotter_arguments=None,
         data_collector_arguments: dict = None,
         plotter_arguments: dict = None,
+        experimental: bool = False,
         **kwargs,
     ):
         """Update both the collected data and the plot(s).
@@ -136,7 +153,11 @@ class BatchCollector:
         self.data_collector = data_collector
         self.plotter = plotter
         self.nick = nick
-        self.backend = backend
+        self.experimental = experimental
+        if backend != "plotly" and not self.experimental:
+            print("WARNING: only plotly is supported at the moment")
+        else:
+            self.backend = backend
         self.collector_name = collector_name or "base"
 
         # Arguments given as default arguments in the subclass have "low" priority (below elevated arguments at least):
@@ -303,6 +324,7 @@ class BatchCollector:
     def render(self):
         self.figure = self.plotter(
             self.data,
+            backend=self.backend,
             journal=self.b.journal,
             units=self.units,
             **self.plotter_arguments,
@@ -478,7 +500,11 @@ class BatchCollector:
             logging.info(f"updating figure with {kwargs}")
             self._update_arguments(plotter_arguments=kwargs)
             self.render()
-        return self.figure
+
+        if self.backend == "seaborn":
+            return self.figure.figure
+        else:
+            return self.figure
 
     def preprocess_data_for_csv(self):
         logging.debug(f"the data layout {self.csv_layout} is not supported yet!")
@@ -527,7 +553,7 @@ class BatchCollector:
             self._image_exporter_plotly(filename_svg)
             self.figure.write_json(filename_json)
             print(f"saved plotly json file: {filename_json}")
-        elif self.backend == "matplotlib":
+        elif self.backend == "seaborn":
             print(f"TODO: implement saving {filename_png}")
             print(f"TODO: implement saving {filename_svg}")
             print(f"TODO: implement saving {filename_json}")
@@ -622,7 +648,6 @@ class BatchSummaryCollector(BatchCollector):
                 key_index_bound[0] to key_index_bound[1].
 
         Elevated plotter args:
-            backend (str): backend used (defaults to Bokeh)
             points (bool): plot points if True
             line (bool): plot line if True
             width: width of plot
@@ -654,7 +679,6 @@ class BatchSummaryCollector(BatchCollector):
 
         elevated_plotter_arguments = {
[…]
