### 2023-12-17 3deb32f
allow for custom jupyter executable for cellpy serve
