### 2023-12-17 d5bca31
allow for custom jupyter executable for cellpy serve
