### 2023-12-17 39922f4
fix a couple of bugs and short-comings for the collectors
