+++ cellpy/utils/collectors.py
@@ -155,7 +155,9 @@ class BatchCollector:
         self.nick = nick
         self.experimental = experimental
         if backend != "plotly" and not self.experimental:
+            print(f"{backend=}")
             print("WARNING: only plotly is supported at the moment")
+            self.backend = "plotly"
         else:
             self.backend = backend
         self.collector_name = collector_name or "base"
@@ -610,7 +612,7 @@ class BatchSummaryCollector(BatchCollector):
         inverse=None,
         inverted: bool = None,
         key_index_bounds=None,
-        backend: str = None,
+        backend: str = "plotly",
         title: str = None,
         points: bool = None,
         line: bool = None,
@@ -625,6 +627,9 @@ class BatchSummaryCollector(BatchCollector):
         **kwargs,
     ):
         """Collects and shows summaries.
+
+        Arguments:
+            backend (str): what plotting backend to use (currently only 'plotly' is supported)
 
         Elevated data collector args:
             max_cycle (int): drop all cycles above this value.
@@ -798,7 +803,7 @@ class BatchICACollector(BatchCollector):
         rate_agg=None,
         inverse=False,
         label_mapper=None,
-        backend=None,
+        backend="plotly",
         cycles_to_plot=None,
         width=None,
         palette=None,
@@ -826,7 +831,6 @@ class BatchICACollector(BatchCollector):
             label_mapper=label_mapper,
         )
         elevated_plotter_arguments = dict(
-            backend=backend,
             cycles_to_plot=cycles_to_plot,
             width=width,
             palette=palette,
@@ -842,6 +846,7 @@ class BatchICACollector(BatchCollector):
             plotter=ica_plotter,
             data_collector=ica_collector,
             collector_name="ica",
+            backend=backend,
             elevated_data_collector_arguments=elevated_data_collector_arguments,
             elevated_plotter_arguments=elevated_plotter_arguments,
             *args,
@@ -888,7 +893,7 @@ class BatchCyclesCollector(BatchCollector):
         rate_agg=None,
         inverse=False,
         label_mapper=None,
-        backend=None,
+        backend="plotly",
         cycles_to_plot=None,
         width=None,
         palette=None,
@@ -905,6 +910,7 @@ class BatchCyclesCollector(BatchCollector):
         Args:
             b:
             plot_type (str): either 'fig_pr_cell' or 'fig_pr_cycle'
+            backend (str): what plotting backend to use (currently only 'plotly' is supported)
             collector_type (str): how the curves are given
                 "back-and-forth" - standard back and forth; discharge
                     (or charge) reversed from where charge (or discharge) ends.
@@ -927,7 +933,6 @@ class BatchCyclesCollector(BatchCollector):
                 Remark! No check are performed to ensure that the new cell labels are unique.
 
         Elevated plotter args:
-            backend (str): backend used (defaults to Bokeh)
             cycles_to_plot (int): plot points if True
             width (float): width of plot
             legend_position (str): position of the legend
@@ -948,7 +953,6 @@ class BatchCyclesCollector(BatchCollector):
             inverse=inverse,
         )
         elevated_plotter_arguments = dict(
-            backend=backend,
             cycles_to_plot=cycles_to_plot,
             width=width,
             palette=palette,
@@ -969,6 +973,7 @@ class BatchCyclesCollector(BatchCollector):
             b,
             plotter=cycles_plotter,
             data_collector=cycles_collector,
+            backend=backend,
             collector_name="cycles",
             elevated_data_collector_arguments=elevated_data_collector_arguments,
             elevated_plotter_arguments=elevated_plotter_arguments,
@@ -1257,38 +1262,38 @@ def sequence_plotter(
 
     Args:
[…]
