### 2023-12-17 cd69336
fix a couple of bugs and short-comings for batch and batch plotter
