### 2023-12-18 7f382e8
automatically remove duplicate rows in data from sql_h5
