### 2023-12-18 1343b06
add checks to CellpyCell objects (has_no_duplicates etc) and add to batch.report
