+++ cellpy/utils/batch.py
@@ -182,6 +182,17 @@ class Batch:
     def name(self):
         return self.experiment.journal.name
 
+    def _check_cell_data_points(self, cell_id):
+        error_code = 0
+        try:
+            c = self.experiment.data[cell_id]
+            if not c.has_no_partial_duplicates(subset="data_point"):
+                error_code = 1
+            return error_code
+        except Exception as e:
+            logging.debug(f"Exception ignored: {e}")
+            return None
+
     def _check_cell_raw(self, cell_id):
         try:
             c = self.experiment.data[cell_id]
@@ -297,28 +308,58 @@ class Batch:
             else:
                 self.pages = self.pages.drop(cell_label)
 
-    def report(self, stylize=True):
+    def report(self, stylize=True, grouped=False, check=False):
         """Create a report on all the cells in the batch object.
 
         Remark! To perform a reporting, cellpy needs to access all the data (and it might take some time).
+
+        Args:
+            stylize (bool): apply some styling to the report (default True).
+            grouped (bool): add information based on the group cell belongs to (default False).
+            check (bool): check if the data seems to be without errors (0 = no errors, 1 = partial duplicates)
+                (default False).
 
         Returns:
             ``pandas.DataFrame``
         """
         pages = self.experiment.journal.pages
-        pages = pages[COLUMNS_SELECTED_FOR_VIEW].copy()
-        # pages["empty"] = pages.index.map(self._check_cell_empty)
-        pages["raw_rows"] = pages.index.map(self._check_cell_raw)
-        pages["steps_rows"] = pages.index.map(self._check_cell_steps)
-        pages["summary_rows"] = pages.index.map(self._check_cell_summary)
-        pages["last_cycle"] = pages.index.map(self._check_cell_cycles)
-        pages["average_capacity"] = pages.index.map(self._check_cell_avg_cap)
-        pages["max_capacity"] = pages.index.map(self._check_cell_max_cap)
-        pages["min_capacity"] = pages.index.map(self._check_cell_min_cap)
-        pages["std_capacity"] = pages.index.map(self._check_cell_std_cap)
+        r_pages = self.experiment.journal.pages[COLUMNS_SELECTED_FOR_VIEW].copy()
+        r_pages["empty"] = pages.index.map(self._check_cell_empty)
+        r_pages["raw_rows"] = pages.index.map(self._check_cell_raw)
+        r_pages["steps_rows"] = pages.index.map(self._check_cell_steps)
+        r_pages["summary_rows"] = pages.index.map(self._check_cell_summary)
+        r_pages["last_cycle"] = pages.index.map(self._check_cell_cycles)
+        r_pages["average_capacity"] = pages.index.map(self._check_cell_avg_cap)
+        r_pages["max_capacity"] = pages.index.map(self._check_cell_max_cap)
+        r_pages["min_capacity"] = pages.index.map(self._check_cell_min_cap)
+        r_pages["std_capacity"] = pages.index.map(self._check_cell_std_cap)
+        if check:
+            r_pages["error_code"] = pages.index.map(self._check_cell_data_points)
 
-        avg_last_cycle = pages.last_cycle.mean()
-        avg_max_capacity = pages.max_capacity.mean()
+        avg_last_cycle = r_pages.last_cycle.mean()
+        avg_max_capacity = r_pages.max_capacity.mean()
+
+        logging.info(f"average last cycle: {avg_last_cycle:.2f}")
+        logging.info(f"average max capacity: {avg_max_capacity:.2f}")
+
+        if grouped:
+            # TODO: currently does not use cumulative values - consider implementing this
+            r_pages["group"] = pages.group
+            r_pages["group_avg_last_cycle"] = r_pages.group.map(
+                r_pages.groupby("group").last_cycle.mean()
+            )
[…]
