+++ cellpy/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.1a1"
+__version__ = "1.0.1a2"
