### 2023-12-19 7836ca6
fix bug in str
