+++ cellpy/readers/instruments/neware_xlsx.py
@@ -1,0 +1,325 @@
+"""neware xlsx exported data"""
+import datetime
+import logging
+import pathlib
+import sys
+import warnings
+
+import pandas as pd
+from dateutil.parser import parse
+
+from cellpy import prms
+from cellpy.exceptions import WrongFileVersion
+from cellpy.parameters.internal_settings import HeaderDict, get_headers_normal
+from cellpy.readers.core import Data, FileID
+from cellpy.readers.instruments.base import BaseLoader
+from pathlib import Path
+
+DEBUG_MODE = prms.Reader.diagnostics  # not used
+ALLOW_MULTI_TEST_FILE = prms._allow_multi_test_file  # not used
+DATE_TIME_FORMAT = prms._date_time_format  # not used
+
+normal_headers_renaming_dict = {
+    "test_id_txt": "Test_ID",
+    "data_point_txt": "Data_Point",
+    "datetime_txt": "Date_Time",
+    "test_time_txt": "Test_Time",
+    "step_time_txt": "Step_Time",
+    "cycle_index_txt": "Cycle_Index",
+    "step_index_txt": "Step_Index",
+    "current_txt": "Current",
+    "voltage_txt": "Voltage",
+    "power_txt": "Power",
+    "charge_capacity_txt": "Charge_Capacity",
+    "discharge_capacity_txt": "Discharge_Capacity",
+    "charge_energy_txt": "Charge_Energy",
+    "discharge_energy_txt": "Discharge_Energy",
+    "internal_resistance_txt": "Internal_Resistance",
+    "ref_voltage_txt": "Aux_Voltage_1",
+}
+
+
+def to_datetime(n):
+    if isinstance(n, int):
+        n = str(n)
+    ms_component = n[-7:]
+    date_time_component = n[:-7]
+    temp = f"{date_time_component}.{ms_component}"
+    datetime_object = datetime.datetime.fromtimestamp(float(temp))
+    time_in_str = datetime_object.strftime(DATE_TIME_FORMAT)
+    return time_in_str
+
+
+class DataLoader(BaseLoader):
+    """Class for loading arbin-data from MS SQL server."""
+
+    instrument_name = "arbin_sql_h5"
+    raw_ext = "h5"
+
+    def __init__(self, *args, **kwargs):
+        """initiates the ArbinSQLLoader class"""
+        self.arbin_headers_normal = (
+            self.get_headers_normal()
+        )  # the column headers defined by Arbin
+        self.cellpy_headers_normal = (
+            get_headers_normal()
+        )  # the column headers defined by cellpy
+
+    @staticmethod
+    def get_headers_normal():
+        """Defines the so-called normal column headings for Arbin SQL Server h5 export"""
+        # covered by cellpy at the moment
+        return get_headers_normal()
+
+    @staticmethod
+    def get_headers_aux(df):
+        """Defines the so-called auxiliary table column headings for Arbin SQL Server h5 export"""
+        headers = HeaderDict()
+        for col in df.columns:
+            if col.startswith("Aux_"):
+                ncol = col.replace("/", "_")
+                ncol = "".join(ncol.split("(")[0])
+                headers[col] = ncol.lower()
+
+        return headers
+
+    @staticmethod
+    def get_raw_units():
+        raw_units = dict()
+        raw_units["current"] = "A"
+        raw_units["charge"] = "Ah"
+        raw_units["mass"] = "g"
+        raw_units["voltage"] = "V"
+        raw_units["nominal_capacity"] = "Ah/g"
+        return raw_units
+
+    @staticmethod
+    def get_raw_limits():
+        """returns a dictionary with resolution limits"""
+        raw_limits = dict()
[…]
