+++ cellpy/readers/instruments/neware_xlsx.py
@@ -1,4 +1,5 @@
 """neware xlsx exported data"""
+from dataclasses import dataclass
 import datetime
 import logging
 import pathlib
@@ -13,30 +14,17 @@ from cellpy.exceptions import WrongFileVersion
 from cellpy.parameters.internal_settings import HeaderDict, get_headers_normal
 from cellpy.readers.core import Data, FileID
 from cellpy.readers.instruments.base import BaseLoader
+from cellpy.readers.instruments.processors import post_processors as pp
 from pathlib import Path
 
 DEBUG_MODE = prms.Reader.diagnostics  # not used
 ALLOW_MULTI_TEST_FILE = prms._allow_multi_test_file  # not used
 DATE_TIME_FORMAT = prms._date_time_format  # not used
 
-normal_headers_renaming_dict = {
-    "test_id_txt": "Test_ID",
-    "data_point_txt": "Data_Point",
-    "datetime_txt": "Date_Time",
-    "test_time_txt": "Test_Time",
-    "step_time_txt": "Step_Time",
-    "cycle_index_txt": "Cycle_Index",
-    "step_index_txt": "Step_Index",
-    "current_txt": "Current",
-    "voltage_txt": "Voltage",
-    "power_txt": "Power",
-    "charge_capacity_txt": "Charge_Capacity",
-    "discharge_capacity_txt": "Discharge_Capacity",
-    "charge_energy_txt": "Charge_Energy",
-    "discharge_energy_txt": "Discharge_Energy",
-    "internal_resistance_txt": "Internal_Resistance",
-    "ref_voltage_txt": "Aux_Voltage_1",
-}
+
+@dataclass
+class ModelParameters:
+    states: dict
 
 
 def to_datetime(n):
@@ -53,43 +41,78 @@ def to_datetime(n):
 class DataLoader(BaseLoader):
     """Class for loading arbin-data from MS SQL server."""
 
-    instrument_name = "arbin_sql_h5"
-    raw_ext = "h5"
+    instrument_name = "neware_xlsx"
+    raw_ext = "xlsx"
 
     def __init__(self, *args, **kwargs):
-        """initiates the ArbinSQLLoader class"""
-        self.arbin_headers_normal = (
-            self.get_headers_normal()
-        )  # the column headers defined by Arbin
+        """initiates the neware xlsx reader class"""
+        self.raw_headers_normal = None  # the column headers defined by Neware
         self.cellpy_headers_normal = (
             get_headers_normal()
         )  # the column headers defined by cellpy
+        self.normal_headers_renaming_dict = None  # renaming dict for the headers
+        self.config_params = None  # configuration parameters
 
     @staticmethod
     def get_headers_normal():
-        """Defines the so-called normal column headings for Arbin SQL Server h5 export"""
+        """Defines the so-called normal column headings for export"""
         # covered by cellpy at the moment
         return get_headers_normal()
 
     @staticmethod
     def get_headers_aux(df):
-        """Defines the so-called auxiliary table column headings for Arbin SQL Server h5 export"""
+        """Defines the so-called auxiliary table column headings for Neware xlsx export
+
+        Warning: not properly implemented yet
+
+        """
         headers = HeaderDict()
         for col in df.columns:
             if col.startswith("Aux_"):
                 ncol = col.replace("/", "_")
                 ncol = "".join(ncol.split("(")[0])
                 headers[col] = ncol.lower()
+            if col.startswith("Step Type"):
+                headers[col] = "step_type"
 
         return headers
 
-    @staticmethod
-    def get_raw_units():
+    def get_normal_headers_renaming_dict(self):
+        units = self.get_raw_units()
+        unit_current = units["current"]
+        unit_voltage = units["voltage"]
+        unit_capacity = units["charge"]
[…]
