### 2023-12-22 #6043 @mbarzegary
Hi @MaximevdHeijden 👋
Considering your field of expertise, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06043/joss.06043/10.21105.joss.06043.pdf
Software: https://github.com/jepegit/cellpy
