### 2023-12-23 f7e1815
working neware_xlsx.py (further development only after refactoring it out as a plugin)
