+++ cellpy/readers/instruments/neware_xlsx.py
@@ -12,7 +12,7 @@ from dateutil.parser import parse
 from cellpy import prms
 from cellpy.exceptions import WrongFileVersion
 from cellpy.parameters.internal_settings import HeaderDict, get_headers_normal
-from cellpy.readers.core import Data, FileID
+from cellpy.readers.core import Data, FileID, Q
 from cellpy.readers.instruments.base import BaseLoader
 from cellpy.readers.instruments.processors import post_processors as pp
 from pathlib import Path
@@ -147,29 +147,42 @@ class DataLoader(BaseLoader):
         # self.name = name
         # self.copy_to_temporary()
         logging.critical("Experimental loader for neware xlsx files")
-        data_df = self._query()
+        data_df, meta = self._query()
+
         self.normal_headers_renaming_dict = self.get_normal_headers_renaming_dict()
         data = Data()
 
-        self.config_params = ModelParameters(
-            {
-                "column_name": "step_type",  # changed from "Step Type" in the _query method
-                "charge_keys": ["CC Chg"],
-                "discharge_keys": ["CC DChg"],
-                "rest_keys": ["Rest"],
-            }
-        )
+        if kwargs.get("config_params"):
+            self.config_params = kwargs.get("config_params")
+        else:
+            self.config_params = ModelParameters(
+                {
+                    "column_name": "step_type",  # changed from "Step Type" in the _query method
+                    "charge_keys": ["CC Chg"],
+                    "discharge_keys": ["CC DChg"],
+                    "rest_keys": ["Rest"],
+                }
+            )
 
         # some metadata is available in the info_df part of the h5 file
         data.loaded_from = self.name
-        # data.channel_index = data_dfs["info_df"]["IV_Ch_ID"].iloc[0]
-        # # data.test_ID = data_dfs["info_df"]["Test_ID"].iloc[0]
-        # data.test_name = self.name.name
-        # data.creator = None
-        # data.schedule_file_name = data_dfs["info_df"]["Schedule_File_Name"].iloc[0]
-        # data.start_datetime = data_dfs["info_df"]["First_Start_DateTime"].iloc[0]
-        # data.mass = data_dfs["info_df"]["SpecificMASS"].iloc[0]
-        # data.nom_cap = data_dfs["info_df"]["SpecificCapacity"].iloc[0]
+
+        if meta.keys():
+            data.test_name = meta.get(name, "unknown")
+            data.custom_info = meta
+
+            if start_time := meta.get("start_time"):
+                # TODO: implement conversion to datetime
+                logging.debug(f"start_time: {start_time}")
+
+            if end_time := meta.get("end_time"):
+                # TODO: implement conversion to datetime
+                logging.debug(f"end-time: {end_time}")
+            if units := meta.get("units"):
+                # TODO: implement handling of units and updating get_raw_units
+                logging.debug(f"units: {units}")
+
+            # TODO: implement handling of the rest of the meta data
 
         # Generating a FileID project:
         self.generate_fid()
@@ -182,11 +195,7 @@ class DataLoader(BaseLoader):
         )  # creating an empty frame - loading summary is not implemented yet
         data = self._post_process(data)
         data = self.identify_last_data_point(data)
-        dev = False
-        if dev:
-            print(data.raw.columns)
-            print(data.raw.head())
-            sys.exit(0)
+
         return data
 
     def _post_process(self, data):
@@ -285,10 +294,13 @@ class DataLoader(BaseLoader):
         file_name = self.temp_file_path
         file_format = file_name.suffix[1:]
         file_name = pathlib.Path(file_name)
[…]
