### 2023-12-25 cfcc034
add option for selecting specifics in get_cap
