+++ cellpy/readers/cellreader.py
@@ -365,6 +365,24 @@ class CellpyCell:
         self.data.mass = self._dump_cellpy_unit(m, "mass")
 
     @property
+    def active_mass(self):
+        """returns the active mass (same as mass)"""
+        return self.data.mass
+
+    @active_mass.setter
+    def active_mass(self, m):
+        self.data.mass = self._dump_cellpy_unit(m, "mass")
+
+    @property
+    def tot_mass(self):
+        """returns the total mass"""
+        return self.data.tot_mass
+
+    @tot_mass.setter
+    def tot_mass(self, m):
+        self.data.tot_mass = self._dump_cellpy_unit(m, "mass")
+
+    @property
     def active_electrode_area(self):
         """returns the area"""
         return self.data.active_electrode_area
@@ -395,11 +413,13 @@ class CellpyCell:
         """Parse for unit, update cellpy_units class, and return magnitude."""
         if isinstance(value, numbers.Number):
             return value
-        logging.debug(f"Parsing {parameter} ({value})")
+        logging.critical(f"Parsing {parameter} ({value})")
+
         try:
             c = Q(value)
             c_unit = c.units
             self.cellpy_units[parameter] = f"{c_unit}"
+            logging.critical(f"Updated your cellpy_units['{parameter}'] to '{c_unit}'")
             c = c.magnitude
         except ValueError:
             logging.debug(f"Could not parse {value}")
@@ -1152,7 +1172,8 @@ class CellpyCell:
         Args:
             file_names (list of raw-file names): uses CellpyCell.file_names if
                 None. If the list contains more than one file name, then the
-                runs will be merged together.
+                runs will be merged together. Remark! the order of the files in
+                the list is important.
             pre_processor_hook (callable): function that will be applied to the data within the loader.
             post_processor_hook (callable): function that will be applied to the
                 cellpy.Dataset object after initial loading.
@@ -1201,7 +1222,11 @@ class CellpyCell:
         max_raw_files_to_merge = 20
         if len(self.file_names) > max_raw_files_to_merge:
             logging.debug("ERROR? Too many files to merge")
-            raise ValueError("Too many files to merge - could be a p2-p3 zip thing")
+            raise ValueError(
+                f"Too many files to merge (max allowed is {max_raw_files_to_merge})"
+                f" - could be a bug in the code (please report if you know you have"
+                f" less than {max_raw_files_to_merge} files)"
+            )
 
         logging.debug("start iterating through file(s)")
         recalc = kwargs.pop("recalc", True)
@@ -4140,6 +4165,11 @@ class CellpyCell:
         inter_cycle_shift=True,
         interpolate_along_cap=False,
         capacity_then_voltage=False,
+        mode="gravimetric",
+        mass=None,
+        area=None,
+        volume=None,
+        cycle_mode=None,
         **kwargs,
     ):
         """Gets the capacity for the run.
@@ -4180,6 +4210,14 @@ class CellpyCell:
                 of along the voltage axis. Defaults to False.
             capacity_then_voltage (bool): return capacity and voltage instead of
                 voltage and capacity. Defaults to False.
+            mode (string): 'gravimetric', 'areal', 'volumetric' or 'absolute'. Defaults
+                to 'gravimetric'.
+            mass (float): mass of active material (in set cellpy unit, typically mg).
+            area (float): area of electrode (in set cellpy units, typically cm2).
+            volume (float): volume of electrode (in set cellpy units, typically cm3).
+            cycle_mode (string): if 'anode' the first step is assumed to be the discharge,
+                else charge (defaults to CellpyCell.cycle_mode).
+            **kwargs: sent to get_ccap and get_dcap.
 
         Returns:
[…]
