### 2023-12-25 6a86bdd
add option for selecting specifics in get_cap
