### 2023-12-25 2e7f7ab
allow empty charge or discharge data for get_cap
