+++ cellpy/readers/cellreader.py
@@ -4228,10 +4228,22 @@ class CellpyCell:
         # TODO: allow for fixing the interpolation range (so that it is possible
         #   to run the function on several cells and have a common x-axis)
 
-        # TODO: allow for missing steps (i.e. if you have a cell with only charge and no discharge in the first cycle)
-
         # if cycle is not given, then this function should
         # iterate through cycles
+
+        def last(df):
+            try:
+                return df.iat[-1]
+            except TypeError:
+                return 0.0
+
+        def first(df):
+            try:
+                return df.iat[0]
+            except TypeError:
+                return 0.0
+
+        experimental = True
 
         cycle_mode = cycle_mode or self.cycle_mode
 
@@ -4259,6 +4271,9 @@ class CellpyCell:
                 insert_nan = True
             else:
                 insert_nan = False
+
+        if insert_nan:
+            _nan = pd.DataFrame({"capacity": [np.nan], "voltage": [np.nan]})
 
         converter_kwargs = dict()
         if mass is not None:
@@ -4306,6 +4321,15 @@ class CellpyCell:
                     as_frame=False,
                     **kwargs,
                 )
+
+            except NullData as e:
+                error = True
+                logging.critical(e)
+                if not ignore_errors:
+                    logging.debug("breaking out of loop")
+                    break
+
+            try:
                 dc, dv = self.get_dcap(
                     current_cycle,
                     converter=specific_converter,
@@ -4320,7 +4344,7 @@ class CellpyCell:
                     logging.debug("breaking out of loop")
                     break
 
-            if not error:
+            if not error or experimental:
                 if cc.empty:
                     logging.debug("get_ccap returns empty cc Series")
 
@@ -4347,32 +4371,36 @@ class CellpyCell:
                     _last_step_v = dv
 
                 if method == "back-and-forth":
-                    # _last = np.amax(_first_step_c)
-                    _last = _first_step_c.iat[-1]
-                    # should change amax to last point
+                    _last = last(_first_step_c)
                     _first = None
                     _new_first = None
+
                     if not inter_cycle_shift:
                         prev_end = 0.0
+
                     if _last_step_c is not None:
                         _last_step_c = _last - _last_step_c + prev_end
+
                     else:
-                        logging.debug("no last charge step found")
+                        logging.critical("no last charge step found")
+
                     if _first_step_c is not None:
-                        _first = _first_step_c.iat[0]
+                        _first = first(_first_step_c)
                         _first_step_c += prev_end
-                        _new_first = _first_step_c.iat[0]
+                        _new_first = first(_first_step_c)
+
                     else:
-                        logging.debug("probably empty (_first_step_c is None)")
+                        logging.critical("probably empty (_first_step_c is None)")
                     # logging.debug(f"current shifts used: prev_end = {prev_end}")
                     # logging.debug(f"shifting start from {_first} to "
                     #                   f"{_new_first}")
 
                     # prev_end = np.amin(_last_step_c)
-                    prev_end = _last_step_c.iat[-1]
+                    prev_end = last(_last_step_c)
+
                 elif method == "forth":
                     # _last = np.amax(_first_step_c)
-                    _last = _first_step_c.iat[-1]
+                    _last = last(_first_step_c)
                     if _last_step_c is not None:
                         _last_step_c += _last + prev_end
                     else:
@@ -4382,9 +4410,8 @@ class CellpyCell:
                     else:
                         logging.debug("no first charge step found")
 
-                    # prev_end = np.amax(_last_step_c)
                     if inter_cycle_shift:
-                        prev_end = _last_step_c.iat[-1]
+                        prev_end = last(_last_step_c)
                     else:
                         prev_end = 0.0
 
@@ -4405,48 +4432,53 @@ class CellpyCell:
[…]
