### 2023-12-25 4b105d5
allow empty charge or discharge data for get_cap
