### 2023-12-25 7ccb8a7
allow empty charge or discharge data for get_cap
