### 2023-12-26 c5e8a80
add alias cycles for cycle in get_cap
