### 2023-12-26 d4a1346
update HISTORY.rst and improve total_time_at_voltage_level (renamed)
