### 2023-12-26 32bb6e8
update HISTORY.rst and improve total_time_at_voltage_level (renamed)
