+++ tests/test_neware.py
@@ -31,6 +31,3 @@ def test_get_neware_from_h5(neware_cell):
     assert len(neware_cell.data.summary) == 4
     t = neware_cell.total_time_at_voltage_level()
     print(f"total time at low voltage: {t} seconds")
-
-
-
