+++ HISTORY.rst
@@ -19,6 +19,7 @@ History
 * Added experimental instrument reader for neware xlsx files
 * Added try-except block for ica post-processing step and add if-clause (suggested by Vajee)
 * Fixed several smaller bugs and improved some of the functionality (most notably in `c.get_cap`)
+* [under progress]: check how cycle_mode and specific is handled and possibly improve
 
 
 1.0.0 (2023)
