### 2023-12-26 5f24578
make example env file when running cellpy setup
