### 2023-12-27 6d23869
make example env file when running cellpy setup
