### 2023-12-27 c300d41
make example env file when running cellpy setup
