### 2023-12-27 51b7e96
make example env file when running cellpy setup
