### 2023-12-27 8330f24
suppress warnings when displaying Data in jupyter
