### 2023-12-27 f01b6b8
suppress warnings when displaying Data in jupyter
