### 2023-12-27 9cabec7
fix bug in cycle_mode (saves as a list in cellpyfile, but needs to (still) be a string in the cellpycell methods.
