### 2024-01-08 #6043 @mbarzegary
Hi @sarahalamdari, @nicolet5, @TomTranter, @mdmurbach 👋
Considering that you have already published your software in JOSS, would any of you be willing to review this submission? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06043/joss.06043/10.21105.joss.06043.pdf
Software: https://github.com/jepegit/cellpy
