### 2024-01-12 #6043 @mbarzegary
Hi @ma-sadeghi, @yangbai90 and @WardLT 👋
Considering that you have already reviewed for JOSS, would any of you be willing to review this submission? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06043/joss.06043/10.21105.joss.06043.pdf
Software: https://github.com/jepegit/cellpy
