### 2024-01-20 2bf60d4
add new template ("single")
