+++ cellpy/parameters/prms.py
@@ -380,6 +380,7 @@ _standard_template_uri = "https://github.com/jepegit/cellpy_cookies.git"
 _registered_templates = {
     "standard": (_standard_template_uri, "standard"),  # (repository, name-of-folder)
     "ife": (_standard_template_uri, "ife"),
+    "single": (_standard_template_uri, "single"),
 }
 
 # used as global variables
