### 2024-01-26 jepegit/cellpy#273 @MaximevdHeijden
JOSS Review - Installation
Dear authors,

The installations manual can be clarified and polished, especially because there are three locations with information on the installation:
-	/latest/, very simple incomplete description
-	/latest/main_description/installation.html, simple but effective installation but you miss the command to install the dependencies using “conda env create -f environment.yml”
-	/latest/examples_and_tutorials/basics.html#getting-started, more complete installation manual with one for beginners and one for standard users. The one for standard users can be improved by adding the installation of the dependencies and jupyter notebook (or ideally add that to the dependencies file), to make sure the installation of your software is as smooth as possible.

Moreover, I tried the installation on Windows and using Linux software on Windows:
Windows: I tried installing cellpy using a virtual environment and the base, using a virtual environment it resulted in the following error: “CondaSSLError: OpenSSL appears to be unavailable on this machine”, which I could not install with conda. Whereas when installing cellpy using the base it gave the mentioned obdc error, which was not solved by running “conda install -c conda-forge pyodbc”.

Linux: The installation works well using a virtual environment; however the installation of jupyter notebook in the dependency list is missing.

I suggest you create only one page for the installation with multiple options: for beginners/standard and for using windows/linux.

This issue is related to https://github.com/openjournals/joss-reviews/issues/6236
