### 2024-01-30 jepegit/cellpy#274 @jepegit
Dear @MaximevdHeijden, Thanks for the valuable suggestions. We will start working on improving the examples and documentation now. Will also will try to make the installation process a bit simpler. I guess we will need a week or so (because of lot of other things happening at work at the moment).
