### 2024-02-01 6ff3c8d
initial suggestion for re-structuring the documentation
