+++ docs/fundamentals/basic_usage.rst
@@ -1,0 +1,15 @@
+============
+Basic usage
+============
+
+Simple usage as a python library
+==============
+
+
+For more details refer to examples
+
+
+Convenience methods and tools
+==============
+
+For more details refer to examples
