+++ docs/fundamentals/data_structure.rst
@@ -1,0 +1,9 @@
+============
+Data structure
+============
+
+
+
+
+
+
