+++ docs/contributing/index.rst
@@ -1,0 +1,9 @@
+Contributing guide
+----------------
+
+.. toctree::
+    :maxdepth: 2
+
+    contributing.rst
+
+    
