+++ docs/fundamentals/fundamentals.rst
@@ -1,0 +1,5 @@
+============
+The fundamentals of ``cellpy``
+============
+
+This is how ``cellpy`` is build up and works
