+++ docs/getting_started/configuration.rst
@@ -1,0 +1,3 @@
+============
+Configuring ``cellpy``
+============
