+++ docs/examples/index.rst
@@ -1,0 +1,9 @@
+Examples & tutorials
+----------------
+
+.. toctree::
+    :maxdepth: 2
+
+    reading_data.rst
+
+    
