+++ docs/fundamentals/index.rst
@@ -1,0 +1,12 @@
+Fundamentals & usage
+----------------
+
+.. toctree::
+    :maxdepth: 2
+
+    fundamentals.rst
+    file_formats.rst
+    data_structure.rst
+    basic_usage.rst
+
+    
