+++ docs/other/index.rst
@@ -1,0 +1,9 @@
+Other
+----------------
+
+.. toctree::
+    :maxdepth: 2
+
+    other.rst
+
+    
