+++ docs/getting_started/index.rst
@@ -1,0 +1,10 @@
+Getting started
+----------------
+
+.. toctree::
+    :maxdepth: 2
+
+    installation.rst
+    checkup.rst
+    configuration.rst
+
