+++ github_actions_environment.yml
@@ -33,8 +33,8 @@ dependencies:
     - pytest
     - pytest-timeout
     - pytest-benchmark
+    - sqlalchemy-access
     - pip:
-        - sqlalchemy-access
         - jinja2_time
 
 
