+++ cellpy/readers/cellreader.py
@@ -14,6 +14,7 @@ Example:
 import collections
 import copy
 import csv
+import functools
 import itertools
 import logging
 import numbers
@@ -5390,6 +5391,45 @@ class CellpyCell:
         last_items = raw[d_txt].isin(steps)
         return last_items
 
+    # TODO: @jepe - this method might be valuable for users and could be made public:
+    def _select_without(self, raw, exclude_types=None, exclude_steps=None):
+        # this function gives a set of indexes pointing to the last
+        # datapoints for each cycle in the dataset given that the steps
+        # is not in the exclude list (either by type of step or by step number)
+
+        steps = self.data.steps
+        d_txt = self.headers_normal.data_point_txt
+
+        if exclude_types is None and exclude_steps is None:
+            last_data_points = (
+                steps.loc[:, ["cycle", "point_last"]]
+                .groupby("cycle")
+                .last()
+                .values.ravel()
+            )
+            return raw[d_txt].isin(last_data_points)
+
+        if not isinstance(exclude_types, (list, tuple)):
+            exclude_types = [exclude_types]
+
+        if not isinstance(exclude_steps, (list, tuple)):
+            exclude_steps = [exclude_steps]
+
+        q = None
+        for exclude_type in exclude_types:
+            _q = ~steps["type"].str.startswith(exclude_type)
+            q = _q if q is None else q & _q
+
+        if exclude_steps:
+            _q = ~steps["step"].isin(exclude_steps)
+            q = _q if q is None else q & _q
+
+        last_data_points = (
+            steps.loc[q, ["cycle", "point_last"]].groupby("cycle").last().values.ravel()
+        )
+        last_items = raw[d_txt].isin(last_data_points)
+        return last_items
+
     # ----------making-summary------------------------------------------------------
     def make_summary(
         self,
@@ -5399,11 +5439,19 @@ class CellpyCell:
         use_cellpy_stat_file=None,
         # all_tests=True,
         ensure_step_table=True,
+        remove_duplicates=True,
         # add_c_rate=True,
         normalization_cycles=None,
         nom_cap=None,
         nom_cap_specifics=None,
         # from_cycle=None,
+        experimental=False,
+        create_copy=False,
+        exclude_types=None,
+        exclude_steps=None,
+        selector_type=None,
+        selector=None,
+        **kwargs,
     ):
         """Convenience function that makes a summary of the cycling data."""
         # TODO: @jepe - include option for omitting steps
@@ -5426,7 +5474,7 @@ class CellpyCell:
         try:
             test = self.data
         except NoDataFound:
-            logging.info(f"Empty test {test})")
+            logging.info(f"Empty test (no data found)")
             return
 
         if isinstance(test.loaded_from, (list, tuple)):
@@ -5436,6 +5484,30 @@ class CellpyCell:
             txt += str(test.loaded_from)
 
         logging.debug(txt)
+        if experimental:
+            data = self._make_summary_experimental(
+                find_ir=find_ir,
+                find_end_voltage=find_end_voltage,
+                use_cellpy_stat_file=use_cellpy_stat_file,
+                ensure_step_table=ensure_step_table,
+                remove_duplicates=remove_duplicates,
+                normalization_cycles=normalization_cycles,
+                nom_cap=nom_cap,
+                nom_cap_specifics=nom_cap_specifics,
+                create_copy=create_copy,
+                exclude_types=exclude_types,
+                exclude_steps=exclude_steps,
+                selector_type=selector_type,
+                selector=selector,
+                **kwargs,
+            )
+            if create_copy:
+                other = copy.deepcopy(self)
+                other.data = data
+                return other
+            else:
+                # TODO: check if anything is using this feature (returning self), if not, remove it.
[…]
