### 2024-02-08 f03b1b3
start clean up of plotting tools and implement overrides for step table
