+++ cellpy/utils/plotutils.py
@@ -13,6 +13,8 @@ import warnings
 from io import StringIO
 from pathlib import Path
 
+import matplotlib.pyplot as plt
+
 from cellpy.parameters.internal_settings import (
     get_headers_journal,
     get_headers_normal,
@@ -21,23 +23,7 @@ from cellpy.parameters.internal_settings import (
 )
 from cellpy.utils import helpers
 
-try:
-    import matplotlib.pyplot as plt
-
-    plt_available = True
-except ImportError:
-    plt_available = False
-
-try:
-    import holoviews as hv
-    from holoviews import opts
-    from holoviews.plotting.links import RangeToolLink
-
-    hv_available = True
-except ImportError:
-    hv_available = False
-
-bokeh_available = importlib.util.find_spec("bokeh") is not None
+plotly_available = importlib.util.find_spec("plotly") is not None
 
 # logger = logging.getLogger(__name__)
 logging.captureWarnings(True)
@@ -152,311 +138,6 @@ hdr_steps = get_headers_step_table()
 hdr_journal = get_headers_journal()
 
 
-def _hv_bokeh_available():
-    warnings.warn(
-        "This utility function will be removed shortly", category=DeprecationWarning
-    )
-    if not hv_available:
-        print("You need holoviews. But I cannot load it. Aborting...")
-        return False
-    if not bokeh_available:
-        print("You need Bokeh. But I cannot find it. Aborting...")
-        return False
-    return True
-
-
-def find_column(columns, label=None, end="cycle_index"):
-    """find columns based on how the column-name ends.
-
-    Args:
-        columns: pandas columns
-        label: if not provided, generate, if provided, return as is
-        end: the string to use for searching
-
-    Returns:
-        column header, label
-    """
-
-    warnings.warn(
-        "This utility function will be removed shortly", category=DeprecationWarning
-    )
-    # TODO @jepe: refactor and use col names directly from HeadersNormal instead
-    hdr = None
-    lab = None
-    for col in columns:
-        if col.endswith(end):
-            hdr = col
-            if label is None:
-                lab = col.replace("_", " ")
-            else:
-                lab = label
-            break
-    return hdr, lab
-
-
-def plot_concatenated(
-    dataframe,
-    title="",
-    x=None,
-    y=None,
-    err=None,
-    xlabel=None,
-    ylabel=None,
-    points=True,
-    line=True,
-    errors=True,
-    hover=True,
-    width=800,
-    height=300,
-    journal=None,
-    file_id_level=0,
-    hdr_level=None,
-    axis=1,
-    mean_end="_mean",
-    std_end="_std",
-    cycle_end="cycle_index",
-    legend_title="cell-type",
-    marker_size=None,
-    cmap="default_colors",
-    spread=False,
-    extension="bokeh",
-    edges=False,
-    keys=None,
-    simple=False,
-    **kwargs,
-):
-    """Create a holoviews plot of the concatenated summary.
-
-    This function is still under development. Feel free to contribute.
-
-    Args:
-        dataframe: the concatenated summary
-        title (str): title of the plot (defaults to empty)
-        x: colum-name for the x variable (not implemented yet)
-        y: colum-name for the y variable (not implemented yet)
-        err: colum-name for the std variable (not implemented yet)
-        xlabel: label for x-axis
-        ylabel: label for y-axis
-        points (bool): plot points if True
-        line (bool): plot line if True
-        errors (bool): plot errors if True
-        hover (bool): add hover tool if True
-        width: width of plot
-        height: height of plot
-        journal: `batch.journal` object
[…]
