### 2024-02-08 bfd00f4
start clean up of plotting tools and implement overrides for step table
