### 2024-02-08 1c4d9bf
start clean up of plotting tools and implement overrides for step table
