+++ cellpy/utils/collectors.py
@@ -1179,6 +1179,8 @@ def ica_collector(
     return collected_curves
 
 
+# plotter functions (consider moving to plotutils)
+
 def remove_markers(trace):
     trace.update(marker=None, mode="lines")
     return trace
