### 2024-02-09 835a083
add plotly option for cycle_info_plot
