### 2024-02-09 6f5139c
add plotly option for cycle_info_plot
