### 2024-02-09 82cedcf
add plotly option for cycle_info_plot
