+++ cellpy/utils/plotutils.py
@@ -199,21 +199,23 @@ def raw_plot(
 ):
     # TODO: missing doc-string
 
+    raw = cell.data.raw.copy()
+
     if y is None:
-        y, y_label = ("voltage", "Voltage (V)")
+        y, y_label = ("voltage", f"Voltage ({cell.data.raw_units.voltage})")
     if x is None:
         x, x_label = ("test_time_hrs", "Time (hours)")
 
     if title is None:
         title = f"{cell.cell_name}"
 
-    raw = cell.data.raw
     if x == "test_time_hrs":
         raw["test_time_hrs"] = raw[hdr_raw["test_time_txt"]] / 3600
 
     if plotly_available and interactive:
         import plotly.express as px
 
+        title = f"<b>{title}</b>"
         if x_label or y_label:
             labels = {}
             if x_label:
@@ -240,7 +242,10 @@ def cycle_info_plot(
     cell,
     cycle,
     get_axes=False,
-    interactive=False,
+    interactive=True,
+    t_unit="hours",
+    v_unit="V",
+    i_unit="mA",
     **kwargs,
 ):
     """Show raw data together with step and cycle information.
@@ -250,28 +255,71 @@ def cycle_info_plot(
         cycle (int or list or tuple): cycle(s) to select (must be int for matplotlib)
         get_axes (bool): return axes (for matplotlib) or figure (for plotly)
         interactive (bool): use interactive plotting (if available)
+        t_unit (str): unit for x-axis (default: "hours")
+        v_unit (str): unit for y-axis (default: "V")
+        i_unit (str): unit for current (default: "mA")
         **kwargs: parameters specific to plotting backend.
 
     Returns:
         ``matplotlib.axes`` or None
     """
+    t_scaler = cell.unit_scaler_from_raw(t_unit, "time")
+    v_scaler = cell.unit_scaler_from_raw(v_unit, "voltage")
+    i_scaler = cell.unit_scaler_from_raw(i_unit, "current")
+
     if plotly_available and interactive:
-        fig = _cycle_info_plot_plotly(cell, cycle, get_axes, **kwargs)
+        fig = _cycle_info_plot_plotly(
+            cell,
+            cycle,
+            get_axes,
+            t_scaler,
+            t_unit,
+            v_scaler,
+            v_unit,
+            i_scaler,
+            i_unit,
+            **kwargs,
+        )
         if get_axes:
             return fig
         return fig
 
-    axes = _cycle_info_plot_matplotlib(cell, cycle, get_axes, **kwargs)
+    axes = _cycle_info_plot_matplotlib(
+        cell,
+        cycle,
+        get_axes,
+        t_scaler,
+        t_unit,
+        v_scaler,
+        v_unit,
+        i_scaler,
+        i_unit,
+        **kwargs,
+    )
+
     if get_axes:
         return axes
 
 
-def _cycle_info_plot_plotly(cell, cycle=None, get_axes=False, **kwargs):
+def _cycle_info_plot_plotly(
+    cell,
+    cycle,
+    get_axes,
+    t_scaler,
+    t_unit,
+    v_scaler,
+    v_unit,
+    i_scaler,
+    i_unit,
+    **kwargs,
+):
     import plotly.express as px
     import plotly.graph_objects as go
     import numpy as np
 
-    # TODO: implement options for units
+    if kwargs.get("xlim"):
+        logging.info("xlim is not supported for plotly")
+
     raw_hdr = get_headers_normal()
[…]
