### 2024-02-10 36464ce
add spread option to plotly summary collector plotter
