+++ cellpy/utils/collectors.py
@@ -24,6 +24,7 @@ from cellpy.utils import ica
 supported_backends = []
 
 try:
+    import plotly
     import plotly.express as px
     import plotly.io as pio
     import plotly.graph_objects as go
@@ -1228,6 +1229,75 @@ def legend_replacer(trace, df, group_legends=True):
         )
 
 
+def spread_plot(curves, plotly_arguments, **kwargs):
+    """Create a spread plot (error-bands instead of error-bars)."""
+
+    colors = plotly.colors.qualitative.Plotly
+    opacity = 0.2
+    color_list = []
+    for color in colors:
+        color_rgb = plotly.colors.hex_to_rgb(color)
+        color_rgb_main = f"rgb({color_rgb[0]}, {color_rgb[1]}, {color_rgb[2]})"
+        color_rgba_spread = (
+            f"rgba({color_rgb[0]}, {color_rgb[1]}, {color_rgb[2]}, {opacity})"
+        )
+        color_list.append((color_rgb_main, color_rgba_spread))
+
+    if plotly_arguments.get("markers"):
+        mode = "lines+markers"
+    else:
+        mode = "lines"
+
+    g = curves.groupby("cell")
+    fig = go.Figure()
+    for i, (cell, data) in enumerate(g):
+        color = color_list[i % len(color_list)]
+        fig.add_trace(
+            go.Scatter(
+                name=cell,
+                x=data["cycle"],
+                y=data["mean"],
+                mode=mode,
+                line=dict(color=color[0]),
+            )
+        )
+        fig.add_trace(
+            go.Scatter(
+                name="Upper Bound",
+                x=data["cycle"],
+                y=data["mean"] + data["std"],
+                mode="lines",
+                marker=dict(
+                    color=color[1],
+                ),
+                line=dict(width=0),
+                showlegend=False,
+            )
+        )
+        fig.add_trace(
+            go.Scatter(
+                name="Lower Bound",
+                x=data["cycle"],
+                y=data["mean"] - data["std"],
+                mode="lines",
+                marker=dict(
+                    color=color[1],
+                ),
+                line=dict(width=0),
+                fillcolor=color[1],
+                fill="tonexty",
+                showlegend=False,
+            )
+        )
+
+    if labels := plotly_arguments.get("labels"):
+        fig.update_layout(xaxis_title=labels.get("cycle", None))
+    if hover_mode := kwargs.pop("hovermode", None):
+        fig.update_layout(hovermode=hover_mode)
+
+    return fig
+
+
 def sequence_plotter(
     collected_curves: pd.DataFrame,
     x: str = "capacity",
@@ -1262,6 +1332,7 @@ def sequence_plotter(
     palette_range: tuple = None,
     height: float = None,
     width: float = None,
+    spread: bool = False,
     **kwargs,
 ) -> Any:
     """create a plot made up of sequences of data (voltage curves, dQ/dV, etc).
@@ -1304,6 +1375,7 @@ def sequence_plotter(
         cols (int): number of columns for layout.
         height (int): plot height.
         width (int): plot width.
+        spread (bool): plot error-bands instead of error-bars if True.
 
         **kwargs: sent to backend (if `backend == "plotly"`, it will be
             sent to `plotly.express` etc.)
@@ -1512,12 +1584,15 @@ def sequence_plotter(
             fig.update_layout(coloraxis_colorbar_title_text=color_bar_txt)
 
         elif method == "summary":
-            fig = px.line(
-                curves,
-                **plotly_arguments,
-                **kwargs,
-            )
-            if group_cells:
+            if spread:
+                fig = spread_plot(curves, plotly_arguments, **kwargs)
+            else:
+                fig = px.line(
+                    curves,
+                    **plotly_arguments,
+                    **kwargs,
[…]
