+++ cellpy/readers/cellreader.py
@@ -5434,15 +5434,9 @@ class CellpyCell:
 
     # TODO: @jepe - this method might be valuable for users and could be made
     #  public when it is fixed:
-    def _select_without(self, raw, exclude_types=None, exclude_steps=None):
-        # this function gives a set of indexes pointing to the last
-        # datapoints for each cycle in the dataset given that the steps
-        # are not in the exclude list (either by type of step or by step number)
-        # It turns out that this is not usable for the summary since the
-        # cycle consists of both charge and discharge steps. Need to use the step
-        # table and subtract the values of the steps that are not wanted.
-
+    def _select_without(self, exclude_types=None, exclude_steps=None, replace_nan=True):
         steps = self.data.steps
+        raw = self.data.raw.copy()
 
         # unravel the headers:
         d_n_txt = self.headers_normal.data_point_txt
@@ -5464,17 +5458,26 @@ class CellpyCell:
 
         _first = "_first"
         _last = "_last"
+        _delta_label = "_diff"
 
-        # TODO: implement also for energy and power (and probably others as well)
+        # TODO: implement also for energy and power (and probably others as well) - this will
+        #  require changing step-table to also include energy and power etc. If implementing
+        #  this, you should also include diff in the step-table. You should preferably also use this
+        #  opportunity to also make both the headers in the tables as well as the names used for
+        #  the headers more aligned (e.g. for header_normal.data_point_txt -> header_normal.point;
+        #  "cycle_index" -> "cycle")
+
+        last_data_points = (
+            steps.loc[:, [c_st_txt, d_st_txt + _last]]
+            .groupby(c_st_txt)
+            .last()
+            .values.ravel()
+        )
+        last_items = raw[d_n_txt].isin(last_data_points)
+        selected = raw[last_items]
 
         if exclude_types is None and exclude_steps is None:
-            last_data_points = (
-                steps.loc[:, [c_st_txt, d_st_txt + _last]]
-                .groupby(c_st_txt)
-                .last()
-                .values.ravel()
-            )
-            return raw[d_n_txt].isin(last_data_points)
+            return selected
 
         if not isinstance(exclude_types, (list, tuple)):
             exclude_types = [exclude_types]
@@ -5491,25 +5494,17 @@ class CellpyCell:
             _q = ~steps[t_st_txt].isin(exclude_steps)
             q = _q if q is None else q & _q
 
-        last_data_points = (
-            steps.loc[q, [c_st_txt, d_st_txt + _last]]
-            .groupby(c_st_txt)
-            .last()
-            .values.ravel()
-        )
-        last_items = raw[d_n_txt].isin(last_data_points)
-
-        # --------- fixing ---------------- (still a mess)
-        #  goal: to find all the deltas (diffs) and subtract them from the raw data
-
-        updated_raw = raw.copy()
-
-        _delta_label = "_diff"
         _delta_columns = [
-            "current",
-            "voltage",
-            "charge",
-            "discharge",
+            i_st_txt,
+            v_st_txt,
+            ch_st_txt,
+            dch_st_txt,
+        ]
+        _raw_columns = [
+            i_n_txt,
+            v_n_txt,
+            ch_n_txt,
+            dch_n_txt,
         ]
         _diff_columns = [f"{col}{_delta_label}" for col in _delta_columns]
 
@@ -5517,33 +5512,23 @@ class CellpyCell:
         delta_last = [f"{col}{_last}" for col in _delta_columns]
         delta_columns = delta_first + delta_last
 
-        delta = steps.loc[~q, [c_st_txt, s_st_txt, d_st_txt + _last, *delta_columns]].copy()
+        delta = steps.loc[~q, [c_st_txt, d_st_txt + _last, *delta_columns]].copy()
 
         for col in _delta_columns:
[…]
