### 2024-02-12 d60420a
fix bug in make_summary (old make_summary can be used by issuing the old=True)
