### 2024-02-12 3368d4e
fix bug in make_summary (old make_summary can be used by issuing the old=True)
