### 2024-02-13 175bb44
add more saving and loading options for collectors as well as allowing for looking for "group_labels" column in pages when creating group names
