### 2024-02-13 211e8be
add more selector_types for make_summary
