### 2024-02-13 32d0c2c
add some docstring to make-summary
