### 2024-02-13 57e8a73
add some docstring to make-summary
