### 2024-02-13 c691064
add some docstring to make-summary
