### 2024-02-14 b7393e9
custom cell name to cycle info plot
