### 2024-02-15 #6236 @jepegit
Hi @mbarzegary. We need to update the documentation and make the installation process easier when using conda. The last item is fixed, but we have still not fixed the documentation (due to sick-leave). It will take at least one more week.
