### 2024-02-15 c9123fe
small future proofing and adding summary plotting utility (v0)
