+++ cellpy/readers/instruments/configurations/maccor_txt_one.py
@@ -120,6 +120,6 @@ post_processors = {
     "rename_headers": True,
     "set_cycle_number_not_zero": True,
     "convert_date_time_to_datetime": True,
-    # "convert_step_time_to_timedelta": True,
-    # "convert_test_time_to_timedelta": True,
+    "convert_step_time_to_timedelta": True,
+    "convert_test_time_to_timedelta": True,
 }
