### 2024-02-15 40dd4bb
add working summary plotting utility (v0.5)
