+++ cellpy/utils/plotutils.py
@@ -188,7 +188,11 @@ def create_colormarkerlist(
 
 
 def create_col_info(c):
-    _cap_cols = ["charge_capacity", "discharge_capacity"]
+    """Create column information for summary plots."""
+
+    # TODO: add support for more column sets and individual columns
+    hdr = c.headers_summary
+    _cap_cols = [hdr.charge_capacity_raw, hdr.discharge_capacity_raw]
     _capacities_gravimetric = [col + "_gravimetric" for col in _cap_cols]
     _capacities_gravimetric_split = (
         _capacities_gravimetric
@@ -202,9 +206,9 @@ def create_col_info(c):
         + [col + "_non_cv" for col in _capacities_areal]
     )
 
-    x_columns = (["cycle_index", "data_point", "test_time", "date_time"],)
+    x_columns = ([hdr.cycle_index, hdr.data_point, hdr.test_time, hdr.datetime],)
     y_cols = dict(
-        voltages=["end_voltage_charge", "end_voltage_discharge"],
+        voltages=[hdr.end_voltage_charge, hdr.end_voltage_discharge],
         capacities_gravimetric=_capacities_gravimetric,
         capacities_areal=_capacities_areal,
         capacities_gravimetric_split_constant_voltage=_capacities_gravimetric_split,
@@ -214,11 +218,12 @@ def create_col_info(c):
 
 
 def create_label_dict(c):
+    hdr = c.headers_summary
     x_axis_labels = {
-        "cycle_index": "Cycle Number",
-        "data_point": "Point",
-        "test_time": f"Test Time ({c.cellpy_units.time})",
-        "data_time": "Date",
+        hdr.cycle_index: "Cycle Number",
+        hdr.data_point: "Point",
+        hdr.test_time: f"Test Time ({c.cellpy_units.time})",
+        hdr.datetime: "Date",
     }
 
     _cap_gravimetric_label = (
@@ -240,7 +245,7 @@ def create_label_dict(c):
 
 def summary_plot(
     c,
-    x: str = "cycle_index",
+    x: str = None,
     y: str = "capacities_gravimetric",
     height: int = 600,
     markers: bool = True,
@@ -249,14 +254,19 @@ def summary_plot(
     y_range: list = None,
     split: bool = False,
     interactive: bool = True,
+    share_y: bool = False,
+    rangeslider: bool = False,
+    **kwargs,
 ):
     """Create a summary plot. Currently only supports plotly.
 
 
     Args:
         c: cellpy object
-        x: x-axis column
-        y: y-axis column
+        x: x-axis column (default: cycle_index)
+        y: y-axis column or column set (predefined sets implemented are: "voltages",
+          "capacities_gravimetric", "capacities_areal", "capacities_gravimetric_split_constant_voltage",
+          "capacities_areal_split_constant_voltage")
         height: height of the plot
         markers: use markers
         title: title of the plot
@@ -264,10 +274,14 @@ def summary_plot(
         y_range: limits for y-axis
         split: split the plot
         interactive: use interactive plotting
+        rangeslider: add a range slider to the x-axis (only for plotly)
+        share_y (bool): share y-axis
+        **kwargs: additional parameters for the plotting backend
+
+    Returns:
+        plotly figure or None
 
     """
-
-    import pandas as pd
 
     if plotly_available and interactive:
         import plotly.express as px
@@ -280,14 +294,15 @@ def summary_plot(
     if title is None:
         title = f"Summary <b>{c.cell_name}</b>"
 
+    if x is None:
+        x = "cycle_index"
+
     x_columns, y_cols = create_col_info(c)
     x_axis_labels, y_axis_label = create_label_dict(c)
 
     # ------------------- main --------------------------------------------
     y_header = "value"
     color = "variable"
[…]
