### 2024-02-15 jepegit/cellpy#276 @ma-sadeghi
Fresh installation fails the health checks
The following was done on a new installation from PyPI on a fresh condo virtualenv:

```shell
❯ cellpy info --check
=================================== checking ===================================
[cellpy] * - Checking cellpy imports
[cellpy] -> succeeded!
[cellpy] * - Checking importing pyodbc
 reading prms
 - ODBC: pyodbc
 - SEARCH_FOR_ODBC_DRIVERS: True
 - use_subprocess: False
 - detect_subprocess_need: False
 - stated office version: 64bit
 checking system
 - running on posix
 - running on a mac
 - python version: 64bit
 - os version:
 checking existence of mdb-export
 - failed finding it
 looks like you are on a mac (driver set to
 /usr/local/lib/libmdbodbc.dylib)
 - but cannot find it!
f[cellpy] -> failed!!!!
[cellpy] * - Checking configuration (prm) file
[cellpy] ->/Users/amin/.cellpy_prms_amin.conf
[cellpy] File does not exist!
[cellpy] check raised an exception (expected str, bytes or os.PathLike object, not NoneType)
[cellpy] OH NO!!! You (or I) failed!
[cellpy] Failed 1 out of 3 checks.
================================================================================
```

- OS: macOS 11.7
- CPU: Intel
