### 2024-02-15 jepegit/cellpy#277 @ma-sadeghi
Drop support for Python 3.9 and begin supporting 3.12
This is a bit opinionated so I don't consider it as a show stopper, but I highly encourage you to consider it.

`cellpy` depends on NumPy, and they will drop support for Python 3.9 in a month ([Apr 4, 2024](https://numpy.org/neps/nep-0029-deprecation_policy.html#drop-schedule)).
