### 2024-02-15 jepegit/cellpy#278 @ma-sadeghi
Minor revisions for the JOSS manuscript
The paper is well-written, just needs a few minor edits. Here's the list of errors. The list was generated by ChatGPT, which I've manually reviewed and removed those that didn't seem reasonable.

- [x] **One of the main tools in battery research are battery cycling experiments.**
   - Correction: One of the main tools in battery research **is** battery cycling experiments.

- [x] **providing the tools to read different data formats,**
   - Suggestion for clarity: providing tools to read different data formats,

- [x] **Converting those into one common data format that also includes relevant battery-specific metadata, and**
   - Correction for consistency: meta-data should be either **metadata** or **meta-data** consistently throughout the document. Here, it's mixed with "meta data" later on.

- [x] **Several open-source libraries focus on battery test-data extraction.**
   - Consistency: test data should be either **test-data** or **test data** consistently throughout the document.

- [x] **Provides a structured interphase for collecting and processing battery test data and exports to text format.**
   - Correction: **interface** instead of "interphase".

- [x] **Accommodating not only data from diverse testers, but also thoughtfully embeding battery-specific metadata**
    - Correction: embedding.

- [x] **Cellpy is implemented in python**
    - Correction: **Python** should be capitalized as it's a proper noun.

- [x] **The cellpy-file format (usually stored in hdf5 format) contains all the data contained in the Data object together with additional relevant metadata,**
    - Consistency: **HDF5** should be in uppercase as it's an acronym for Hierarchical Data Format version 5. Also, decide on either "metadata" or "meta data" and use it consistently.
