### 2024-02-15 #6236 @ma-sadeghi
[WIP] List of issues to be addressed:

- [ ] jepegit/cellpy/issues/276
- [ ] jepegit/cellpy/issues/277 (optional, but recommended)
- [ ] jepegit/cellpy/issues/278
