### 2024-02-15 jepegit/cellpy#276 @ma-sadeghi
This could be a macOS-related issue, since I was able to install it on my Ubuntu machine. I also noticed that your CIs only run on Ubuntu and Windows.
