### 2024-02-16 jepegit/cellpy#276 @jepegit
Thanks! Yes, I found it a bit tricky to set up CIs, and in particularly for macOS. But its been some years since I last tried it, so well worth to try once more. Will do (see you opened another issue for that, so will use that for making a branch). Must admit that I also find it a bit fiddly to make tests for the cli tool and in particular things that are related to installation and setup. But from your print-out I see that the cellpy output is not very easy to understand (even I had to think a while, and I wrote it). (I think) the message tries to communicate that the computer is missing a library that it needs for parsing raw-data files from Arbin testers (old versions, .res format). I will double check and improve the feedback from cellpy.
