### 2024-02-16 b6ef539
update github actions and add for macos-latest and macos-14
