+++ .github/workflows/pytest_macos.yml
@@ -1,25 +1,28 @@
-name: Run pytest on win - NOT WORKING DUE TO MISSING ACCESSDATABASEENGINE
+name: Run pytest on linux
 on:
-#  push:
-#    branches:
-#      - "**"
-#  pull_request:
-#    branches: [ master ]
+  push:
+    branches:
+      - "**"
+  pull_request:
+    branches: [ master ]
   workflow_dispatch:
 
 jobs:
   Install-miniconda-and-run-pytest:
-    runs-on: windows-latest
+    runs-on: macos-latest
+    defaults:
+      run:
+        shell: bash -l {0}
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11"]
     steps:
-      - uses: actions/checkout@v4
-      - uses: conda-incubator/setup-miniconda@v3
+      - uses: actions/checkout@v3
+      - uses: conda-incubator/setup-miniconda@v2
         with:
           auto-update-conda: true
           activate-environment: cellpy_dev
-          environment-file: dev_environment.yml
+          environment-file: github_actions_environment.yml
           python-version: ${{ matrix.python-version }}
 
       - name: Conda info
@@ -27,6 +30,11 @@ jobs:
 
       - name: Conda list
         run: conda list
+
+      - name: Install mdbtools
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y mdbtools
 
       - name: pytest
         run: pytest
