+++ .github/workflows/pytest_macos_14.yml
@@ -1,15 +1,18 @@
-name: Run pytest on win - NOT WORKING DUE TO MISSING ACCESSDATABASEENGINE
+name: Run pytest on linux
 on:
-#  push:
-#    branches:
-#      - "**"
-#  pull_request:
-#    branches: [ master ]
+  push:
+    branches:
+      - "**"
+  pull_request:
+    branches: [ master ]
   workflow_dispatch:
 
 jobs:
   Install-miniconda-and-run-pytest:
-    runs-on: windows-latest
+    runs-on: macos-14
+    defaults:
+      run:
+        shell: bash -l {0}
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11"]
@@ -19,7 +22,7 @@ jobs:
         with:
           auto-update-conda: true
           activate-environment: cellpy_dev
-          environment-file: dev_environment.yml
+          environment-file: github_actions_environment.yml
           python-version: ${{ matrix.python-version }}
 
       - name: Conda info
@@ -27,6 +30,11 @@ jobs:
 
       - name: Conda list
         run: conda list
+
+      - name: Install mdbtools
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y mdbtools
 
       - name: pytest
         run: pytest
