### 2024-02-16 jepegit/cellpy#277 @jepegit
Adding support for Python 3.12 requires some work on the `OtherPath` object since it builds on `pathlib` that had a major upgrade going from Python 3.11 to Python 3.12. Should be implemented before v1.1.
