### 2024-02-16 0f742f2
update github actions and add for macos-latest and macos-14
