### 2024-02-16 adde72a
update github actions and add for macos-latest and macos-14 (3.9 fix)
