### 2024-02-16 096f253
update github actions and add for macos-latest and macos-14 (3.9 fix, better naming)
