+++ .github/workflows/pytest_posix.yml
@@ -1,4 +1,4 @@
-name: Run pytest on linux
+name: Run pytest on linux (conda)
 on:
   push:
     branches:
