### 2024-02-16 0bb3ca3
use brew for all macos actions and add miniconda version for the conda dependent macos actions
