+++ .github/workflows/pytest_macos_14.yml
@@ -21,6 +21,7 @@ jobs:
       - uses: conda-incubator/setup-miniconda@v3
         with:
           auto-update-conda: true
+          miniconda-version: latest
           activate-environment: cellpy_dev
           environment-file: github_actions_environment.yml
           python-version: ${{ matrix.python-version }}
@@ -33,8 +34,8 @@ jobs:
 
       - name: Install mdbtools
         run: |
-          sudo apt-get update
-          sudo apt-get install -y mdbtools
+          brew install mdbtools
+#          sudo apt-get install -y mdbtools
 
       - name: pytest
         run: pytest
