+++ docs/getting_started/installation.rst
@@ -3,8 +3,16 @@
 ============
 Installation
 ============
-``cellpy`` is available on Windows and Linux and can be installed using ``pip`` or ``conda``, or can be installed from source.
+``cellpy`` is available on (Windows :ref:`_Cellpy_Install_Windows`) and Linux (:ref:`Cellpy_Setup_Linux`)) and can be installed using ``pip``
+or ``conda``, or from source (:ref:`_Cellpy_Install_Sources`).
 
+For more details on dependencies, have a look (here :ref:`Cellpy_Dependencies`).
+
+For a tea-spoon explanation on how to install ``cellpy`` on windows, including the
+installation of python and setup of virtual environments, (see below
+:ref:`Cellpy_Setup_Teaspoon`).
+
+.. _Cellpy_Install_Windows:
 
 Windows
 ==============
@@ -12,7 +20,7 @@ Windows
 Conda
 -----
 
-Usually, the easiest way to install ``cellpy`` is by using conda:
+The easiest way to install ``cellpy`` is by using conda:
 
 .. code-block:: console
 
@@ -22,25 +30,58 @@ Usually, the easiest way to install ``cellpy`` is by using conda:
 This will also install all of the critical dependencies, as well as ``jupyter``
 that comes in handy when working with ``cellpy``.
 
-You can install pre-releases by adding the ``--pre`` flag.
+
+In general, we recommend to install ``cellpy`` in a virtual environment.
+
+To make sure your environment contains the correct packages and dependencies
+required for running ``cellpy``, you can create an environment based on the available
+``environment.yml`` file. Download the
+`environment.yml <https://github.com/jepegit/cellpy/blob/master/environment.yml>`_
+file and place it in the directory shown in your Anaconda Prompt, and run:
+
+.. code-block:: console
+
+    conda env create -f environment.yml
+
+Then activate your environment:
+
+.. code-block:: console
+
+    conda activate cellpy
+
+and install ``cellpy`` within the environment.
 
 
 Pip
 ---
 
-If you would like to install only ``cellpy``, you should install using pip.
-Note that ``cellpy`` uses several packages that are a bit cumbersome to install
-on windows (e.g. ``scipy``, ``numpy`` and ``pytables`` ).
-
-Install ``cellpy`` by running this command in your terminal:
+If you would like to install 'only' ``cellpy``, you can use pip:
 
 .. code-block:: console
 
-    $ python -m pip install cellpy
+    $ pip install cellpy
 
 
-You can install pre-releases by adding the ``--pre`` flag.
+Note that ``cellpy`` uses several packages that are a bit cumbersome to install
+on windows (e.g. ``scipy``, ``numpy`` and ``pytables`` ) and when using pip,
+you have to take care of this yourself. To make sure to install
+all the required dependencies, we recommend to create an environment based
+the provided ``environment.yml`` (see installation with conda above):
 
+.. code-block:: console
+
+    $ conda env create -f environment.yml
+
+
+Pre-releases can be installed by adding the ``--pre`` flag to the installation command:
+
+.. code-block:: console
+
+   python -m pip install --pre cellpy
+
+
+
+.. _Cellpy_Install_Linux:
 
 
 Linux
@@ -56,16 +97,174 @@ Pip
 This is how to install ``cellpy`` using conda on Linux.
 Be aware that you might have to install additional packages.
 
+.. _Cellpy_Install_Sources:
 
 Installation from sources
 ==============
 
-This is how to install ``cellpy`` from sources.
+The sources for ``cellpy`` can be downloaded from the `Github repo`_.
+
+You can clone the public repository by:
+
+.. code-block:: console
+
+    $ git clone git://github.com/jepegit/cellpy
 
 
[…]
