### 2024-02-16 ab72603
macos py10 and py11 only - updated installation steps for pytables
