### 2024-02-16 779dac4
macos pip - install unixodbc using brew
