### 2024-02-16 074a741
279 ci workflows dont test macos (#281)
* update github actions and add for macos-latest and macos-14

* update github actions and add for macos-latest and macos-14

* update github actions and add for macos-latest and macos-14 (3.9 fix)

* update github actions and add for macos-latest and macos-14 (3.9 fix, better naming)

* use brew for all macos actions and add miniconda version for the conda dependent macos actions

* macos py10 and py11 only - updated installation steps for pytables

* macos pip - install unixodbc using brew
