+++ .github/workflows/pip-install-macos-14.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Installing using pip on posix
+name: Installing using pip on Mac M chip
 
 on:
   push:
@@ -14,11 +14,11 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: macos-14
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v4
@@ -28,13 +28,20 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install mdbtools
       run: |
-        sudo apt-get update
-        sudo apt-get install -y mdbtools
-#        sudo apt-get install -y libhdf5-serial-dev
+        brew install mdbtools
+    - name: Install unixodbc
+      run: |
+            brew install unixodbc
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         python -m pip install pytest pytest-timeout pytest-benchmark lmfit
+        python -m pip install cython
+        brew install hdf5
+        brew install c-blosc
+        export HDF5_DIR=/opt/homebrew/opt/hdf5
+        export BLOSC_DIR=/opt/homebrew/opt/c-blosc
+        python -m pip install tables
         # if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Install cellpy
       run: |
