+++ .github/workflows/pytest_macos_14.yml
@@ -1,4 +1,4 @@
-name: Run pytest on linux (conda)
+name: Run pytest on macos M (conda)
 on:
   push:
     branches:
@@ -9,18 +9,19 @@ on:
 
 jobs:
   Install-miniconda-and-run-pytest:
-    runs-on: ubuntu-latest
+    runs-on: macos-14
     defaults:
       run:
         shell: bash -l {0}
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.10", "3.11"]
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
         with:
           auto-update-conda: true
+          miniconda-version: latest
           activate-environment: cellpy_dev
           environment-file: github_actions_environment.yml
           python-version: ${{ matrix.python-version }}
@@ -33,8 +34,8 @@ jobs:
 
       - name: Install mdbtools
         run: |
-          sudo apt-get update
-          sudo apt-get install -y mdbtools
+          brew install mdbtools
+#          sudo apt-get install -y mdbtools
 
       - name: pytest
         run: pytest
