+++ .github/workflows/pip-install-macos.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Installing using pip on posix
+name: Installing using pip on Mac Intel
 
 on:
   push:
@@ -14,7 +14,7 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
     strategy:
       fail-fast: false
       matrix:
@@ -28,9 +28,10 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install mdbtools
       run: |
-        sudo apt-get update
-        sudo apt-get install -y mdbtools
-#        sudo apt-get install -y libhdf5-serial-dev
+        brew install mdbtools
+    - name: Install unixodbc
+      run: |
+        brew install unixodbc
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
