+++ .github/workflows/pytest_macos.yml
@@ -1,4 +1,4 @@
-name: Run pytest on linux (conda)
+name: Run pytest on macos (conda)
 on:
   push:
     branches:
@@ -9,7 +9,7 @@ on:
 
 jobs:
   Install-miniconda-and-run-pytest:
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
     defaults:
       run:
         shell: bash -l {0}
@@ -17,10 +17,11 @@ jobs:
       matrix:
         python-version: ["3.9", "3.10", "3.11"]
     steps:
-      - uses: actions/checkout@v4
-      - uses: conda-incubator/setup-miniconda@v3
+      - uses: actions/checkout@v3
+      - uses: conda-incubator/setup-miniconda@v2
         with:
           auto-update-conda: true
+          miniconda-version: latest
           activate-environment: cellpy_dev
           environment-file: github_actions_environment.yml
           python-version: ${{ matrix.python-version }}
@@ -33,8 +34,8 @@ jobs:
 
       - name: Install mdbtools
         run: |
-          sudo apt-get update
-          sudo apt-get install -y mdbtools
+          brew install mdbtools
+#          sudo apt-get install -y mdbtools
 
       - name: pytest
         run: pytest
