### 2024-02-16 67dc09b
add dependency on rich
