### 2024-02-16 82f5504
run pytest using python -m pytest instead of only issuing pytest
