### 2024-02-17 a8468ce
use only conda for installing on macOS
