+++ .github/workflows/pip-install-macos-14.yml
@@ -1,54 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
-
-name: Installing using pip on Mac M chip
-
-on:
-  push:
-    branches:
-      - "**"
-  pull_request:
-    branches: [ "master" ]
-  workflow_dispatch:
-
-jobs:
-  build:
-
-    runs-on: macos-14
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.10", "3.11"]
-
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install mdbtools
-      run: |
-        brew install mdbtools
-    - name: Install unixodbc
-      run: |
-            brew install unixodbc
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install pytest pytest-timeout pytest-benchmark lmfit
-        python -m pip install cython
-        brew install hdf5
-        brew install c-blosc
-        export HDF5_DIR=/opt/homebrew/opt/hdf5
-        export BLOSC_DIR=/opt/homebrew/opt/c-blosc
-        python -m pip install tables
-        # if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-    - name: Install cellpy
-      run: |
-        python -m pip install -e .
-    - name: Test with pytest
-      run: |
-        python -m pytest
-    - name: Run cellpy setup
-      run: |
-        cellpy setup --silent
