### 2024-02-17 cfc63fa
add cellpy checks to CI
