+++ .github/workflows/pip-install-win.yml
@@ -40,3 +40,6 @@ jobs:
     - name: Run cellpy setup
       run: |
         cellpy setup --silent
+    - name: Run the cellpy check
+      run: |
+        cellpy info --check
