+++ .github/workflows/pip-install-macos-14.yml
@@ -52,3 +52,6 @@ jobs:
     - name: Run cellpy setup
       run: |
         cellpy setup --silent
+    - name: Run the cellpy check
+      run: |
+        cellpy info --check
