### 2024-02-17 3f97d21
clean up check text
