### 2024-02-17 687e290
clean up check text
