### 2024-02-17 7e0d25e
add non-dev-mode to save example env-file in setup
