### 2024-02-17 df53b1b
add non-dev-mode to save example env-file in setup (.write_text instead of .write)
