+++ .github/workflows/pip-install-macos-14.yml
@@ -55,3 +55,6 @@ jobs:
     - name: Run the cellpy check
       run: |
         cellpy info --check
+    - name: Show the configuration
+      run: |
+        cellpy info -p
