+++ .github/workflows/pip-install-win.yml
@@ -43,3 +43,6 @@ jobs:
     - name: Run the cellpy check
       run: |
         cellpy info --check
+    - name: Show the configuration
+      run: |
+        cellpy info -p
