### 2024-02-17 57774ac
add non-dev-mode to save example env-file in setup (.write_text instead of .write)
