### 2024-02-17 357ae59
add non-dev-mode to save example env-file in setup (fix 3)
