### 2024-02-17 045664d
add non-dev-mode to save example env-file in setup (fix 3)
