### 2024-02-17 jepegit/cellpy#276 @jepegit
Updated how check outputs. Should be more understandable now.
