+++ docs/API/index.rst
@@ -1,0 +1,6 @@
+API Reference
+-------------
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
