### 2024-02-18 3e2312e
check if myst_parser can be added as extension to sphinx without breaking anything
