+++ docs/API/index.rst
@@ -3,4 +3,4 @@ API Reference
 
 * :ref:`genindex`
 * :ref:`modindex`
-* :ref:`search`
+
