+++ docs/index.rst
@@ -25,4 +25,4 @@ Indices and tables
 
 * :ref:`genindex`
 * :ref:`modindex`
-* :ref:`search`
+
