+++ docs/environment_rtd.yml
@@ -28,4 +28,5 @@ dependencies:
     - python-dotenv
     - fabric
     - sphinx_rtd_theme
+    - myst-parser
 
