+++ docs/source/cellpy.readers.instruments.loader_specific_modules.rst
@@ -1,21 +1,0 @@
-cellpy.readers.instruments.loader\_specific\_modules package
-============================================================
-
-Submodules
-----------
-
-cellpy.readers.instruments.loader\_specific\_modules.biologic\_file\_format module
-----------------------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.loader_specific_modules.biologic_file_format
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: cellpy.readers.instruments.loader_specific_modules
-   :members:
-   :undoc-members:
-   :show-inheritance:
