+++ docs/source/cellpy.readers.rst
@@ -1,69 +1,0 @@
-cellpy.readers package
-======================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   cellpy.readers.instruments
-
-Submodules
-----------
-
-cellpy.readers.cellreader module
---------------------------------
-
-.. automodule:: cellpy.readers.cellreader
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.core module
---------------------------
-
-.. automodule:: cellpy.readers.core
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.dbreader module
-------------------------------
-
-.. automodule:: cellpy.readers.dbreader
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.do module
-------------------------
-
-.. automodule:: cellpy.readers.do
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.filefinder module
---------------------------------
-
-.. automodule:: cellpy.readers.filefinder
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.sql\_dbreader module
------------------------------------
-
-.. automodule:: cellpy.readers.sql_dbreader
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: cellpy.readers
-   :members:
-   :undoc-members:
-   :show-inheritance:
