+++ docs/source/cellpy.internals.rst
@@ -1,21 +1,0 @@
-cellpy.internals package
-========================
-
-Submodules
-----------
-
-cellpy.internals.core module
-----------------------------
-
-.. automodule:: cellpy.internals.core
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: cellpy.internals
-   :members:
-   :undoc-members:
-   :show-inheritance:
