+++ docs/source/cellpy.rst
@@ -1,0 +1,48 @@
+cellpy package
+==============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   cellpy.internals
+   cellpy.parameters
+   cellpy.readers
+   cellpy.utils
+
+Submodules
+----------
+
+cellpy.cli module
+-----------------
+
+.. automodule:: cellpy.cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.exceptions module
+------------------------
+
+.. automodule:: cellpy.exceptions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.log module
+-----------------
+
+.. automodule:: cellpy.log
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: cellpy
+   :members:
+   :undoc-members:
+   :show-inheritance:
