+++ docs/source/cellpy.readers.instruments.rst
@@ -1,0 +1,143 @@
+cellpy.readers.instruments package
+==================================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   cellpy.readers.instruments.configurations
+   cellpy.readers.instruments.loader_specific_modules
+   cellpy.readers.instruments.processors
+
+Submodules
+----------
+
+cellpy.readers.instruments.arbin\_res module
+--------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.arbin_res
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.arbin\_sql module
+--------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.arbin_sql
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.arbin\_sql\_7 module
+-----------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.arbin_sql_7
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.arbin\_sql\_csv module
+-------------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.arbin_sql_csv
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.arbin\_sql\_h5 module
+------------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.arbin_sql_h5
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.arbin\_sql\_xlsx module
+--------------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.arbin_sql_xlsx
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.base module
+--------------------------------------
+
+.. automodule:: cellpy.readers.instruments.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.biologics\_mpr module
+------------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.biologics_mpr
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.custom module
+----------------------------------------
+
+.. automodule:: cellpy.readers.instruments.custom
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.ext\_nda\_reader module
+--------------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.ext_nda_reader
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.local\_instrument module
+---------------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.local_instrument
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.maccor\_txt module
+---------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.maccor_txt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.neware\_txt module
+---------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.neware_txt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.neware\_xlsx module
+----------------------------------------------
+
+.. automodule:: cellpy.readers.instruments.neware_xlsx
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.readers.instruments.pec\_csv module
[…]
