+++ docs/source/cellpy.utils.rst
@@ -1,0 +1,109 @@
+cellpy.utils package
+====================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   cellpy.utils.batch_tools
+
+Submodules
+----------
+
+cellpy.utils.batch module
+-------------------------
+
+.. automodule:: cellpy.utils.batch
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.collectors module
+------------------------------
+
+.. automodule:: cellpy.utils.collectors
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.diagnostics module
+-------------------------------
+
+.. automodule:: cellpy.utils.diagnostics
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.easyplot module
+----------------------------
+
+.. automodule:: cellpy.utils.easyplot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.example\_data module
+---------------------------------
+
+.. automodule:: cellpy.utils.example_data
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.helpers module
+---------------------------
+
+.. automodule:: cellpy.utils.helpers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.ica module
+-----------------------
+
+.. automodule:: cellpy.utils.ica
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.live module
+------------------------
+
+.. automodule:: cellpy.utils.live
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.ocv\_rlx module
+----------------------------
+
+.. automodule:: cellpy.utils.ocv_rlx
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.plotutils module
+-----------------------------
+
+.. automodule:: cellpy.utils.plotutils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.processor module
+-----------------------------
+
+.. automodule:: cellpy.utils.processor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: cellpy.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
