+++ docs/source/cellpy.utils.batch_tools.rst
@@ -1,0 +1,101 @@
+cellpy.utils.batch\_tools package
+=================================
+
+Submodules
+----------
+
+cellpy.utils.batch\_tools.batch\_analyzers module
+-------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_analyzers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_core module
+--------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_experiments module
+---------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_experiments
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_exporters module
+-------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_exporters
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_helpers module
+-----------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_helpers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_journals module
+------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_journals
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_plotters module
+------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_plotters
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.batch\_reporters module
+-------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.batch_reporters
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.dumpers module
+----------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.dumpers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.engines module
+----------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.engines
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cellpy.utils.batch\_tools.sqlite\_from\_excel\_db module
+--------------------------------------------------------
+
+.. automodule:: cellpy.utils.batch_tools.sqlite_from_excel_db
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: cellpy.utils.batch_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
