### 2024-02-18 f0773b9
add sphinx-autoapi (can be used for creating api-documentation without installing cellpy on readthedocs)
