### 2024-02-18 3794851
remove the self-made API reference - rely on the autoapi to do the work for us
