+++ docs/API/index.rst
@@ -1,6 +1,0 @@
-API Reference
--------------
-
-* :ref:`genindex`
-* :ref:`modindex`
-
