### 2024-02-18 44a469a
add autoapi_template_dir
