### 2024-02-19 a233282
modify autoapi settings
