+++ docs/_autoapi_templates/index.rst
@@ -1,7 +1,9 @@
 API Reference
 =============
 
-This page contains auto-generated API reference documentation [#f1]_.
+.. note::
+
+    This page contains auto-generated API reference documentation [#f1]_.
 
 .. toctree::
    :titlesonly:
