### 2024-02-19 5d0806c
add basic text on checkup and configuration
