+++ docs/getting_started/configuration.rst
@@ -1,3 +1,260 @@
 ============
-Configuring ``cellpy``
+Setup and configuration
 ============
+
+To get the most out of ``cellpy`` it is to best to set it up properly.
+
+To help with this, you can use the ``setup`` command. If you include
+the  ``-i`` option, the setup will be done interactively, and you will
+be prompted for your preferred location for the different folders and
+directories ``cellpy`` uses:
+
+.. code-block:: shell
+
+    cellpy setup -i
+
+This creates the cellpy configuration file ``.cellpy_prms_USERNAME.conf``
+in your home directory (USERNAME = your user name), or update it if it exists,
+and try to create the following directory structure:
+
+.. code-block:: shell
+
+    batchfiles/
+    cellpyfiles/
+    db/
+    examples/
+    instruments/
+    logs/
+    notebooks/
+    out/
+    raw/
+    templates/
+
+
+If you want to specify a root folder different from the default (your HOME
+folder), you can use the ``-d`` option *e.g.*
+``cellpy setup -i -d /Users/kingkong/cellpydir``
+
+
+.. hint:: You can always edit your configurations directly in the cellpy configuration
+   file ``.cellpy_prms_USER.conf``. This file should be located inside your
+   home directory, /. in posix and c:\users\USERNAME in not-too-old windows.
+
+To check, where your configuration file is actually located, you can run:
+
+.. code-block:: shell
+
+    $ cellpy info --configloc
+    [cellpy] -> C:\Users\jepe\.cellpy_prms_jepe.conf
+
+To get the filepath to your config file AND other cellpy info, run:
+
+.. code-block:: console
+
+    cellpy info -l
+
+.. note::
+
+    It is recommended to re-run setup each time you update ``cellpy``.
+
+
+The configuration file
+----------------------
+The paths to raw data, the cellpy data base file, file locations etc. are set in
+the ``.cellpy_prms_USER.conf`` file that is located in your home directory.
+``cellpy`` tries to read this .conf file when imported the first time,
+and looks in your user directory after files named ``.cellpy_prms_SOMENAME.conf``.
+(If you have run ``cellpy setup``, the configuration file will be placed in
+the appropriate place with the appropriate filename.)
+
+The configuration file is a YAML-file and it is reasonably easy to read and edit
+in a text editor (remember that YAML is rather strict with regards to spaces and
+indentations).
+
+Within the config file, the paths are the most important parts that need to
+be set up correctly. This tells ``cellpy`` where to find (and save) different files,
+such as the database file and raw data.
+
+Furthermore, the config file contains details about the database-file
+(see below :ref:`_Cellpy_Database_File`) to be used for cell info and metadata
+(i.e. type and structure of the database file such as column headers etc.).
+
+
+Configuration file - example
+.............................
+
+As an example, here are the first lines
+from one of the authors' configuration file:
+
+.. code-block:: yaml
+
+    ---
+    Paths:
+        outdatadir: C:\scripts\processing_cellpy\out
+        rawdatadir: I:\Org\MPT-BAT-LAB\Arbin-data
+        cellpydatadir: C:\scripts\processing_cellpy\cellpyfiles
+        db_path: C:\scripts\processing_cellpy\db
+        filelogdir: C:\scripts\processing_cellpy\logs
+        examplesdir: C:\scripts\processing_cellpy\examples
+        notebookdir: C:\scripts\processing_cellpy\notebooks
+        templatedir: C:\scripting\processing_cellpy\templates
+        batchfiledir: C:\scripts\processing_cellpy\batchfiles
+        db_filename: 2023_Cell_Analysis_db_001.xlsx
+        env_file: .env_cellpy
+
+
+    FileNames:
+        file_name_format: YYYYMMDD_[NAME]EEE_CC_TT_RR
+
+
[…]
