### 2024-02-19 69aef56
modify autoapi settings
