+++ _old_docs/adapted_readme.rst
@@ -1,56 +1,0 @@
-.. image:: _static/cellpy-icon-long.svg
-  :height: 100
-  :alt: cellpy
-
-===================================================================
- - *a library for assisting in analysing batteries and cells*
-===================================================================
-
-
-.. image:: https://img.shields.io/pypi/v/cellpy.svg
-        :target: https://pypi.python.org/pypi/cellpy
-
-.. image:: https://readthedocs.org/projects/cellpy/badge/?version=latest
-        :target: https://cellpy.readthedocs.io/en/latest/?badge=latest
-        :alt: Documentation Status
-
-.. image:: https://static.pepy.tech/badge/cellpy
-        :target: https://pepy.tech/project/cellpy
-
-
-This Python Package was developed to help the
-researchers at IFE, Norway, in their cumbersome task of
-interpreting and handling data from cycling tests of
-batteries and cells.
-
-|ProjectVersion|
-
-Installation and dependencies
-=============================
-
-The easiest way to install ``cellpy`` is to install with conda or pip.
-
-With conda::
-
-   conda install -c conda-forge cellpy
-
-Or if you prefer installing using pip::
-
-   python -m pip install cellpy
-
-Have a look at the documentation for more detailed installation procedures, especially
-with respect to "difficult" dependencies when installing with pip.
-
-Licence
-=======
-
-``cellpy`` is free software made available under the MIT License.
-
-Features
-========
-
-* Load test-data and store in a common format.
-* Summarize and compare data.
-* Filter out the steps of interest.
-* Process and plot the data.
-* And more...
