+++ docs/other/other.md
@@ -1,0 +1,11 @@
+# Credits
+
+## Development lead
+
+## Contributors
+
+# History
+
+# Documentation Archive
+
+include link to old documentation
