+++ docs/source/cellpy.readers.md
@@ -1,0 +1,74 @@
+# cellpy.readers package
+
+## Subpackages
+
+```{toctree}
+:maxdepth: 4
+
+cellpy.readers.instruments
+```
+
+## Submodules
+
+## cellpy.readers.cellreader module
+
+```{eval-rst}
+.. automodule:: cellpy.readers.cellreader
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## cellpy.readers.core module
+
+```{eval-rst}
+.. automodule:: cellpy.readers.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## cellpy.readers.dbreader module
+
+```{eval-rst}
+.. automodule:: cellpy.readers.dbreader
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## cellpy.readers.do module
+
+```{eval-rst}
+.. automodule:: cellpy.readers.do
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## cellpy.readers.filefinder module
+
+```{eval-rst}
+.. automodule:: cellpy.readers.filefinder
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## cellpy.readers.sql_dbreader module
+
+```{eval-rst}
+.. automodule:: cellpy.readers.sql_dbreader
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## Module contents
+
+```{eval-rst}
+.. automodule:: cellpy.readers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
