+++ _old_docs/main_description/authors.md
@@ -1,0 +1,3 @@
+```{eval-rst}
+.. include:: ../../AUTHORS.rst
+```
