+++ docs/fundamentals/fundamentals.rst
@@ -1,5 +1,0 @@
-============
-The fundamentals of ``cellpy``
-============
-
-This is how ``cellpy`` is build up and works
