+++ docs/index.md
@@ -1,0 +1,28 @@
+% cellpy documentation master file, created by
+% sphinx-quickstart on Tue Jul  9 22:26:36 2013.
+% You can adapt this file completely to your liking, but it should at least
+% contain the root `toctree` directive.
+
+# Welcome to cellpy's documentation!
+
+```{eval-rst}
+.. include::
+    adapted_readme.rst
+```
+
+# Contents
+
+```{toctree}
+:maxdepth: 2
+
+getting_started/index.rst
+fundamentals/index.rst
+examples/index.rst
+contributing/index.rst
+other/index.rst
+```
+
+# Indices and tables
+
+- {ref}`genindex`
+- {ref}`modindex`

[…]
