+++ _old_docs/examples_and_tutorials/utils/collectors.rst
@@ -1,6 +1,0 @@
-.. _utils-collectors::
-
-Collectors
-==========
-
-Todo.
