+++ _old_docs/examples_and_tutorials/loaders/06_custom.rst
@@ -1,7 +1,0 @@
-.. _instruments-custom:
-
-Custom
-------
-
-TODO.
-
