+++ _old_docs/examples_and_tutorials/loaders/01_arbin.rst
@@ -1,6 +1,0 @@
-.. _instruments-arbin:
-
-Arbin
------
-
-TODO.
