+++ _old_docs/examples_and_tutorials/utils/easyplot.rst
@@ -1,6 +1,0 @@
-.. _utils-easyplot::
-
-Easyplot
-========
-
-Todo.
