+++ _old_docs/examples_and_tutorials/utils/templates.rst
@@ -1,11 +1,0 @@
-.. _utils-templates:
-
-Templates
-=========
-
-.. note:: This chapter would benefit from some more love and care. Any help
-    on that would be highly appreciated.
-
-    Some info can be found in one of the example notebooks (:ref:`notebooks`).
-    It would be more logical if that information was here instead.
-    Sorry for that.
