+++ docs/conf.py
@@ -337,7 +337,7 @@ texinfo_documents = [
 
 nbsphinx_kernel_name = "python3"
 autoapi_dirs = ["../cellpy"]
-autoapi_template_dir = "_autoapi_templates"
+# autoapi_template_dir = "_autoapi_templates"
 autoapi_options = [
     "members",
     "undoc-members",
