+++ docs/source/cellpy.readers.rst
@@ -1,74 +1,69 @@
-# cellpy.readers package
+cellpy.readers package
+======================
 
-## Subpackages
+Subpackages
+-----------
 
-```{toctree}
-:maxdepth: 4
+.. toctree::
+   :maxdepth: 4
 
-cellpy.readers.instruments
-```
+   cellpy.readers.instruments
 
-## Submodules
+Submodules
+----------
 
-## cellpy.readers.cellreader module
+cellpy.readers.cellreader module
+--------------------------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers.cellreader
    :members:
    :undoc-members:
    :show-inheritance:
-```
 
-## cellpy.readers.core module
+cellpy.readers.core module
+--------------------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers.core
    :members:
    :undoc-members:
    :show-inheritance:
-```
 
-## cellpy.readers.dbreader module
+cellpy.readers.dbreader module
+------------------------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers.dbreader
    :members:
    :undoc-members:
    :show-inheritance:
-```
 
-## cellpy.readers.do module
+cellpy.readers.do module
+------------------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers.do
    :members:
    :undoc-members:
    :show-inheritance:
-```
 
-## cellpy.readers.filefinder module
+cellpy.readers.filefinder module
+--------------------------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers.filefinder
    :members:
    :undoc-members:
    :show-inheritance:
-```
 
-## cellpy.readers.sql_dbreader module
+cellpy.readers.sql\_dbreader module
+-----------------------------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers.sql_dbreader
    :members:
    :undoc-members:
    :show-inheritance:
-```
 
-## Module contents
+Module contents
+---------------
 
-```{eval-rst}
 .. automodule:: cellpy.readers
    :members:
    :undoc-members:
    :show-inheritance:
-```
