+++ docs/source/modules.md
@@ -1,7 +1,0 @@
-# cellpy
-
-```{toctree}
-:maxdepth: 4
-
-cellpy
-```
