+++ docs/source/cellpy.readers.instruments.loader_specific_modules.md
@@ -1,21 +1,0 @@
-# cellpy.readers.instruments.loader_specific_modules package
-
-## Submodules
-
-## cellpy.readers.instruments.loader_specific_modules.biologic_file_format module
-
-```{eval-rst}
-.. automodule:: cellpy.readers.instruments.loader_specific_modules.biologic_file_format
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-## Module contents
-
-```{eval-rst}
-.. automodule:: cellpy.readers.instruments.loader_specific_modules
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
