+++ docs/source/cellpy.parameters.md
@@ -1,47 +1,0 @@
-# cellpy.parameters package
-
-## Subpackages
-
-```{toctree}
-:maxdepth: 4
-
-cellpy.parameters.legacy
-```
-
-## Submodules
-
-## cellpy.parameters.internal_settings module
-
-```{eval-rst}
-.. automodule:: cellpy.parameters.internal_settings
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-## cellpy.parameters.prmreader module
-
-```{eval-rst}
-.. automodule:: cellpy.parameters.prmreader
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-## cellpy.parameters.prms module
-
-```{eval-rst}
-.. automodule:: cellpy.parameters.prms
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-## Module contents
-
-```{eval-rst}
-.. automodule:: cellpy.parameters
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
