+++ docs/_autoapi_templates/python/function.md
@@ -1,18 +1,0 @@
-{% if obj.display %}
-.. py:function:: {{ obj.short_name }}({{ obj.args }}){% if obj.return_annotation is not none %} -> {{ obj.return_annotation }}{% endif %}
-
-{% for (args, return_annotation) in obj.overloads %}
-
-: {{ obj.short_name }}({{ args }}){% if return_annotation is not none %} -> {{ return_annotation }}{% endif %}
-
-{% endfor %}
-
-: {% for property in obj.properties %}
-  \:{{ property }}:
-  {% endfor %}
-
-  {% if obj.docstring %}
-  {{ obj.docstring|indent(3) }}
-  {% endif %}
-
-{% endif %}

[…]
