+++ docs/conf.py
@@ -1,80 +1,35 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-#
-# cellpy documentation build configuration file, created by
-# sphinx-quickstart on Tue Jul  9 22:26:36 2013.
-#
-# This file is execfile()d with the current directory set to its
-# containing dir.
-#
-# Note that not all possible configuration values are present in this
-# autogenerated file.
-#
-# All configuration values have a default; values that are commented out
-# serve to show the default.
 import os
 import sys
 
-# importing mock for modules that cant be installed by RTD
+INCLUDE_SOURCE_FILES = True
 
-# from unittest.mock import MagicMock # python3
-# from mock import Mock as MagicMock
-#
-# class Mock(MagicMock):
-#     @classmethod
-#     def __getattr__(cls, name):
-#             return Mock()
-#
-# MOCK_MODULES = ['pyodbc', 'lmfit', 'pytables', 'pandas']
-# sys.modules.update((mod_name, Mock()) for mod_name in MOCK_MODULES)
-
-
-# If extensions (or modules to document with autodoc) are in another
-# directory, add these directories to sys.path here. If the directory is
-# relative to the documentation root, use os.path.abspath to make it
-# absolute, like shown here.
-# sys.path.insert(0, os.path.abspath('.'))
-
-# Get the project root dir, which is the parent dir of this
-cwd = os.getcwd()
-project_root = os.path.dirname(cwd)
-project_prmsdir = os.path.join(project_root, r"cellpy\parameters")
-project_utils = os.path.join(project_root, r"cellpy\utils")
-project_scripts = os.path.join(project_root, r"cellpy\scripts")
-project_readers = os.path.join(project_root, r"cellpy\readers")
-project_internals = os.path.join(project_root, r"cellpy\internals")
-
-# print project_root
-# print project_prmsdir
-# print project_utils
-# print project_scripts
-# print project_readers
-
+project_root = os.path.abspath("../")
 name = "cellpy"
 version_ns = {}
 with open(os.path.join(project_root, name, "_version.py")) as f:
     exec(f.read(), {}, version_ns)
 
 
-# Insert the project root dir as the first element in the PYTHONPATH.
-# This lets us ensure that the source package is imported, and that its
-# version is used.
+# -- Path setup -------------------------------------------------------
 
-sys.path.insert(0, project_root)
-sys.path.insert(0, project_prmsdir)
-sys.path.insert(0, project_utils)
-sys.path.insert(0, project_scripts)
-sys.path.insert(0, project_readers)
-sys.path.insert(0, project_internals)
+if INCLUDE_SOURCE_FILES:
+    project_prmsdir = os.path.join(project_root, r"cellpy\parameters")
+    project_utils = os.path.join(project_root, r"cellpy\utils")
+    project_scripts = os.path.join(project_root, r"cellpy\scripts")
+    project_readers = os.path.join(project_root, r"cellpy\readers")
+    project_internals = os.path.join(project_root, r"cellpy\internals")
+
+    sys.path.insert(0, project_root)
+    sys.path.insert(0, project_prmsdir)
+    sys.path.insert(0, project_utils)
+    sys.path.insert(0, project_scripts)
+    sys.path.insert(0, project_readers)
+    sys.path.insert(0, project_internals)
 
 
 # -- General configuration ---------------------------------------------
 
-# If your documentation needs a minimal Sphinx version, state it here.
-# needs_sphinx = '1.0'
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
+# Extensions
 extensions = [
     "sphinx.ext.autodoc",
[…]
