+++ docs/adapted_readme.md
@@ -1,4 +1,4 @@
-<img src="static/cellpy-icon-long.svg" height="300" alt="cellpy-icon">
+<img src="_static/cellpy-icon-long.svg" height="300" alt="cellpy-icon">
 
 # - *a library for assisting in analysing batteries and cells*
 
