+++ docs/source/cellpy.readers.instruments.configurations.rst
@@ -1,61 +1,0 @@
-cellpy.readers.instruments.configurations package
-=================================================
-
-Submodules
-----------
-
-cellpy.readers.instruments.configurations.maccor\_txt\_four module
-------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.configurations.maccor_txt_four
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.instruments.configurations.maccor\_txt\_one module
------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.configurations.maccor_txt_one
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.instruments.configurations.maccor\_txt\_three module
--------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.configurations.maccor_txt_three
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.instruments.configurations.maccor\_txt\_two module
------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.configurations.maccor_txt_two
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.instruments.configurations.maccor\_txt\_zero module
-------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.configurations.maccor_txt_zero
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cellpy.readers.instruments.configurations.neware\_txt\_zero module
-------------------------------------------------------------------
-
-.. automodule:: cellpy.readers.instruments.configurations.neware_txt_zero
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: cellpy.readers.instruments.configurations
-   :members:
-   :undoc-members:
-   :show-inheritance:
