+++ _old_docs/adapted_readme.md
@@ -23,7 +23,6 @@ researchers at IFE, Norway, in their cumbersome task of
 interpreting and handling data from cycling tests of
 batteries and cells.
 
-{{ ProjectVersion }}
 
 ## Installation and dependencies
 
