+++ docs/getting_started/installation.rst
@@ -1,0 +1,24 @@
+.. highlight:: shell
+
+============
+Installation
+============
+``cellpy`` is available on :ref:`Cellpy_Install_Windows` and :ref:`Cellpy_Setup_Linux` and can be installed using ``pip``
+or ``conda``, or :ref:`Cellpy_Install_Sources`.
+
+For more details on dependencies, have a look at :ref:`Cellpy_Dependencies`.
+
+For a tea-spoon explanation on how to install ``cellpy`` on windows, including the
+installation of python and setup of virtual environments, (see below
+:ref:`Cellpy_Setup_Teaspoon`).
+
+
+
+After installing ``cellpy``, continue to
+- :ref: `Setup and configuration`
+- :ref: `Check your cellpy installation`
+
+.. _Cellpy_Install_Windows:
+
+Installing on Windows
+=====================
