### 2024-02-19 24ab309
try to fix the installation.rst file so that it can be converted to markdown
