+++ docs/getting_started/installation.rst
@@ -22,3 +22,233 @@ After installing ``cellpy``, continue to
 
 Installing on Windows
 =====================
+
+
+Conda
+-----
+
+The easiest way to install ``cellpy`` is by using conda:
+
+.. code-block:: console
+
+    $ conda install -c conda-forge cellpy
+
+
+This will also install all of the critical dependencies, as well as ``jupyter``
+that comes in handy when working with ``cellpy``.
+
+
+In general, we recommend to install ``cellpy`` in a virtual environment.
+
+
+Pip
+---
+
+If you would like to install 'only' ``cellpy``, you can use pip:
+
+.. code-block:: console
+
+    $ pip install cellpy
+
+
+Note that ``cellpy`` uses several packages that are a bit cumbersome to install
+on windows (e.g. ``pytables``) and when using pip,
+you have to take care of this yourself.
+
+Pre-releases can be installed by adding the ``--pre`` flag to the installation command:
+
+.. code-block:: console
+
+    $ python -m pip install --pre cellpy
+
+
+.. _Cellpy_Install_Linux:
+
+Linux
+=====
+
+
+Conda
+-----
+This is how to install ``cellpy`` using conda on Linux.
+Be aware that you might have to install additional packages.
+
+This includes ``libobdc``, that can be installed on Ubuntu like this:
+
+.. code-block:: console
+
+    $ sudo apt update
+    $ sudo apt install unixodbc-dev
+
+
+Pip
+---
+This is how to install ``cellpy`` using conda on Linux.
+Be aware that you might have to install additional packages.
+
+.. _Cellpy_Install_Sources:
+
+Installation from sources
+=========================
+
+
+The sources for ``cellpy`` can be downloaded from the `Github repo`_.
+
+You can clone the public repository by:
+
+.. code-block:: console
+
+    $ git clone git://github.com/jepegit/cellpy
+
+
+To make sure to install all the required dependencies, we recommend
+to create an environment based the provided ``environment.yml``:
+
+.. code-block:: console
+
+    $ conda env create -f environment.yml
+
+
+Once you have a copy of the source, you can install in development
+mode using pip:
+
+.. code-block:: console
+
+    $ pip install -e .
+
+(assuming that you are in the project folder, *i.e.* the folder that
+contains the setup.py file)
+
+
+.. _Cellpy_Dependencies:
+
+Fixing dependencies
+===================
+``cellpy`` relies on a number of other python package and these need
+to be installed. Most of these packages are included when installing
+``cellpy`` using conda or when creating the environment based on the
+``environment.yml`` file as outlined above.
+
+Here is an additional overview on the required dependencies:
+
+Basic dependencies
+------------------
+
+In general, you need the typical scientific python pack, including
+
+- ``numpy``
+- ``scipy``
+- ``pandas``
+
+Additional dependencies are:
+
+- ``pytables`` is needed for working with the HDF5 files (the cellpy-files):
+
+.. code-block:: console
+
+    conda install -c conda-forge pytables
+
+- ``lmfit`` is required to use some of the fitting routines in ``cellpy``:
+
+.. code-block:: console
+
+    conda install -c conda-forge lmfit
+
+- ``holoviz`` and ``plotly``: plotting library used in several of our example notebooks.
+
+- ``jupyter``: used for tutorial notebooks and in general very useful tool
+  for working with and sharing your ``cellpy`` results.
+
+For more details, have a look at the documentation of these packages.
+
+Additional requirements for .res files
+--------------------------------------
+
+Note! .res files from Arbin testers are actually in a Microsoft Access format.
+
+**For Windows users:** if you do not have one of the
+most recent Office versions, you might not be allowed to install a driver
+of different bit than your office version is using (the installers can be found
+`here <https://www.microsoft.com/en-US/download/details.aspx?id=13255>`__).
[…]
