### 2024-02-19 3ee579d
installation.rst converted to markdown
