+++ docs/contributing/index.md
@@ -3,5 +3,5 @@
 ```{toctree}
 :maxdepth: 2
 
-contributing.rst
+contributing.md
 ```
