+++ docs/fundamentals/index.md
@@ -3,8 +3,8 @@
 ```{toctree}
 :maxdepth: 2
 
-fundamentals.rst
-file_formats.rst
-data_structure.rst
-basic_usage.rst
+fundamentals.md
+file_formats.md
+data_structure.md
+basic_usage.md
 ```
