+++ docs/getting_started/installation.md
@@ -1,0 +1,225 @@
+```{highlight} shell
+```
+
+# Installation
+
+`cellpy` is available on {ref}`Cellpy_Install_Windows` and {ref}`Cellpy_Setup_Linux` and can be installed using `pip`
+or `conda`, or {ref}`Cellpy_Install_Sources`.
+
+For more details on dependencies, have a look at {ref}`Cellpy_Dependencies`.
+
+For a tea-spoon explanation on how to install `cellpy` on windows, including the
+installation of python and setup of virtual environments, (see below
+{ref}`Cellpy_Setup_Teaspoon`).
+
+After installing `cellpy`, continue to
+\- :ref: `Setup and configuration`
+\- :ref: `Check your cellpy installation`
+
+(cellpy-install-windows)=
+
+## Installing on Windows
+
+### Conda
+
+The easiest way to install `cellpy` is by using conda:
+
+```console
+$ conda install -c conda-forge cellpy
+```
+
+This will also install all of the critical dependencies, as well as `jupyter`
+that comes in handy when working with `cellpy`.
+
+In general, we recommend to install `cellpy` in a virtual environment.
+
+### Pip
+
+If you would like to install 'only' `cellpy`, you can use pip:
+
+```console
+$ pip install cellpy
+```
+
+Note that `cellpy` uses several packages that are a bit cumbersome to install
+on windows (e.g. `pytables`) and when using pip,
+you have to take care of this yourself.
+
+Pre-releases can be installed by adding the `--pre` flag to the installation command:
+
+```console
+$ python -m pip install --pre cellpy
+```
+
+(cellpy-install-linux)=
+
+## Linux
+
+### Conda
+
+This is how to install `cellpy` using conda on Linux.
+Be aware that you might have to install additional packages.
+
+This includes `libobdc`, that can be installed on Ubuntu like this:
+
+```console
+$ sudo apt update
+$ sudo apt install unixodbc-dev
+```
+
+### Pip
+
+This is how to install `cellpy` using conda on Linux.
+Be aware that you might have to install additional packages.
+
+(cellpy-install-sources)=
+
+## Installation from sources
+
+The sources for `cellpy` can be downloaded from the [Github repo].
+
+You can clone the public repository by:
+
+```console
+$ git clone git://github.com/jepegit/cellpy
+```
+
+To make sure to install all the required dependencies, we recommend
+to create an environment based the provided `environment.yml`:
+
+```console
+$ conda env create -f environment.yml
+```
+
+Once you have a copy of the source, you can install in development
+mode using pip:
+
+```console
+$ pip install -e .
+```
+
+(assuming that you are in the project folder, *i.e.* the folder that
+contains the setup.py file)
+
+(cellpy-dependencies)=
+
+## Fixing dependencies
+
+`cellpy` relies on a number of other python package and these need
+to be installed. Most of these packages are included when installing
+`cellpy` using conda or when creating the environment based on the
+`environment.yml` file as outlined above.
+
+Here is an additional overview on the required dependencies:
+
+### Basic dependencies
+
+In general, you need the typical scientific python pack, including
+
+- `numpy`
+- `scipy`
+- `pandas`
+
+Additional dependencies are:
+
+- `pytables` is needed for working with the HDF5 files (the cellpy-files):
+
+```console
+conda install -c conda-forge pytables
+```
+
+- `lmfit` is required to use some of the fitting routines in `cellpy`:
+
+```console
+conda install -c conda-forge lmfit
+```
+
+- `holoviz` and `plotly`: plotting library used in several of our example notebooks.
+- `jupyter`: used for tutorial notebooks and in general very useful tool
+  for working with and sharing your `cellpy` results.
+
[…]
