+++ docs/conf.py
@@ -77,10 +77,6 @@ version = version_ns["__version__"]
 # The full version, including alpha/beta/rc tags.
 release = version_ns["__version__"]
 
-# Exposing the version number to the docs (use |ProjectVersion| in the docs)
-rst_epilog = f"""
-.. |ProjectVersion| replace:: Version: {version}
-"""
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -297,7 +293,13 @@ autoapi_options = [
     "show-module-summary",
     "show-inheritance-diagram",
 ]
+
+# myst settings
+# see https://myst-parser.readthedocs.io/en/latest/syntax/optional.html
 myst_enable_extensions = [
     "colon_fence",
     "substitution",
 ]
+myst_substitutions = {
+    "ProjectVersion": version,
+}
