### 2024-02-20 bd4fec6
remove env items for rtd (test)
