### 2024-02-20 6d2cee7
remove env items for rtd (test)
