+++ docs/other/authors.md
@@ -1,0 +1,3 @@
+
+```{include} ../../AUTHORS.md
+```
