+++ docs/other/history.md
@@ -1,0 +1,3 @@
+
+```{include} ../../HISTORY.MD
+```
