+++ docs/other/other.md
@@ -1,11 +1,0 @@
-# Credits
-
-## Development lead
-
-## Contributors
-
-# History
-
-# Documentation Archive
-
-include link to old documentation
