+++ docs/other/code-of-conduct.md
@@ -1,0 +1,4 @@
+
+
+```{include} ../../CODE_OF_CONDUCT.md
+```
