+++ HISTORY.md
@@ -1,11 +1,8 @@
-=======
-History
-=======
+# History
 
+## 1.0.1 (under development)
 
-1.0.1 (under development)
-=========================
-
+* General: supports only python 3.9 and up to 3.11
 * Batch: `naked` and `init(empty=True)` easier method for creating batch with empty pages
 * File handling: new fix in `find_files`
 * Batch / Utils: refactored and updated `Collectors` (using `plotly`)
@@ -20,16 +17,14 @@ History
 * Added try-except block for ica post-processing step and add if-clause (suggested by Vajee)
 * Fixed several smaller bugs and improved some of the functionality (most notably in `c.get_cap`)
 * Added CI for macOS
-* Added conda package including sqlalchemy-access
+* Added conda package including `sqlalchemy-access`
 * [under progress]: check how cycle_mode and specific is handled and possibly improve
 * [under progress]: start moving/implementing all methods that modify the data to the `do` module
 * [under progress]: improve plotting tools
 * [under progress]: improve documentation
 * [under progress]: improve feedback from the CLI
 
-
-1.0.0 (2023)
-============
+## 1.0.0 (2023)
 
 * Unit handling: new unit handling (using pint)
 * Unit handling: renaming summary headers
@@ -40,28 +35,24 @@ History
 * Templates: adding more documentation
 * File handling: allow for external raw files (ssh)
 * Readers: neware.txt (one version/model)
-* Readers: arbin_sql7 (experimental, @jtgibson91)
+* Readers: `arbin_sql7` (experimental, @jtgibson91)
 * Batch plotting: collectors for both data collection, plotting and saving
 * OCV-rlx: improvements of the OCV-rlx tools
-* Internals: rename main classes (CellpyData -> CellpyCell, Cell -> Data)
-* Internals: rename .cell property to .data
-* Internals: allow for only one Data-object pr CellpyCell object
+* Internals: rename main classes (`CellpyData` -> `CellpyCell`, `Cell` -> `Data`)
+* Internals: rename `.cell` property to `.data`
+* Internals: allow for only one `Data` object pr `CellpyCell` object
 * CLI: general improvements and bug fixes
 * CLI: move editing of db-file to the edit sub-command
 
 
-
-0.4.3 (2023)
-============
+## 0.4.3 (2023)
 
 * Neware txt loader (supports one specific format only, other formats will have to wait for v.1.0)
 
-
-0.4.2 (2022)
-============
+## 0.4.2 (2022)
 
 * Changed definition of Coulombic Difference (negative of previous)
-* Updated loaders with hooks and additional base class TxtLoader with configuration mechanism
+* Updated loaders with hooks and additional base class `TxtLoader` with configuration mechanism
 * Support for Maccor txt files
 * Supports only python 3.8 and up
 * Optional parameters through batch and pages
@@ -69,71 +60,55 @@ History
 * Restrict use of instrument label to only one option
 * Fix bug in example file (@kevinsmia1939)
 
-
-0.4.1 (2021)
-============
+## 0.4.1 (2021)
 
 * Updated documentations
 * CLI improvements
-* New argument for get_cap: max_cycle
+* New argument for get_cap: `max_cycle`
 * Reverting from using Documents to user home for location of prm file in windows.
 * Easyplot by Amund
 * Arbin sql reader by Muhammad
 
-
-0.4.0 (2020)
-============
+## 0.4.0 (2020)
 
 * Reading arbin .res files with auxiliary data should now work.
 * Many bugs have been removed - many new introduced.
[…]
