+++ docs/other/license.md
@@ -1,0 +1,3 @@
+
+```{include} ../../LICENSE
+```
