+++ CONTRIBUTING.md
@@ -1,21 +1,15 @@
-.. highlight:: shell
-
-============
-Contributing
-============
+# Contributing
 
 Contributions are welcome, and they are greatly appreciated! Every
 little bit helps, and credit will always be given.
 
 You can contribute in many ways:
 
-Types of Contributions
-======================
+## Types of Contributions
 
-Report Bugs
------------
+### Report Bugs
 
-Report bugs at https://github.com/jepegit/cellpy/issues.
+Report bugs in the [cellpy repository at GitHub](https://github.com/jepegit/cellpy/issues).
 
 If you are reporting a bug, please include:
 
@@ -23,29 +17,26 @@ If you are reporting a bug, please include:
 * Any details about your local setup that might be helpful in troubleshooting.
 * Detailed steps to reproduce the bug.
 
-Fix Bugs
---------
+### Fix Bugs
 
 Look through the GitHub issues for bugs. Anything tagged with "bug"
 and "help wanted" is open to whoever wants to implement it.
 
-Implement Features
-------------------
+### Implement Features
 
 Look through the GitHub issues for features. Anything tagged with "enhancement"
 and "help wanted" is open to whoever wants to implement it.
 
-Write Documentation
--------------------
+### Write Documentation
 
 cellpy could always use more documentation, whether as part of the
 official cellpy docs, in docstrings, or even on the web in blog posts,
 articles, and such.
 
-Submit Feedback
----------------
+### Submit Feedback
 
-The best way to send feedback is to file an issue at https://github.com/jepegit/cellpy/issues.
+The best way to send feedback is to file an issue in the
+[cellpy repository at GitHub](https://github.com/jepegit/cellpy/issues).
 
 If you are proposing a feature:
 
@@ -54,65 +45,75 @@ If you are proposing a feature:
 * Remember that this is a volunteer-driven project, and that contributions
   are welcome :)
 
-Get Started!
-============
+## Get Started
 
 Ready to contribute? Here's how to set up ``cellpy`` for local development.
 
 1. Fork the ``cellpy`` repo on GitHub.
-2. Clone your fork locally::
+2. Clone your fork locally:
 
-    $ git clone git@github.com:your_name_here/cellpy.git
+    ```shell
+    git clone git@github.com:your_name_here/cellpy.git
+    ```
 
-3. Create a local python virtual environment and activate it using python's venv utility::
+3. Create a local python virtual environment and activate it using python's venv utility:
 
-    $ python -m venv .venv
-    $ source .venv/bin/activate  # or .venv\Scripts\activate on Windows
+    ```shell
+    python -m venv .venv
+    source .venv/bin/activate  # or .venv\Scripts\activate on Windows
+    ```
 
    Or use ``conda`` environments. See the conda documentation for more information.
    A suitable environment yaml configuration file
    can be found in the root of the repository (``dev_environment.yml``; to create the environment,
    run ``conda env create -f dev_environment.yml``).
 
-4. Install your local copy into your virtualenv::
+4. Install your local copy into your virtualenv:
 
-    $ python -m pip install . -e
+    ```shell
+    python -m pip install . -e
+    ```
 
-5. Create a branch for local development::
+5. Create a branch for local development:
 
-    $ git checkout -b name-of-your-bugfix-or-feature
+    ```shell
+    git checkout -b name-of-your-bugfix-or-feature
+    ```
 
    Now you can make your changes locally.
 
-6. When you're done making changes, check that your changes pass the tests::
+6. When you're done making changes, check that your changes pass the tests:
 
-    $ pytest
+    ```shell
+    pytest
+    ```
 
    If there are any libraries missing (it could happen) just pip install them into your virtual environment (or
[…]
