### 2024-02-20 23fe28a
fix setup and manifest (rst -> md)
