### 2024-02-20 7e14d00
improving getting started section
