+++ _old_docs/examples_and_tutorials/utils.md
@@ -1,33 +1,0 @@
-# Using some of the cellpy special utilities
-
-```{eval-rst}
-.. include:: utils/tut_ocv_rlx.rst
-```
-
-```{eval-rst}
-.. include:: utils/ica.rst
-```
-
-```{eval-rst}
-.. include:: utils/batch.rst
-```
-
-```{eval-rst}
-.. include:: utils/collectors.rst
-```
-
-```{eval-rst}
-.. include:: utils/plotting.rst
-```
-
-```{eval-rst}
-.. include:: utils/easyplot.rst
-```
-
-```{eval-rst}
-.. include:: utils/templates.rst
-```
-
-```{eval-rst}
-.. include:: utils/custom_file_loaders.rst
-```
