+++ docs/other/_old_docs/_examples_and_tutorials/loaders.md
@@ -1,0 +1,27 @@
+(loaders)=
+
+# Loaders
+
+```{eval-rst}
+.. include:: loaders/01_arbin.md
+```
+
+```{eval-rst}
+.. include:: loaders/02_maccor.md
+```
+
+```{eval-rst}
+.. include:: loaders/03_PEC.md
+```
+
+```{eval-rst}
+.. include:: loaders/04_Neware.md
+```
+
+```{eval-rst}
+.. include:: loaders/05_biologics.md
+```
+
+```{eval-rst}
+.. include:: loaders/06_custom.md
+```
