+++ _old_docs/main_description/index.md
@@ -1,12 +1,0 @@
-# Main description
-
-```{toctree}
-:maxdepth: 2
-
-installation.rst
-usage.rst
-formats.rst
-contributing.rst
-authors.rst
-history.rst
-```
