+++ docs/other/_old_docs/_examples_and_tutorials/utils.md
@@ -1,0 +1,33 @@
+# Using some of the cellpy special utilities
+
+```{eval-rst}
+.. include:: utils/tut_ocv_rlx.md
+```
+
+```{eval-rst}
+.. include:: utils/ica.md
+```
+
+```{eval-rst}
+.. include:: utils/batch.md
+```
+
+```{eval-rst}
+.. include:: utils/collectors.md
+```
+
+```{eval-rst}
+.. include:: utils/plotting.md
+```
+
+```{eval-rst}
+.. include:: utils/easyplot.md
+```
+
+```{eval-rst}
+.. include:: utils/templates.md
+```
+
+```{eval-rst}
+.. include:: utils/custom_file_loaders.md
+```
