+++ _old_docs/examples_and_tutorials/basics.md
@@ -1,33 +1,0 @@
-# Basics usage
-
-```{eval-rst}
-.. include:: basic_interactions/01_getting_started_tutorial.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/02_read_cell_data.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/03_more_about_get.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/04_other_interactions.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/05_configuring.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/06_pandas.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/07_data_mining.rst
-```
-
-```{eval-rst}
-.. include:: basic_interactions/08_the_cellpy_cmd.rst
-```
