+++ _old_docs/main_description/contributing.md
@@ -1,3 +1,0 @@
-```{eval-rst}
-.. include:: ../../CONTRIBUTING.rst
-```
