+++ _old_docs/developers_guide/index.md
@@ -1,14 +1,0 @@
-# Developers guide
-
-```{toctree}
-:maxdepth: 2
-
-dev_cellpy_setup.rst
-dev_cellpy_data_structure.rst
-dev_cellpy_folder_structure.rst
-dev_loaders_and_instruments.rst
-dev_docs.rst
-dev_cellpy_packaging_pypi.rst
-dev_conda_package.rst
-dev_various.rst
-```
