+++ docs/other/_old_docs/_main_description/index.md
@@ -1,0 +1,9 @@
+# Main description
+
+```{toctree}
+:maxdepth: 2
+
+installation.md
+usage.md
+formats.md
+```
