+++ docs/other/index.md
@@ -1,11 +1,12 @@
 # Other
 
 ```{toctree}
-:maxdepth: 2
+:maxdepth: 1
 
 authors.md
 history.md
 licence.md
 code-of-conduct.md
 contributing.md
+legacy.md
 ```
