+++ docs/other/_old_docs/index.md
@@ -1,0 +1,8 @@
+```{include} ./_main_description/index.md
+```
+
+```{include} ./_examples_and_tutorials/index.md
+```
+
+```{include} ./_developers_guide/index.md
+```
