+++ _old_docs/main_description/authors.md
@@ -1,3 +1,0 @@
-```{eval-rst}
-.. include:: ../../AUTHORS.rst
-```
