### 2024-02-20 0699ab7
small improvements on basic usage
