### 2024-02-20 1f68fd8
Add some text+paper figures to docs/fundamentals
