+++ docs/other/_old_docs/_examples_and_tutorials/index.md
@@ -3,10 +3,10 @@
 ```{toctree}
 :maxdepth: 2
 
-basics
-utils
-notebooks
-examples
-tips_and_tricks
-loaders
+basics.md
+utils.md
+notebooks.md
+examples.md
+tips_and_tricks.md
+loaders.md
 ```
