+++ docs/other/legacy.md
@@ -1,13 +1,10 @@
 # Legacy Documentation
 
-This is the legacy documentation for cellpy. It is not maintained and will be removed in the future.
+The legacy documentation for cellpy still exists.
+
+It is located in the `docs/other/_old_docs` folder and can be found [here](https://github.com/jepegit/cellpy/tree/273-joss-review-installation/docs/other/_old_docs).
+
+It is not maintained and will be removed in the future.
 It is kept here for the time being while we are working with the new documentation (some of it can be
 reused, we hope).
 
-```{toctree}
-:maxdepth: 3
-```
-
-
-```{include} _old_docs/index.md
-```
