+++ setup.py
@@ -12,7 +12,7 @@ import os
 
 from setuptools import find_packages, setup
 
-with open("README.rst") as readme_file:
+with open("README.md") as readme_file:
     readme = readme_file.read()
 
 with open("HISTORY.md") as history_file:
