+++ README.rst
@@ -1,70 +1,0 @@
-.. image:: https://raw.githubusercontent.com/jepegit/cellpy/master/docs/_static/cellpy-icon-long.svg
-        :height: 100
-        :alt: cellpy
-
-===================================================================
-cellpy - *a library for assisting in analysing batteries and cells*
-===================================================================
-
-
-.. image:: https://img.shields.io/pypi/v/cellpy.svg
-        :target: https://pypi.python.org/pypi/cellpy
-
-.. image:: https://readthedocs.org/projects/cellpy/badge/?version=latest
-        :target: https://cellpy.readthedocs.io/en/latest/?badge=latest
-        :alt: Documentation Status
-
-.. image:: https://static.pepy.tech/badge/cellpy
-        :target: https://pepy.tech/project/cellpy
-
-
-This Python Package was developed to help the
-researchers at IFE, Norway, in their cumbersome task of
-interpreting and handling data from cycling tests of
-batteries and cells.
-
-
-Documentation
-=============
-
-The documentation for ``cellpy`` is hosted on `Read the docs
-<https://cellpy.readthedocs.io>`_.
-
-
-Installation and dependencies
-=============================
-
-The easiest way to install ``cellpy`` is to install with conda or pip.
-
-With conda::
-
-   conda install -c conda-forge cellpy
-
-Or if you prefer installing using pip::
-
-   python -m pip install cellpy
-
-Have a look at the documentation for more detailed installation procedures, especially
-with respect to "difficult" dependencies when installing with pip.
-
-License
-=======
-
-``cellpy`` is free software made available under the MIT License.
-
-Features
-========
-
-* Convert testing data into a common format
-* Summarize and compare data
-* Filter out the steps of interest
-* Process and plot the data
-* Do simple and advanced analysis on single datasets as well as collections of datasets
-* And more!
-
-Contributing
-============
-
-Contributions are welcome, and they are greatly appreciated! Every
-little bit helps, and credit will always be given. If you want to contribute,
-please have a look at contributing guidelines.
