### 2024-02-20 9f9262e
small improvement in notification in inv task (build)
