### 2024-02-20 66b7a25
some work on fundamentals - datastructure
