+++ docs/fundamentals/data_structure.md
@@ -1,1 +1,314 @@
 # Data structure
+The most important file formats and data structures for cellpy are
+summarized here. It is also possible to look into the source-code at the
+repository <https://github.com/jepegit/cellpy>.
+
+## CellpyCell - main structure
+
+The **CellpyCell** is the main work-horse for cellpy, containing
+all the data, stored in the **Data** object, as well as all the functions
+for reading, selecting, and tweaking your data.
+It also contains the header definitions, both for the cellpy HDF5 format,
+and for the various cell-tester file-formats that can be read. The class
+can contain several tests and each test is stored in a list. The class also
+contains several attributes that can be assigned directly.
+
+
+```{eval-rst}
+.. graphviz::
+
+   digraph {
+      "CellpyCell" -> "Data";
+      "CellpyCell" -> "session metadata";
+      "CellpyCell" -> "cellpy metadata";
+      "CellpyCell" -> "methods";
+   }
+
+```
+
+## Methods
+The **CellpyCell** object contains lots of methods for manipulating, extracting
+and summarising the data from the run(s).
+The following two methods are typically automatically run upon loading your data using
+`cellpy.get(filename)` and thereby creating your **CellpyCell** object:
+
+> - `make_step_table`: creates a statistical summary of all the steps in the run(s) and categorizes
+>   the step type from that. It is also possible to give the step types directly (step_specifications).
+> - `make_summary`: create a summary based on cycle number.
+
+Other common methods worth mentioning are:
+
+> - `load`: load a cellpy file.
+> - `load_raw`: load raw data file(s) (merges automatically if several filenames are given as a list).
+> - `get_cap`: get the capacity-voltage graph from one or more cycles in three different formats as well
+>   as optionally interpolated, normalized and/or scaled.
+> - `get_cycle_numbers`: get the cycle numbers for your run.
+> - `get_ocv`: get the rest steps after each charge and discharge step.
+
+Take a look at API section (Module index, `cellpy.readers.cellreader.CellpyCell`) for more info.
+
+## Data
+The data is stored as an instance of the Data class, `CellpyCell.data`
+(a `cellpy.cellreader.Data` instance).
+
+```{eval-rst}
+.. graphviz::
+
+   digraph {
+    "CellpyCell" -> "Data";
+        "Data" -> "cell metadata (cell)";
+        "Data" -> "cell metadata (test)";
+        "Data" -> "methods";
+        "Data" -> "raw";
+        "Data" -> "steps";
+        "Data" -> "summary";
+   }
+```
+
+The Data object contains the data and the metadata for the cell characterisation experiment(s).
+
+The actual measurement data, information, and summary are stored in three `pandas.DataFrames`:
+
+> - `raw`: raw data from the run.
+> - `steps`: stats from each step (and step type), created using the `CellpyCell.make_step_table` method.
+> - `summary`: summary data vs. cycle number (e.g. coulombic efficiency), created using the `CellpyCell.make_summary` method.
+
+For details on column headings, see below.
+
+#### Metadata
+The Data object contains the following metadata:
+
+```python
+cell_no = None
+mass = prms.Materials.default_mass  # active material (in mg)
+tot_mass = prms.Materials.default_mass  # total material (in mg)
+no_cycles = 0.0
+charge_steps = None
+discharge_steps = None
+ir_steps = None
+ocv_steps = None
+nom_cap = prms.DataSet.nom_cap  # mAh/g (for finding c-rates)
+mass_given = False
[…]
