+++ docs/fundamentals/fundamentals.md
@@ -1,7 +1,6 @@
 # The fundamentals of cellpy
 `cellpy` is implemented in Python and can be used as either a library within Python scripts, or as a stand-alone application for analysing battery cell test data. Internally, `cellpy` utilises the rich ecosystem of scientific tools available for Python. In particular, `cellpy` uses `pandas` DataFrames as the “storage containers” for the collected data within the `cellpy` Data object. This offers full flexibility and makes it easy for the user to apply advanced methods, analyses of or transformations to the data in addition to the features implemented in `cellpy`.
 
-## The CellpyCell
 The core of `cellpy` is the **CellpyCell** object (see \autoref{fig:2}) that contains both the data (stored in the **Data** object) as well as central methods required to read, process and store battery testing data. The CellpyCell provides the appropriate interface and coordination of the resources needed, such as loading configurations (*e.g* default reader, default raw-data location), selecting readers for different data formats and exporters for saving the data.
 
 ![Illustration of the core object within ``cellpy``, the **CellpyCell**.\label{fig:2}](..\figures\CellpyCell.jpg)
@@ -14,8 +13,6 @@ The most common data processing routines, such as extraction of charge/discharge
 
 ![The `cellpy` library contains multiple utilities that assists in data analysis. A utility can work on (A) a single **CellpyCell** object, or (B) a set of CellpyCell objects such as the Batch utility that helps the user in automating and comparing results from many data sets.\label{fig:4}](Figures/Cellpy-Utils.jpg)
 
-
-## File Formats
 The `cellpy`-file format (usually stored in HDF5 format) contains all the data contained in the Data object together with additional relevant metadata, including information about the file version.
 
 (Ref: [paper.md](https://github.com/jepegit/cellpy/tree/master/paper))
