### 2024-02-20 c325fbb
added notebook on loading .res data, still need to link to it from index
