+++ docs/examples/index.md
@@ -4,4 +4,5 @@
 :maxdepth: 2
 
 reading_data.md
+01_loading_data.ipynb
 ```
