### 2024-02-20 e0eb568
Include info on installing on POSIX
