+++ docs/getting_started/installation.md
@@ -1,15 +1,12 @@
-```{highlight} shell
-```
-
 # Installation
 
-`cellpy` is available on [Windows](#cellpy_install_windows) and [Linux](#cellpy_install_linux) and can be installed using `pip`
-or `conda`, or [installed from sources](#cellpy_install_sources).
+`cellpy` is available on {ref}`Windows <cellpy_install_windows>` and {ref}`Linux <cellpy_install_linux>` and can be
+installed using `pip` or `conda`, or {ref}`installed from sources <cellpy_install_sources>`.
 
-For more details on dependencies, have a look at [cellpy dependencies](#cellpy_dependencies).
+For more details on dependencies, have a look at {ref}`cellpy dependencies <cellpy_dependencies>`.
 
-For a tea-spoon explanation on how to install `cellpy` on Windows, including the
-installation of Python and setup of virtual environments, [see below](#cellpy_setup_teaspoon).
+For a teaspoon explanation on how to install `cellpy` on Windows, including the
+installation of Python and setup of virtual environments, {ref}`see below <cellpy_setup_teaspoon>`.
 
 After installing `cellpy`, continue to
 - [Setup and configuration](configuration.md)
@@ -22,21 +19,21 @@ After installing `cellpy`, continue to
 The easiest way to install cellpy is by using conda:
 
 ```console
-$ conda install -c conda-forge cellpy
+conda install -c conda-forge cellpy
 ```
 
-This will also install all of the critical dependencies, as well as `jupyter`
+This will also install all the critical dependencies, as well as `jupyter`
 that comes in handy when working with cellpy.
 
-In general, we recommend to install cellpy in a virtual environment (if
-you do not know what this means, have a look the [teaspoon explanation](#cellpy_setup_teaspoon)).
+In general, we recommend to install cellpy in a virtual environment (if you do not know what this means,
+have a look the at the {ref}`teaspoon explanation <cellpy_setup_teaspoon>`).
 
 ### Pip
 
 If you would like to install 'only' cellpy, you can use pip:
 
 ```console
-$ pip install cellpy
+pip install cellpy
 ```
 
 Note that `cellpy` uses several packages that are a bit cumbersome to install
@@ -48,27 +45,70 @@ based on the provided [environment.yml](https://github.com/jepegit/cellpy/blob/m
 :::
 
 (cellpy_install_linux)=
-## Installation on Linux
+## Installation on Linux (and macOS)
+
+It is especially recommended to install `cellpy` in a virtual environment on Linux and macOS.
 
 ### Conda
 
-This is how to install `cellpy` using conda on Linux.
-Be aware that you might have to install additional packages.
-
-This includes, e.g., `libobdc`, that can be installed on Ubuntu like this:
+To be able to get a fully functional installation of `cellpy` on Linux using`conda`,
+you might need to install some additional packages. For parsing .res files, you must have
+the `mdbtools` installed. If it is missing, you install it with the Linux package manager, for example:
 
 ```console
-$ sudo apt update
-$ sudo apt install unixodbc-dev
+sudo apt update
+sudo apt-get install -y mdbtools
+```
+
+For macOS, you can use `brew`:
+
+```console
+brew install mdbtools
+```
+
+Then you can install `cellpy` using conda:
+
+```console
+conda install -c conda-forge cellpy
 ```
 
 ### Pip
 
-This is how to install `cellpy` using pip on Linux.
-Be aware that you might have to install additional packages.
+Installing `cellpy` using pip on Linux might be a bit more cumbersome than using conda.
+A recipe that seems to work well in most cases is to first install the `mdbtools` (and possibly also `unixodbc`)
+and then install `cellpy` using pip. For example, on Ubuntu, you can do this by running the following commands
+in a terminal:
+
+```console
[…]
