### 2024-02-23 jepegit/cellpy#273 @juliawind
Installation notes updated and summarized all in one place.
