### 2024-02-23 5ecb159
tests fail on linux (test_sql) - claims it cant import pyodbc. Hard to understand why. Trying to pip install pyodbc instead (changing the github_actions_environment.yml so it will affect all the pytest actions)
