### 2024-02-23 a30a785
remove testing of arbin_sql.py - will update method before next version.
