### 2024-02-23 9f2bcab
many other tests need pyodbc - so - did not fix it. Moving test_sql.py back into tests.
