+++ tests/test_sql.py
@@ -2,7 +2,6 @@ import pandas as pd
 import pytest
 
 from cellpy import log
-from cellpy.readers.instruments import arbin_sql
 from cellpy.readers.instruments import base
 from cellpy.readers.core import Data
 
@@ -17,11 +16,13 @@ def raw_mock(parameters):
 
 
 def test_import_arbin_sql():
+    from cellpy.readers.instruments import arbin_sql
     loader = arbin_sql.DataLoader()
     assert isinstance(loader, base.BaseLoader)
 
 
 def test_post_process_rename_headers_defined():
+    from cellpy.readers.instruments import arbin_sql
     keywords = {
         "fix_datetime": False,
         "set_index": False,
@@ -39,6 +40,7 @@ def test_post_process_rename_headers_defined():
 
 
 def test_post_process_rename_headers_from_file(raw_mock):
+    from cellpy.readers.instruments import arbin_sql
     keywords = {
         "fix_datetime": True,
         "set_index": True,
