### 2024-02-23 jepegit/cellpy#277 @jepegit
Closing it for now - will reopen it soon when we start making the necessary changes in `OtherPaths` for supporting py3.12
