### 2024-02-23 8d1666b
remove old examples and notebooks and start to update the cellpy_batch_processing.ipynb notebook
