+++ examples/simple_arbin_reader_no_cellpy.py
@@ -1,218 +1,0 @@
-# -*- coding: utf-8 -*-
-"""simple script for reading .res-files from arbin
-
-This script does not rely on any of the modules in cellpy.
-"""
-
-import os
-import shutil
-import sys
-import tempfile
-import time
-
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-
-try:
-    import pyodbc as dbloader
-except ImportError:
-    print("could not import dbloader (pyodbc)")
-    print("this script will not work without")
-    sys.exit(0)
-
-
-def convert2mAhg(d, mass=0.02):
-    """
-
-    Args:
-        d: capacity
-        mass (float): mass in mg
-
-    Returns:
-        np.array with capacity in mAh/g
-    """
-    d = np.array(d)
-    return d * 1000000 / mass
-
-
-def main():
-    # This script relies on using Microsoft Access Driver.
-    # If you are on a non-Microsoft OS, this will not work.
-    # Even on a Windows machine, you will need to make sure that
-    # the appropriate driver is installed (same bit as your python)
-
-    # write file name here:
-    filename = r"..\testdata\data\20160805_test001_45_cc_01.res"
-    # write min value for capacity (points less than that will be removed):
-    lim_low = 10
-    # write max value for capacity (points larger than that will be removed):
-    lim_high = 6050
-    # mass of (active material of) electrode in mg:
-    mass = 0.8369
-    # write path to were you want to save the data:
-    outdir = r"..\.."
-
-    tablename_normal = "Channel_Normal_Table"
-    tablename_global = "Global_Table"
-    tablename_statistic = "Channel_Statistic_Table"
-    fname = os.path.basename(filename)
-    fname = os.path.splitext(fname)[0]
-    outfile = os.path.join(outdir, fname + "_out.csv")
-    outfile2 = os.path.join(outdir, fname + "_cdc.csv")
-
-    print("filename:"), filename
-
-    # ------making temporary file-------------
-    temp_dir = tempfile.gettempdir()
-    temp_filename = os.path.join(temp_dir, os.path.basename(filename))
-    print("Copying to tmp-file")
-    print("temp_filename:"), temp_filename
-
-    shutil.copy2(filename, temp_dir)
-    print("Finished to tmp-file")
-
-    constructor = (
-        "Driver={Microsoft Access Driver (*.mdb, *.accdb)};Dbq=" + temp_filename
-    )
-    conn = dbloader.connect(constructor, autocommit=True)
-    cur = conn.cursor()
-    sql = "select * from %s ORDER BY Test_Time" % tablename_normal
-    cur.execute(sql)
-    col_names = [i[0] for i in cur.description]
-
-    # all_data=cur.fetchall()
-    print("COLS:")
-    for cname in col_names:
-        print(cname)
-
-    print("\nreading file")
-
-    # print all_data
-    t = time.time()
-    state_counter = 0
-    limit_counter = 0
-    limit = 11213440
-    # limit = 100000
-
-    V = []  # "Voltage"
-    T = []  # "Test_Time"
-    D = []  # "Discharge_Capacity"
-    C = []  # "Charge_Capacity"
-    I = []  # "Cycle_Index"
-
-    step = 0
-    cycle = 0
-    v = 0.0
-    t = 0.0
-    d = 0.0
-    c = 0.0
-
-    ofile = open(outfile, "w")
-    header = "time;cycle;step;voltage;discharge_cap;charge_cap\n"
-    ofile.write(header)
-
-    for row in cur:
-        if not row:
-            break
-
-        limit_counter += 1
-        if limit_counter >= limit:
-            print("x"),
-            break
-
-        state_counter += 1
-        # getting the cycle and step index
[…]
