### 2024-02-23 00adc79
add pyarrow as dependency
