### 2024-02-23 1d0fa9e
small change to cycle_info_plot to not require input of cycle number
