### 2024-02-23 f808c69
try fixing missing plotly figures in docs and use book theme
