### 2024-02-24 2fe5dca
remove test notebook
