### 2024-02-24 4d23425
tried to use myst_nb again - but failed (not able to get the plotly figures alive).
