+++ docs/other/index.md
@@ -4,8 +4,8 @@
 :maxdepth: 1
 
 authors.md
-history.md
-licence.md
+project-history.md
+project-license.md
 code-of-conduct.md
 contributing.md
 legacy.md
