+++ docs/other/code-of-conduct.md
@@ -1,4 +1,3 @@
 
-
 ```{include} ../../CODE_OF_CONDUCT.md
 ```
