+++ docs/conf.py
@@ -9,9 +9,6 @@ version_ns = {}
 with open(os.path.join(project_root, name, "_version.py")) as f:
     exec(f.read(), {}, version_ns)
 
-
-# -- Path setup -------------------------------------------------------
-
 if INCLUDE_SOURCE_FILES:
     project_prmsdir = os.path.join(project_root, r"cellpy\parameters")
     project_utils = os.path.join(project_root, r"cellpy\utils")
@@ -26,10 +23,6 @@ if INCLUDE_SOURCE_FILES:
     sys.path.insert(0, project_readers)
     sys.path.insert(0, project_internals)
 
-
-# -- General configuration ---------------------------------------------
-
-# Extensions
 extensions = [
     "sphinx.ext.inheritance_diagram",
     "sphinx.ext.viewcode",
@@ -44,14 +37,13 @@ extensions = [
 ]
 
 # Note about myst_nb:
-# It is possible to use myst_nb (wich can parse notebooks with myst markdown)
+# It is possible to use myst_nb (that can parse notebooks with myst markdown)
 # instead of nbsphinx, but at the moment it is difficult to get the plotly plots to
 # work properly. Setting plotly to make the plots in iframes is a workaround, but
 # this will only work smoothly if the jupyter notebook are in the root of the documentation
 # folder. If not, the paths to the plots (iframes) will be wrong (and you will have to copy the folder with
 # the iframes somehow to the same location as the built notebooks).
 
-# Napoleon settings
 napoleon_google_docstring = True
 napoleon_numpy_docstring = True
 napoleon_include_init_with_doc = False
@@ -64,42 +56,14 @@ napoleon_use_ivar = False
 napoleon_use_param = True
 napoleon_use_rtype = True
 napoleon_use_keyword = True
-
-# Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
-
-# The suffix of source file_names.
 source_suffix = ".rst"
-
-# The master toctree document.
 master_doc = "index"
-
-# General information about the project.
 project = "cellpy"
 copyright = "2024, Jan Petter Maehlen"
-
-# The version info for the project you're documenting, acts as replacement
-# for |version| and |release|, also used in various other places throughout
-# the built documents.
-#
-# The short X.Y version.
 version = version_ns["__version__"]
-# The full version, including alpha/beta/rc tags.
 release = version_ns["__version__"]
 
-
-# The language for content autogenerated by Sphinx. Refer to documentation
-# for a list of supported languages.
-# language = None
-
-# There are two options for replacing |today|: either, you set today to
-# some non-false value, then it is used:
-# today = ''
-# Else, today_fmt is used as the format for a strftime call.
-# today_fmt = '%B %d, %Y'
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
 exclude_patterns = [
     "_build",
     "_autoapi_templates",
@@ -107,119 +71,17 @@ exclude_patterns = [
     "examples/.ipnb_checkpoints",
 ]
 
-# The reST default role (used for this markup: `text`) to use for all
-# documents.
-# default_role = None
-
-# If true, '()' will be appended to :func: etc. cross-reference text.
 add_function_parentheses = True
-
-# If true, the current module name will be prepended to all description
-# unit titles (such as .. function::).
 add_module_names = False
-
-# If true, sectionauthor and moduleauthor directives will be shown in the
-# output. They are ignored by default.
-# show_authors = False
-
-# The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
-
-# A list of ignored prefixes for module index sorting.
 modindex_common_prefix = ["cellpy."]
 
[…]
