### 2024-02-25 50c21c0
clean up api docs a little bit (but not completely)
