+++ tests/test_cell_readers.py
@@ -302,11 +302,6 @@ def test_get_ir(dataset):
     dataset.get_ir()
 
 
-@pytest.mark.xfail(raises=DeprecatedFeature)
-def test_get_diagnostics_plot(dataset):
-    dataset.get_diagnostics_plots()
-
-
 def test_check64bit():
     a = cellpy.readers.core.check64bit()
     b = cellpy.readers.core.check64bit("os")
