+++ cellpy/readers/cellreader.py
@@ -103,14 +103,12 @@ module_logger = logging.getLogger(__name__)
 class CellpyCell:
     """Main class for working and storing data.
 
-    This class is the main work-horse for cellpy where all the functions for
+    This class is the main work-horse for cellpy where methods for
     reading, selecting, and tweaking your data is located. It also contains the
     header definitions, both for the cellpy hdf5 format, and for the various
-    cell-tester file-formats that can be read. The class can contain
-    several cell-tests and each test is stored in a list. If you see what I mean...
+    cell-tester file-formats that can be read.
 
     Attributes:
-        # TODO v.1.0.1: update this
         data: cellpy.Data object
         cellpy_units: cellpy.units object
         cellpy_datadir: path to cellpy data directory
@@ -218,8 +216,7 @@ class CellpyCell:
         output_units=None,
         debug=False,
     ):
-        """CellpyCell object
-
+        """
         Args:
             filenames: list of files to load.
             selected_scans:
@@ -1044,7 +1041,7 @@ class CellpyCell:
         selector=None,
         **kwargs,
     ):
-        """Loads data for given cells.
+        """Loads data for given cells (soon to be deprecated).
 
         Args:
             raw_files (list): name of res-files
@@ -1309,11 +1306,11 @@ class CellpyCell:
                 return False
         return v
 
-    def partial_load(self, **kwargs):
+    def _partial_load(self, **kwargs):
         """Load only a selected part of the cellpy file."""
         raise NotImplementedError
 
-    def link(self, **kwargs):
+    def _link(self, **kwargs):
         """Create a link to a cellpy file.
 
         If the file is very big, it is sometimes better to work with the data
@@ -2589,10 +2586,18 @@ class CellpyCell:
     def load_step_specifications(self, file_name, short=False):
         """Load a table that contains step-type definitions.
 
-        This function loads a file containing a specification for each step or
-        for each (cycle_number, step_number) combinations if short==False. The
-        step_cycle specifications that are allowed are stored in the variable
-        cellreader.list_of_step_types.
+        This method loads a file containing a specification for each step or
+        for each (cycle_number, step_number) combinations if `short==False`, and
+        runs the `make_step_table` method. The step_cycle specifications that
+        are allowed are stored in the variable `cellreader.list_of_step_types`.
+
+        Args:
+            file_name (str): name of the file to load
+            short (bool): if True, the file only contains step numbers and
+                step types. If False, the file contains cycle numbers as well.
+
+        Returns:
+            None
         """
 
         # if short:
@@ -2659,13 +2664,13 @@ class CellpyCell:
 
         The format of the steps is:
 
-            index: cycleno - stepno - sub-step-no - ustep
-            Time info: average, stdev, max, min, start, end, delta
-            Logging info: average, stdev, max, min, start, end, delta
-            Current info: average, stdev, max, min, start, end, delta
-            Voltage info: average,  stdev, max, min, start, end, delta
-            Type: (from pre-defined list) - SubType
-            Info: not used.
+        - index: cycleno - stepno - sub-step-no - ustep
+        - Time info: average, stdev, max, min, start, end, delta
+        - Logging info: average, stdev, max, min, start, end, delta
+        - Current info: average, stdev, max, min, start, end, delta
+        - Voltage info: average,  stdev, max, min, start, end, delta
+        - Type: (from pre-defined list) - SubType
+        - Info: not used.
 
         Args:
             step_specifications (pandas.DataFrame): step specifications
@@ -3830,6 +3835,10 @@ class CellpyCell:
         logging.debug(f"(dt: {(time.time() - time_00):4.2f}s)")
 
     def get_mass(self):
+        """Returns the mass of the active material (in mg).
+
+        This method will be deprecated in the future.
+        """
[…]
