+++ cellpy/parameters/prms.py
@@ -53,6 +53,8 @@ class CellPyConfig:
 # This can stay global:
 @dataclass
 class PathsClass(CellPyConfig):
+    """Paths used in cellpy."""
+
     outdatadir: Union[Path, str] = wdir
     _rawdatadir: Union[OtherPath, str] = op_wdir
     _cellpydatadir: Union[OtherPath, str] = op_wdir
@@ -85,6 +87,8 @@ class PathsClass(CellPyConfig):
 
 @dataclass
 class BatchClass(CellPyConfig):
+    """Settings for batch processing."""
+
     auto_use_file_list: bool = False
     template: str = "standard"
     fig_extension: str = "png"
@@ -104,6 +108,8 @@ class BatchClass(CellPyConfig):
 
 @dataclass
 class FileNamesClass(CellPyConfig):
+    """Settings for file names and file handling."""
+
     file_name_format: str = "YYYYMMDD_[NAME]EEE_CC_TT_RR"
     raw_extension: str = "res"
     reg_exp: str = None
@@ -116,6 +122,8 @@ class FileNamesClass(CellPyConfig):
 
 @dataclass
 class ReaderClass(CellPyConfig):
+    """Settings for reading data."""
+
     diagnostics: bool = False
     filestatuschecker: str = "size"
     force_step_table_creation: bool = True
@@ -141,6 +149,7 @@ class ReaderClass(CellPyConfig):
 
 @dataclass
 class DbClass(CellPyConfig):
+    """Settings for the handling the simple database."""
     db_type: str = "simple_excel_reader"
     db_table_name: str = "db_table"  # used for simple excel db reader
     db_header_row: int = 0  # used for simple excel db reader
@@ -155,6 +164,7 @@ class DbClass(CellPyConfig):
 
 @dataclass
 class DbColsClass(CellPyConfig):  # used for simple excel db reader
+    """Names of the columns in the simple database."""
     # Note to developers:
     #  1) This is ONLY for the excel-reader (dbreader.py)! More advanced
     #     readers should get their own way of handling the db-columns.
@@ -202,6 +212,7 @@ class DbColsClass(CellPyConfig):  # used for simple excel db reader
 
 @dataclass
 class DbColsUnitClass(CellPyConfig):
+    """Unit of the columns in the simple database."""
     # Note to developers:
     #  1) This is ONLY for the excel-reader (dbreader.py)! More advanced
     #     readers should get their own way of handling the db-columns.
@@ -296,6 +307,7 @@ Batch = BatchClass(backend="plotly")
 # remark! using box.Box for each instrument
 @dataclass
 class InstrumentsClass(CellPyConfig):
+    """Settings for the instruments."""
     tester: Union[str, None]
     custom_instrument_definitions_file: Union[str, None]
     Arbin: box.Box
