+++ cellpy/readers/core.py
@@ -3,6 +3,7 @@
 It also contains functions that are used by readers and utils.
 And it has the file version definitions.
 """
+
 import abc
 import datetime
 import importlib
