+++ cellpy/parameters/internal_settings.py
@@ -1,4 +1,5 @@
 """Internal settings and definitions and functions for getting them."""
+
 import logging
 import warnings
 from collections import UserDict
@@ -155,69 +156,68 @@ class CellpyMetaCommon(CellpyMeta):
     # about cell
     material: Optional[prms.CellPyDataConfig] = prms.Materials.default_material
     # TODO @jepe: Maybe we should use values with units here instead (pint)?
-    mass: Optional[
-        prms.CellPyDataConfig
-    ] = prms.Materials.default_mass  # active material
-    tot_mass: Optional[
-        prms.CellPyDataConfig
-    ] = prms.Materials.default_mass  # total material
+    mass: Optional[prms.CellPyDataConfig] = (
+        prms.Materials.default_mass
+    )  # active material
+    tot_mass: Optional[prms.CellPyDataConfig] = (
+        prms.Materials.default_mass
+    )  # total material
     # volume: Optional[prms.CellPyDataConfig] = prms.CellInfo.volume
-    nom_cap: Optional[
-        prms.CellPyDataConfig
-    ] = prms.Materials.default_nom_cap  # nominal capacity   # used as property
-    nom_cap_specifics: Optional[
-        prms.CellPyDataConfig
-    ] = (
+    nom_cap: Optional[prms.CellPyDataConfig] = (
+        prms.Materials.default_nom_cap
+    )  # nominal capacity   # used as property
+    nom_cap_specifics: Optional[prms.CellPyDataConfig] = (
         prms.Materials.default_nom_cap_specifics
     )  # nominal capacity type  # used as property
 
-    active_electrode_area: Optional[
-        prms.CellPyDataConfig
-    ] = prms.CellInfo.active_electrode_area
-    active_electrode_thickness: Optional[
-        prms.CellPyDataConfig
-    ] = prms.CellInfo.active_electrode_thickness
-    electrolyte_volume: Optional[
-        prms.CellPyDataConfig
-    ] = prms.CellInfo.electrolyte_volume
+    active_electrode_area: Optional[prms.CellPyDataConfig] = (
+        prms.CellInfo.active_electrode_area
+    )
+    active_electrode_thickness: Optional[prms.CellPyDataConfig] = (
+        prms.CellInfo.active_electrode_thickness
+    )
+    electrolyte_volume: Optional[prms.CellPyDataConfig] = (
+        prms.CellInfo.electrolyte_volume
+    )
 
     electrolyte_type: Optional[prms.CellPyDataConfig] = prms.CellInfo.electrolyte_type
-    active_electrode_type: Optional[
-        prms.CellPyDataConfig
-    ] = prms.CellInfo.active_electrode_type
-    counter_electrode_type: Optional[
-        prms.CellPyDataConfig
-    ] = prms.CellInfo.counter_electrode_type
-    reference_electrode_type: Optional[
-        prms.CellPyDataConfig
-    ] = prms.CellInfo.reference_electrode_type
+    active_electrode_type: Optional[prms.CellPyDataConfig] = (
+        prms.CellInfo.active_electrode_type
+    )
+    counter_electrode_type: Optional[prms.CellPyDataConfig] = (
+        prms.CellInfo.counter_electrode_type
+    )
+    reference_electrode_type: Optional[prms.CellPyDataConfig] = (
+        prms.CellInfo.reference_electrode_type
+    )
     experiment_type: Optional[prms.CellPyDataConfig] = prms.CellInfo.experiment_type
     cell_type: Optional[prms.CellPyDataConfig] = prms.CellInfo.cell_type
     separator_type: Optional[prms.CellPyDataConfig] = prms.CellInfo.separator_type
-    active_electrode_current_collector: Optional[
-        prms.CellPyDataConfig
[…]
