### 2024-02-25 c4c1193
more api docs cleanup
