### 2024-02-26 1b17821
add nbmake test in actions on cellpy_batch_processing.ipynb
