### 2024-02-26 ac10840
add nbmake test in actions on cellpy_batch_processing.ipynb with named kernel
