### 2024-02-26 6a960b0
include registering conda env using ipykernel
