### 2024-02-26 8ce784c
add plotly as dependency in github_actions_environment.yml
