+++ github_actions_environment.yml
@@ -34,6 +34,7 @@ dependencies:
     - pytest-timeout
     - pytest-benchmark
     - nbmake
+    - plotly
     - pip:
         - sqlalchemy-access
         - jinja2_time
