### 2024-02-26 b45337e
update cycling summary plotting test (renamed method)
