+++ github_actions_environment.yml
@@ -12,7 +12,7 @@ dependencies:
     - cryptography
     - click
     - scipy
-    - pandas
+    - pandas <=2.1
     - setuptools
     - ruamel.yaml
     - pyodbc
