### 2024-02-26 48c9aff
started on dqdv notebook - work in progress
