### 2024-02-26 4b6e7f5
hack to allow for new pandas (but not unpinning yet) and add installing cellpy within action container (posix)
