### 2024-02-26 cdc0eeb
set more prms in notebook, update .gitignore (cellpy-files etc) and include kaleido in requirements (was missing in the conda envs).
