### 2024-02-26 1f6cca1
remove kaleido==0.1.* in github_actions_environment.yml (does not exist for macOS M-chip)
