+++ environment.yml
@@ -33,6 +33,7 @@ dependencies:
     - ipykernel
     - rich
     - sqlalchemy-access
+# This version of kaleido is not available for macOS M-chip:
     - kaleido==0.1.*
     - jinja2_time
 
