+++ github_actions_environment.yml
@@ -38,7 +38,6 @@ dependencies:
     - pip:
         - sqlalchemy-access
         - jinja2_time
-        - kaleido==0.1.*
 
 
 
