### 2024-02-26 72ee3e4
start implementing better example data solution
