### 2024-02-26 05d33c6
start implementing better example data solution
