### 2024-02-26 35efeca
start implementing better example data solution
