### 2024-02-26 aa76c9c
start implementing better example data solution
