### 2024-02-26 4297b45
start implementing better example data solution
