### 2024-02-27 2520100
update names and docstrings for dqdv functions
