### 2024-02-27 jepegit/cellpy#274 @jepegit
Summary of tasks:

- [ ] Remove the examples from the basic usage page and split them into smaller pages related to their topic/example.

- [ ] Add a clear example documentation per type of example with a link to the correct example script (jupyter notebook) with the correct example input data (.res file) to test the scripts.

- [ ] In the example documentation, make a section dedicated to the datafiles that can be used for different battery testers so that the user knows how to use the software for their data.
