### 2024-02-27 c6be8cf
fix bug in type hints for example_data.py
