### 2024-02-27 e177425
fix bug in type hints for example_data.py
