+++ cellpy/parameters/prms.py
@@ -356,6 +356,9 @@ Instruments = InstrumentsClass(
 # Other secret- or non-config (only for developers)
 # ------------------------------------------------------------------------------
 
+_url_example_cellpy_data = (
+    "https://raw.githubusercontent.com/jepegit/cellpy/master/examples/data/"
+)
 _db_cols_unit = DbColsUnitClass()
 _debug = False
 _variable_that_is_not_saved_to_config = "Hei"
