+++ requirements.txt
@@ -23,3 +23,4 @@ fabric
 dateparser
 ipykernel
 rich
+requests
