+++ cellpy/utils/example_data.py
@@ -1,10 +1,12 @@
 """Tools for getting some data to play with"""
 
+from enum import Enum
 import logging
 import os
 from pathlib import Path
 
 import requests
+from tqdm.auto import tqdm
 
 import cellpy
 from cellpy import prms
@@ -12,6 +14,21 @@ from cellpy import prms
 logging.info("Ready to help you to get some data to play with.")
 CURRENT_PATH = Path(os.path.dirname(os.path.realpath(__file__)))
 DATA_PATH = CURRENT_PATH / "data"
+DO_NOT_REMOVE_THESE_FILES = [".gitkeep"]
+
+
+class ExampleData(Enum):
+    """Enum for example data files"""
+
+    ARBIN = "20160805_test001_45_cc_01.res"
+    CELLPY = "20180418_sf033_4_cc.h5"
+    OLD_CELLPY = "20160805_test001_45_cc.h5"
+    RATE = "20231115_rate_cc.h5"
+    AUX_MULTI_X = "aux_multi_x.res"
+    # BIOL_MPR = "biol.mpr"
+    # MACCOR_TXT = "maccor.txt"
+    # NEWARE_CSV = "neware.csv"
+    PEC_CSV = "pec.csv"
 
 
 def download_file(url, local_filename):
@@ -25,7 +42,13 @@ def download_file(url, local_filename):
     with requests.get(url, stream=True) as r:
         r.raise_for_status()
         with open(local_filename, "wb") as f:
+            if prms._url_example_data_download_with_progressbar:
+                pbar = tqdm(
+                    total=int(r.headers["Content-Length"]), unit="B", unit_scale=True
+                )
             for chunk in r.iter_content(chunk_size=8192):
+                if chunk and prms._url_example_data_download_with_progressbar:
+                    pbar.update(len(chunk))
                 f.write(chunk)
 
 
@@ -34,6 +57,32 @@ def _download_if_missing(filename: str) -> Path:
     if not p.is_file():
         _download_example_data(filename)
     return p
+
+
+def _remove_file(filename: str):
+    p = DATA_PATH / filename
+    logging.debug(f"Removing file: {p}")
+    p.unlink(missing_ok=False)
+
+
+def _remove_all_files():
+    for f in DATA_PATH.glob("*"):
+        logging.debug(f"Removing file: {f}")
+        if f.name not in DO_NOT_REMOVE_THESE_FILES:
+            f.unlink(missing_ok=True)
+            logging.debug(f"{f.name} removed")
+        else:
+            logging.debug(f"{f.name} not removed (protected)")
+
+
+def _is_downloaded(filename: str) -> bool:
+    p = DATA_PATH / filename
+    return p.is_file()
+
+
+def _download_all_files():
+    for f in ExampleData:
+        _download_example_data(f.value)
 
 
 def _download_example_data(filename: str):
@@ -47,7 +96,7 @@ def _download_example_data(filename: str):
 
     """
     logging.info(f"{filename} not found. Trying to access it from GitHub...")
-    base_url = prms._url_example_cellpy_data
+    base_url = prms._url_example_data
     if not os.path.exists(DATA_PATH):
         raise FileNotFoundError(f"Could not find {DATA_PATH}")
 
@@ -55,6 +104,11 @@ def _download_example_data(filename: str):
     download_file(base_url + filename, os.path.join(DATA_PATH, filename))
 
     logging.debug("File downloaded successfully.")
+
+
+def download_all_files():
+    """Download all example data files from the cellpy-data repository."""
+    _download_all_files()
 
 
 def raw_file(
@@ -91,60 +145,51 @@ def cellpy_file(testing: bool = False) -> cellpy.cellreader.CellpyCell:
     return cellpy.get(filename=file_path, testing=testing)
 
 
-@property
[…]
