### 2024-02-27 c6598b4
add (commented out) example data enums not existing yet
