### 2024-02-28 e71223b
save example data in experiment folder as first option instead of in the cellpy package folder
