### 2024-02-28 4bef4de
save example data in experiment folder as first option instead of in the cellpy package folder
