### 2024-02-28 190dbfe
save example data in experiment folder as first option instead of in the cellpy package folder
