+++ docs/environment_rtd.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python >=3.9
+  - python >=3.10
   - pip
   - Sphinx
   - pandoc
