### 2024-02-28 3d2df9e
remove Py3.9 CI tests in GitHub actions
