### 2024-02-28 56d625a
remove Py3.9 CI tests in Appveyor
