+++ HISTORY.md
@@ -2,7 +2,8 @@
 
 ## 1.0.1 (under development)
 
-* General: supports only python 3.9 and up to 3.11
+* Utils: example_data now includes auto-download of example data
+* General: supports only python 3.10 and up to 3.11
 * Batch: `naked` and `init(empty=True)` easier method for creating batch with empty pages
 * File handling: new fix in `find_files`
 * Batch / Utils: refactored and updated `Collectors` (using `plotly`)
