### 2024-02-28 9d93f63
add batch processing incl data to the docs
