+++ tests/test_batch.py
@@ -342,6 +342,7 @@ def test_interact_with_cellpydata_get_cap(updated_cycling_experiment, parameters
     assert len(capacity_voltage_df) == 1105
 
 
+@pytest.mark.skip(reason="shaky test - fails sometimes on appveyor")
 def test_cycling_summary_plotter(populated_batch):
     populated_batch.combine_summaries()
     populated_batch.plot()
