+++ cellpy/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.1b3"
+__version__ = "1.0.1b4"
