### 2024-02-28 23893c9
294 move notebooks out of documentation (#298)
* typo in notebook

* include png rendering in example notebooks

* move contributing from other into its own section

* add batch processing incl data to the docs

* copy notebooks from docs into examples
