+++ docs/other/index.md
@@ -7,6 +7,5 @@ authors.md
 project-history.md
 project-license.md
 code-of-conduct.md
-contributing.md
 legacy.md
 ```
