+++ docs/examples/index.md
@@ -1,5 +1,12 @@
 # Examples & tutorials
 
+Here we provide a few basic examples to get you started with cellpy - from reading the data to creating some basic plots. These notebooks, the used datafiles and a little more can be found in the [examples
+folder](https://github.com/jepegit/cellpy/tree/master/examples) in the cellpy GitHub repository.
+
+To exploit the full potential of cellpy, have a look at the [API](/autoapi/index).
+
+Contributions to more example notebooks are of course very welcome!
+
 ```{toctree}
 :maxdepth: 2
 
@@ -9,4 +16,5 @@
 03_capacity_vs_voltage.ipynb
 04_incremental_capacity_analysis.ipynb
 05_GITT.ipynb
+batch_utility/cellpy_batch_processing_docs.ipynb
 ```
