+++ docs/index.md
@@ -13,7 +13,7 @@
 getting_started/index.md
 fundamentals/index.md
 examples/index.md
-contributing/index.md
+contributing/contributing.md
 other/index.md
 ```
 
