+++ docs/contributing/contributing.md
@@ -1,0 +1,3 @@
+
+```{include} ../../CONTRIBUTING.md
+```
