+++ docs/other/contributing.md
@@ -1,3 +1,0 @@
-
-```{include} ../../CONTRIBUTING.md
-```
