+++ docs/contributing/developers_guide/dev_cellpy_folder_structure.md
@@ -1,0 +1,148 @@
+# Structure of the `cellpy` package
+
+## Folder structure
+
+The `cellpy` repository is structured as follows:
+
+```
+📂 cellpy               # the main folder for the cellpy repository
+├── 📂 .github          # github specific files (e.g. github actions)
+├── 📂 bin              # binary files (mdbtools for win)
+├── 📂 cellpy           # the main folder for the cellpy package
+├── 📂 docs             # the main folder for the cellpy documentation
+├── 📂 examples
+├── 📂 test_journal     # data etc for the tests
+├── 📂 testdata         # data etc for the tests
+├── 📂 tests            # here are the tests
+├── 📄 .coverage
+├── 📄 .env_example
+├── 📄 .gitattributes
+├── 📄 .gitignore
+├── 📄 .readthedocs.yaml
+├── 📄 bumpver.toml
+├── 📄 appveyor.yml
+├── 📄 AUTHORS.rst        <-- picked up by sphinx (in docs)
+├── 📄 README.rst         <-- picked up by sphinx (in docs)
+├── 📄 CODE_OF_CONDUCT.md
+├── 📄 CONTRIBUTING.rst   <-- picked up by sphinx (in docs)
+├── 📄 HISTORY.rst        <-- picked up by sphinx (in docs)
+├── 📄 LICENSE            <-- picked up by sphinx (in docs)
+├── 📄 MANIFEST.in
+├── 📄 notes.md           <-- log of notes
+├── 📄 pyproject.toml
+├── 📄 dev_environment.yml
+├── 📄 environment.yml
+├── 📄 requirements_dev.txt
+├── 📄 requirements.txt
+├── 🐍 noxfile.py
+├── 🐍 setup.py
+└── 🐍 tasks.py           <-- invoke tasks
+```
+
+The `cellpy` source code is structured as follows:
+
+```
+📂 cellpy\cellpy
+├── 📂 internals
+│   ├── 🐍 __init__.py
+│   └── 🐍 core.py
+├── 📂 parameters
+│   ├── 📂 legacy
+│   │   ├── 🐍 __init__.py
+│   │   └── 🐍 update_headers.py
+│   ├── 📄 .cellpy_prms_default.conf
+│   ├── 🐍 __init__.py
+│   ├── 🐍 internal_settings.py
+│   ├── 🐍 prmreader.py
+│   └── 🐍 prms.py
+├── 📂 readers
+│   ├── 📂 instruments
+│   │   ├── 📂 .benchmarks
+│   │   ├── 📂 configurations
+│   │   │   ├── 🐍 __init__.py
+│   │   │   ├── 🐍 maccor_txt_four.py
+│   │   │   ├── 🐍 maccor_txt_one.py
+│   │   │   ├── 🐍 maccor_txt_three.py
+│   │   │   ├── 🐍 maccor_txt_two.py
+│   │   │   ├── 🐍 maccor_txt_zero.py
+│   │   │   └── 🐍 neware_txt_zero.py
+│   │   ├── 📂 loader_specific_modules
+│   │   │   ├── 🐍 __init__.py
+│   │   │   └── 🐍 biologic_file_format.py
+│   │   ├── 📂 processors
+│   │   │   ├── 🐍 __init__.py
+│   │   │   ├── 🐍 post_processors.py
+│   │   │   └── 🐍 pre_processors.py
+│   │   ├── 🐍 __init__.py
+│   │   ├── 🐍 arbin_res.py
[…]
