+++ docs/contributing/index.md
@@ -2,4 +2,7 @@
 
 ```{toctree}
 :maxdepth: 2
+
+contributing.md
+developers_guide/index.md
 ```

[…]
