+++ cellpy/utils/collectors.py
@@ -168,6 +168,7 @@ class BatchCollector:
         **kwargs,
     ):
         """Update both the collected data and the plot(s).
+
         Args:
             b (cellpy.utils.Batch): the batch object.
             data_collector (callable): method that collects the data.
@@ -183,7 +184,9 @@ class BatchCollector:
             plotter_arguments (dict): keyword arguments sent to the plotter.
             update_name (bool): update the name (using automatic name generation) based on new settings.
             **kwargs: set Collector attributes.
+
         """
+
         self.b = b
         self.data_collector = data_collector
         self.plotter = plotter
@@ -352,6 +355,8 @@ class BatchCollector:
         self.toolbar = kwargs.get("toolbar", True)
 
     def generate_name(self):
+        """Generate a name for the collection."""
+
         names = ["collector", self.collector_name]
         if self.nick:
             names.insert(0, self.nick)
@@ -430,9 +435,11 @@ class BatchCollector:
         self, data_collector_arguments: dict = None, plotter_arguments: dict = None
     ):
         """Reset the arguments to the defaults.
+
         Args:
             data_collector_arguments (dict): optional additional keyword arguments for the data collector.
             plotter_arguments (dict): optional additional keyword arguments for the plotter.
+
         """
         self._data_collector_arguments = self._default_data_collector_arguments.copy()
         self._plotter_arguments = self._default_plotter_arguments.copy()
@@ -440,6 +447,7 @@ class BatchCollector:
 
     def parse_units(self, **kwargs):
         """Look through your cellpy objects and search for units."""
+
         b = self.b
         c_units = []
         r_units = []
@@ -473,6 +481,7 @@ class BatchCollector:
         update_plot: bool = True,
     ):
         """Update both the collected data and the plot(s).
+
         Args:
             data_collector_arguments (dict): keyword arguments sent to the data collector.
             plotter_arguments (dict): keyword arguments sent to the plotter.
@@ -480,6 +489,7 @@ class BatchCollector:
             update_data (bool): update the data before updating the plot even if data has been collected before.
             update_name (bool): update the name (using automatic name generation) based on new settings.
             update_plot (bool): update the plot.
+
         """
         if reset:
             self.reset_arguments(data_collector_arguments, plotter_arguments)
@@ -517,11 +527,12 @@ class BatchCollector:
     def show(self, **kwargs):
         """Show the figure.
 
-        Note that show returns the `figure` object and  if the `backend` used
-        does not provide automatic rendering in the editor / running environment you
-        are using, you might have to issue the rendering yourself. For example, if you
-        are using `plotly` and running it as a script in a typical command shell,
-        you will have to issue `.show()` on the returned `figure` object.
+        Notes:
+            The ``show`` method returns the ``figure`` object and  if the ``backend`` used
+            does not provide automatic rendering in the editor / running environment you
+            are using, you might have to issue the rendering yourself. For example, if you
+            are using `plotly` and running it as a script in a typical command shell,
+            you will have to issue ``.show()`` on the returned ``figure`` object.
 
         Args:
             **kwargs: sent to the plotter.
@@ -529,6 +540,7 @@ class BatchCollector:
         Returns:
             Figure object
         """
+
         if not self._figure_valid():
             return
 
@@ -548,6 +560,13 @@ class BatchCollector:
         return self.data
 
     def to_csv(self, serial_number=None):
+        """Save to csv file.
+
+        Args:
+            serial_number (int): serial number to append to the filename.
+
+        """
+
         filename = self._output_path(serial_number)
         filename = filename.with_suffix(".csv")
         if self.csv_layout != "long":
[…]
