### 2024-02-28 f4fb720
fix developers guide for making conda package
