### 2024-02-28 9e2531a
fix developers guide for documentation and pypi packaging
