+++ docs/contributing/developers_guide/dev_cellpy_packaging_pypi.md
@@ -20,10 +20,10 @@ The following instructions are copied from the PyPI website (<https://pypi.org/h
 Please see the PyPI website for the latest instructions.
 
 Get the API token:
-: - Verify your email address (check your account settings).
+  - Verify your email address (check your account settings).
   - In your account settings, go to the API tokens section and select "Add API token"
 
 To use an API token:
-: - Set your username to "\_\_token\_\_"
+  - Set your username to "\_\_token\_\_"
   - Set your password to the token value, including the pypi- prefix
 :::
