+++ docs/contributing/developers_guide/dev_cellpy_folder_structure.md
@@ -1,4 +1,4 @@
-# Structure of the `cellpy` package
+# Structure of the cellpy package
 
 ## Folder structure
 
