+++ docs/contributing/index.md
@@ -1,8 +1,6 @@
 # Contributing guide
 
 ```{toctree}
-:maxdepth: 1
-
 contributing.md
 developers_guide/index.md
 ```

[…]
