### 2024-02-29 d0aecee
re-run and clean up notebooks in doc and fix bug in cycle_info_plot (needed sorting after merging with step-table)
