### 2024-02-29 5edf46e
update example on dqdv in the basic_usage.md file and remove link to legacy documentation
