+++ docs/other/legacy.md
@@ -1,10 +1,0 @@
-# Legacy Documentation
-
-The legacy documentation for cellpy still exists.
-
-It is located in the `_old_docs` folder and can be found [here](https://github.com/jepegit/cellpy/tree/a09c46f05a7acdd6684e5b7eb5820e58f747cd12/_old_docs).
-
-It is not maintained and will be removed in the future.
-It is kept here for the time being while we are working with the new documentation (some of it can be
-reused, we hope).
-
