### 2024-02-29 97c3335
update example on dqdv in the basic_usage.md file and remove link to legacy documentation
