### 2024-03-03 abade4e
add method to get instrument loader information and include that in the notebook
