+++ cellpy/readers/instruments/neware_txt.py
@@ -52,6 +52,7 @@ from cellpy.readers.instruments.base import TxtLoader
 
 
 SUPPORTED_MODELS = {
+    "ONE": "neware_txt_zero",
     "UIO": "neware_txt_zero",
 }
 
