+++ examples/data/maccor_two.txt
@@ -1,7 +1,7 @@
-Name:	CW-SIMBA-FC-CT-NAPF6-013-2.081
+Name:	KIT-SIMBA-013-2.081
 Channel:	81
 Tester	Maccor #4
-Procedure	CW_MZ_SIMBA_CT.000
+Procedure	SIMBA.000
 Description	SIMBA-Capacity_test
 Started	06/05/2021 11:23:46
 SCap	0,017
@@ -13,217 +13,217 @@ DisCh. max. I	5,000 A
 
 Rec	Cycle P	Cycle C	Step	TestTime	StepTime	Cap. [Ah]	Ener. [Wh]	Current [A]	Voltage [V]	Md	ES	DPT Time	VAR1	VAR2	VAR3	VAR4	VAR5	VAR6	VAR7	VAR8	VAR9	VAR10	VAR11	VAR12	VAR13	VAR14	VAR15
 
-1	0	1	1	0,000000	0,000000	0,00000000	0,00000000	0,00000000	0,1377	R	0	06/05/2021 11:23:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-2	0	1	1	1,000000	1,000000	0,00000000	0,00000000	0,00000000	0,1385	R	1	06/05/2021 11:24:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-3	0	1	1	2,000000	2,000000	0,00000000	0,00000000	0,00000000	0,1394	R	1	06/05/2021 11:25:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-4	0	1	1	3,000000	3,000000	0,00000000	0,00000000	0,00000000	0,1403	R	1	06/05/2021 11:26:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-5	0	1	1	4,000000	4,000000	0,00000000	0,00000000	0,00000000	0,1409	R	1	06/05/2021 11:27:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-6	0	1	1	5,000000	5,000000	0,00000000	0,00000000	0,00000000	0,1416	R	1	06/05/2021 11:28:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-7	0	1	1	6,000000	6,000000	0,00000000	0,00000000	0,00000000	0,1421	R	1	06/05/2021 11:29:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
-8	0	1	1	7,000000	7,000000	0,00000000	0,00000000	0,00000000	0,1425	R	1	06/05/2021 11:30:46	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	0,0000	
[…]
