### 2024-03-04 85c67e4
add neware to notebook and example data
