### 2024-03-05 01c8650
add doc on templates
