### 2024-03-05 eb611e2
fix so that instrument files can be provided also as pathlib.Path objects.
