### 2024-03-06 3791df8
notebook 7 ok-ish
