+++ cellpy/readers/instruments/biologics_mpr.py
@@ -1,4 +1,5 @@
 """This file contains methods for importing Bio-Logic mpr-type files"""
+
 # This is based on the work by Chris Kerr
 # (https://github.com/chatcannon/galvani/blob/master/galvani/BioLogic.py)
 import datetime
@@ -215,7 +216,9 @@ class DataLoader(BaseLoader):
         self.mpr_log = None
         self.mpr_settings = None
 
+        print(" loading mpr-data ".center(80, "-"))
         self._load_mpr_data(temp_filename, bad_steps)
+
         length_of_test = self.mpr_data.shape[0]
         logging.debug(f"length of test: {length_of_test}")
 
@@ -228,7 +231,7 @@ class DataLoader(BaseLoader):
             txt = "\nCould not find any summary (stats-file)!"
             txt += " (summary_df.empty = True)"
             txt += "\n -> issue make_summary(use_cellpy_stat_file=False)"
-            warnings.warn(txt)
+            logging.debug(txt)
 
         data.summary = summary_df
         data.raw = self.mpr_data
@@ -383,7 +386,7 @@ class DataLoader(BaseLoader):
 
         dtype_dict = OrderedDict()
         flags_dict = OrderedDict()
-
+        print(" -> column_types")
         for col in column_types:
             print(f"{col=}")
             if col in bl_flags.keys():
@@ -404,13 +407,17 @@ class DataLoader(BaseLoader):
             )
         bulk = main_data
         bulk_data = np.fromstring(bulk, dtype=dtype)
+        print(" -> bulk_data")
         print(bulk_data)
         mpr_data = pd.DataFrame(bulk_data)
-        print(mpr_data.head())
-        print(self.flags_dict)
-        print(mpr_data.columns)
-        print(mpr_data["flags"].unique())
-        print(mpr_data["flags2"].unique())
+        print(" -> as dataframe")
+        print(mpr_data.T.head(30))
+        print(f"{self.flags_dict=}")
+        print(f"{mpr_data.columns=}")
+        # print(f"{self.flags_dict=}")
+        # print(mpr_data.columns)
+        # print(mpr_data["flags"].unique())
+        # print(mpr_data["flags2"].unique())
         logging.debug(mpr_data.columns)
         self.logger.debug(mpr_data.head())
 
@@ -474,11 +481,16 @@ class DataLoader(BaseLoader):
         # (while for arbindata it is stored as str)
 
     def _generate_step_index(self):
-        # TODO: @jepe - check and optionally fix me
+        # TODO: @jepe - fix me
+        # Difficult to get the steps right
         cellpy_header_txt = "step_index_txt"
-        biologics_header_txt = "flags2"
-        self._rename_header(cellpy_header_txt, biologics_header_txt)
-        self.mpr_data[self.cellpy_headers[cellpy_header_txt]] += 1
+        try:
+            biologics_header_txt = "flags2"
+            self._rename_header(cellpy_header_txt, biologics_header_txt)
+            self.mpr_data[self.cellpy_headers[cellpy_header_txt]] += 1
+        except KeyError as e:
+            self.logger.info(f"Problem during conversion to cellpy-format ({e})")
+            self.mpr_data[self.cellpy_headers[cellpy_header_txt]] = 1
 
     def _generate_step_time(self):
         k = self.cellpy_headers["step_time_txt"]
@@ -542,7 +554,7 @@ class DataLoader(BaseLoader):
         mpr_settings = self.mpr_settings
         # TODO: @jepe - finalise making summary of mpr-files
[…]
