### 2024-03-06 7884128
fix bug in concat_summaries
