+++ cellpy/readers/instruments/biologics_mpr.py
@@ -62,10 +62,9 @@ MINIMUM_SELECTION = [
 
 def _read_modules(fileobj):
     module_magic = fileobj.read(len(b"MODULE"))
+    logging.debug(f"-")
     hdr_bytes = fileobj.read(hdr_dtype.itemsize)
-    hdr = np.fromstring(
-        hdr_bytes, dtype=hdr_dtype, count=1
-    )  # TODO: change to frombuffer
+    hdr = np.fromstring(hdr_bytes, dtype=hdr_dtype, count=1)
     hdr_dict = dict(((n, hdr[n][0]) for n in hdr_dtype.names))
     hdr_dict["offset"] = fileobj.tell()
     hdr_dict["data"] = fileobj.read(hdr_dict["length"])
@@ -242,14 +241,7 @@ class DataLoader(BaseLoader):
 
     def _parse_mpr_log_data(self):
         if not self.mpr_log:
-            print("no mpr_log")
-            # self.mpr_log["end_date"] = self.mpr_settings["start_date"]
-            # self.mpr_log["length2"] = 0
-            # self.mpr_log["end2"] = 0
-            # self.mpr_log["offset2"] = 0
-            # self.mpr_log["version2"] = 0
-            # self.mpr_log["data"] = None
-            # self.mpr_log["file"] = None
+            logging.debug("OBS! No mpr_log!")
 
         for value in bl_log_pos_dtype:
             # print(value)
@@ -279,11 +271,11 @@ class DataLoader(BaseLoader):
         mpr_settings["offset"] = settings_mod["offset"]
         mpr_settings["version"] = settings_mod["version"]
         mpr_settings["data"] = settings_mod["data"]
-        print(f"mpr_settings: {mpr_settings}")
         self.mpr_settings = mpr_settings
-        return None
 
     def _get_flag(self, flag_name):
+        # TODO: next, find out what where the Ns changes are and Ns
+        #  seems to be a method there to get the cycle number as well
         print(f"flag_name: {flag_name}")
         if flag_name in self.flags_dict:
             mask, dtype = self.flags_dict[flag_name]
@@ -313,9 +305,9 @@ class DataLoader(BaseLoader):
         stats_info = os.stat(filename)
         mpr_modules = []
 
-        mpr_log = None
-        mpr_data = None
-        mpr_settings = None
+        # VMP LOG: log module
+        # VMP data: data module
+        # VMP Set: settings module
         with open(filename, mode="rb") as file_obj:
             label = file_obj.read(len(mpr_label))
             self.logger.debug(f"label: {label}")
@@ -332,7 +324,7 @@ class DataLoader(BaseLoader):
                     self.logger.info(txt)
                     break
 
-        # ------------- set -----------------------------------
+        # ------------- settings -------------------------------
         settings_mod = None
         for m in mpr_modules:
             if m["shortname"].strip().decode() == "VMP Set":
@@ -357,7 +349,7 @@ class DataLoader(BaseLoader):
         n_columns = np.fromstring(data_module["data"][4:5], dtype="u1")
         n_columns = n_columns[0]
 
-        # print(f"data (points, cols): {n_data_points}, {n_columns}")
+        logging.debug(f"data (points, cols): {n_data_points}, {n_columns}")
 
         if data_version == 0:
             logging.debug("data version 0")
@@ -381,45 +373,27 @@ class DataLoader(BaseLoader):
 
         whats_left = remaining_headers.strip(b"\x00").decode("utf8")
         if whats_left:
-            self.logger.debug("UPS! you have some columns left")
+            self.logger.debug(" *** UPS! you have some columns left")
             self.logger.debug(whats_left)
 
         dtype_dict = OrderedDict()
[…]
