+++ cellpy/readers/instruments/loader_specific_modules/biologic_file_format.py
@@ -1,8 +1,8 @@
-"""File-format dicts for BioLogic"""
+"""File-format dicts for BioLogic" (building on the ``galvani`` package)"""
 
 import numpy as np
 
-__version__ = "0.1"
+__version__ = "0.2"
 
 mpr_label = b"BIO-LOGIC MODULAR FILE\x1a                         \x00\x00\x00\x00"
 
@@ -16,6 +16,29 @@ hdr_dtype = np.dtype(
     ]
 )
 
+# # from galvani 2024:
+# VMPmodule_hdr_v2 = np.dtype(
+#     [
+#         ("shortname", "S10"),
+#         ("longname", "S25"),
+#         ("max length", "<u4"),
+#         ("length", "<u4"),
+#         ("version", "<u4"),
+#         ("unknown2", "<u4"),  # 10 for set, log and loop, 11 for data
+#         ("date", "S8"),
+#     ]
+# )
+# # These column IDs define flags which are all stored packed in a single byte
+# # The values in the map are (name, bitmask, dtype)
+# VMPdata_colID_flag_map = {
+#     1: ("mode", 0x03, np.uint8),
+#     2: ("ox/red", 0x04, np.bool_),
+#     3: ("error", 0x08, np.bool_),
+#     21: ("control changes", 0x10, np.bool_),
+#     31: ("Ns changes", 0x20, np.bool_),
+#     65: ("counter inc.", 0x80, np.bool_),
+# }
+
 #
 bl_log_pos_dtype = {
     ("User", 139, 143, "a4"),
@@ -27,17 +50,582 @@ bl_log_pos_dtype = {
 }
 
 bl_flags = {
-    1: ["mode", (2, np.uint8)],
-    2: ["ox/red", (4, np.bool_)],
-    3: ["error", (8, np.bool_)],
-    21: ["control changes", (16, np.bool_)],
-    31: ["Ns changes", (32, np.bool_)],
-    65: ["counter inc.", (128, np.bool_)],
-    131: ["Ns", (256, np.bool_)],  # step number
+    1: ["mode", (0x03, np.uint8)],
+    2: ["ox_red", (0x04, np.bool_)],
+    3: ["error", (0x08, np.bool_)],
+    21: ["control_changes", (0x10, np.bool_)],
+    31: ["Ns_changes", (0x20, np.bool_)],
+    65: ["counter_inc", (0x80, np.bool_)],
+}
+
+bl_dtypes = {
+    1: ["u1", "flags", "int", "mode", ""],
+    2: ["u1", "flags", "int", "ox/red", ""],
+    3: ["u1", "flags", "int", "error", ""],
+    4: ["<f8", "time", "s", "time/s", "test_time_txt"],
+    5: ["<f4", "control", "V/mA", "control/V/mA", ""],  # Updated dtype from <f8 to <f4
+    6: ["<f4", "Ewe", "V", "Ewe/V", "voltage_txt"],
+    7: ["<f8", "dQ", "mAh", "dq/mA.h", ""],
+    8: ["<f4", "I", "mA", "I/mA", "current_txt"],
+    9: ["<f4", "Ece", "", "Ece/V", "ref_voltage_txt"],
+    10: ["<f4", "NN_10", "", "", ""],
+    11: ["<f8", "I", "mA", "<I>/mA", ""],
+    12: ["<f4", "NN_12", "", "", ""],
[…]
