+++ cellpy/readers/instruments/biologics_mpr.py
@@ -106,8 +106,8 @@ class DataLoader(BaseLoader):
         """
 
         raw_units = dict()
-        raw_units["current"] = "A"
-        raw_units["charge"] = "Ah"
+        raw_units["current"] = "mA"
+        raw_units["charge"] = "mAh"
         raw_units["mass"] = "g"
         raw_units["voltage"] = "V"
         return raw_units
@@ -178,8 +178,8 @@ class DataLoader(BaseLoader):
         Returns:
             new test
         """
-        print("bad steps: %s" % bad_steps)
-        print(f"kwargs: {kwargs}")
+        # print("bad steps: %s" % bad_steps)
+        # print(f"kwargs: {kwargs}")
         # self.name = file_name
 
         # creating temporary file and connection
@@ -215,7 +215,7 @@ class DataLoader(BaseLoader):
         self.mpr_log = None
         self.mpr_settings = None
 
-        print(" loading mpr-data ".center(80, "-"))
+        # print(" loading mpr-data ".center(80, "-"))
         self._load_mpr_data(temp_filename, bad_steps)
 
         length_of_test = self.mpr_data.shape[0]
@@ -276,18 +276,18 @@ class DataLoader(BaseLoader):
     def _get_flag(self, flag_name):
         # TODO: next, find out what where the Ns changes are and Ns
         #  seems to be a method there to get the cycle number as well
-        print(f"flag_name: {flag_name}")
+        # print(f"flag_name: {flag_name}")
         if flag_name in self.flags_dict:
             mask, dtype = self.flags_dict[flag_name]
-            print(f"flag: {flag_name}, mask: {mask}, dtype: {dtype}")
+            # print(f"flag: {flag_name}, mask: {mask}, dtype: {dtype}")
             bin_str = f"{mask:010b}"
-            print(f"bin_str: {bin_str}")
-            print([int(x) for x in bin_str])
-            print(self.mpr_data["flags"])
+            # print(f"bin_str: {bin_str}")
+            # print([int(x) for x in bin_str])
+            # print(self.mpr_data["flags"])
             stuff = np.array(
                 self.mpr_data["flags"] & mask, dtype=dtype
             )  # need to fix this!
-            print(f"stuff: {stuff}")
+            # print(f"stuff: {stuff}")
             return np.array(
                 self.mpr_data["flags"] & mask, dtype=dtype
             )  # need to fix this!
@@ -376,9 +376,12 @@ class DataLoader(BaseLoader):
             self.logger.debug(" *** UPS! you have some columns left")
             self.logger.debug(whats_left)
 
+        self.cols = []
         dtype_dict = OrderedDict()
         flags_dict = OrderedDict()
         for col in column_types:
+            # print(f"col: {col} - {bl_dtypes[col]}")
+            self.cols.append(col)
             if col in bl_flags.keys():
                 flags_dict[bl_flags[col][0]] = bl_flags[col][1]
             dtype_dict[bl_dtypes[col][1]] = bl_dtypes[col][0]
@@ -418,10 +421,28 @@ class DataLoader(BaseLoader):
         self.mpr_log = mpr_log
         self._parse_mpr_log_data()
         self.mpr_data = mpr_data
-        self.mpr_data.to_csv(
-            r"C:\scripting\cellpy_dev_resources\dev_data\biologic\out\intermediate.csv",
-            sep=";",
-        )
+        self._unpack_flags()
+
+        # # temporary export while developing
+        # temp = self.mpr_data.copy()
+        #
+        # temp.to_csv(
+        #     r"C:\scripting\cellpy_dev_resources\dev_data\biologic\out\intermediate.csv",
[…]
