### 2024-03-11 jepegit/cellpy#301 @MaximevdHeijden
Running example scripts (installation issues)
Dear cellpy team,

I still encounter problems when running an example script (01_loading_data).
I get the following error: "NoSuchModuleError: Can't load plugin: sqlalchemy.dialects:access.pyodbc", which is related to the odbc driver installation. I checked your documentation again and you do refer to this (for Windows users) by provinding a link to the website to install the driver. However the weblink does not work.

The example also does not work with Linux as I get the following error: InterfaceError: (pyodbc.InterfaceError) ('IM002', '[IM002] [Microsoft][ODBC Driver Manager] Data source name not found and no default driver specified (0) (SQLDriverConnect)')
(Background on this error at: https://sqlalche.me/e/20/rvf5)"". I did install the mbdtools as referred to in the documentation and I should have the correct version of sqlalchemy as I installed the requirements.txt file.

Can you please fix this and try to install cellpy and run it on another machine? Thanks!

Can you also still add jupyter notebook to the requirements.txt file (and refer to it in the installation)?

This issue is related to https://github.com/openjournals/joss-reviews/issues/6236
