### 2024-03-11 jepegit/cellpy#301 @MaximevdHeijden
I realized that my python and drives dont have the same bit on that PC so I tried the installation on another PC as well (Windows).
There it seemed to run the script AFTER specifically installing the requirements.txt file. Can you add this to the installation please?
