### 2024-03-11 jepegit/cellpy#301 @MaximevdHeijden
Regarding the same example script, "tables" needs to have a specific version (at least 3.7). please add this to the requirements.txt
