### 2024-03-11 jepegit/cellpy#301 @MaximevdHeijden
03_capacity_vs_voltage: required the installation of the kaleido package
