### 2024-03-11 jepegit/cellpy#301 @MaximevdHeijden
04_incremental_capacity_analysis: the ica.dqdv(c) does not run, I get the following error: "TypeError: dqdv() missing 1 required positional argument: 'capacity'"
