### 2024-03-11 jepegit/cellpy#301 @MaximevdHeijden
Please fix the above issuesand rerun all examples scripts. Once you have addressed the issues, the software is ready to use as far as I'm concerned!
