### 2024-03-13 cba2320
update installation notes adressing #301
