### 2024-03-13 jepegit/cellpy#301 @juliawind
link to access drivers fixed and reference to requirements.txt added in installation notes.
(issue with dqdv is fixed the master branch)
