+++ .github/workflows/check_conda_forge.yml
@@ -31,3 +31,8 @@ jobs:
           conda create -n cellpy_conda
           conda activate cellpy_conda
           conda install -c conda-forge cellpy
+
+      - name: check cellpy
+        run: |
+          cellpy setup
+          
