+++ .github/workflows/check_conda_forge.yml
@@ -35,5 +35,5 @@ jobs:
       - name: check cellpy
         run: |
           conda activate cellpy_conda
-          cellpy setup
+          cellpy setup -s
           
