### 2024-03-15 jepegit/cellpy#301 @jepegit
> Dear cellpy team,
>
> I still encounter problems when running an example script (01_loading_data). I get the following error: "NoSuchModuleError: Can't load plugin: sqlalchemy.dialects:access.pyodbc", which is related to the odbc driver installation. I checked your documentation again and you do refer to this (for Windows users) by provinding a link to the website to install the driver. However the weblink does not work.
>

Updated the link :-)

> The example also does not work with Linux as I get the following error: InterfaceError: (pyodbc.InterfaceError) ('IM002', '[IM002] [Microsoft][ODBC Driver Manager] Data source name not found and no default driver specified (0) (SQLDriverConnect)') (Background on this error at: https://sqlalche.me/e/20/rvf5)"". I did install the mbdtools as referred to in the documentation and I should have the correct version of sqlalchemy as I installed the requirements.txt file.

I forgot to make a conda package of the latest `cellpy` version. Now it is made.
Tried it on several linux distros (ubuntu and RHEL) without being able to reproduce the error. The intended behavior of `cellpy` is to check if it is on windows or posix first by itself and then only use `sqlalchemy `and `pyodbc` if it is on windows, while using `mdbtools` and `pandas.read_csv`, when not on windows.  Could be that the error was fixed "unintentionally" while fixing another error so that it works on the newest version while not on the older one...

>
> Can you please fix this and try to install cellpy and run it on another machine? Thanks!

So, not sure if I have fixed it, but at least it runs on the ubuntu and RHEL linux distros I have access to.

> Can you also still add jupyter notebook to the requirements.txt file (and refer to it in the installation)?

Yes, added `jupyterlab` and jupyter `notebook` to requirements.txt

> This issue is related to [openjournals/joss-reviews#6236](https://github.com/openjournals/joss-reviews/issues/6236)
