### 2024-03-15 jepegit/cellpy#301 @jepegit
> 04_incremental_capacity_analysis: the ica.dqdv(c) does not run, I get the following error: "TypeError: dqdv() missing 1 required positional argument: 'capacity'"

This should now have been fixed in v1.0.1b5 for both pip and conda
