### 2024-03-15 jepegit/cellpy#301 @jepegit
> Regarding the same example script, "tables" needs to have a specific version (at least 3.7). please add this to the requirements.txt

Added to requirements.txt
