### 2024-03-15 jepegit/cellpy#301 @jepegit
> I realized that my python and drives dont have the same bit on that PC so I tried the installation on another PC as well (Windows).
> There it seemed to run the script after specifically installing the requirements.txt file. Can you add this to the installation please?

Added "hint" about installing requirements using the "requirements.txt" file.
Added link to note about need of same bit on driver and python there as well.
