### 2024-03-15 jepegit/cellpy#301 @jepegit
> 02_Initial_data_inspection: somehow it does not recognize the .summary_plot function for me, although its in the plotutils script.

Runs well now
