### 2024-03-15 jepegit/cellpy#301 @jepegit
> 03_capacity_vs_voltage: required the installation of the kaleido package (pip install -U kaleido)

Now a part of the conda package as well and referenced in the "installation guide" (under "Dependencies")
