### 2024-03-15 jepegit/cellpy#301 @jepegit
> Please fix the above issuesand rerun all examples scripts. Once you have addressed the issues, the software is ready to use as far as I'm concerned!

All fixed and all examples scripts were rerun also on linux without crashing :-)
