+++ requirements.txt
@@ -7,7 +7,7 @@ click>=6.6
 lmfit>=0.9.7
 scipy>=0.19.0
 pandas<=2.1
-tables>=3.5.1
+tables>=3.7
 setuptools>=35.0.1
 ruamel.yaml
 pyodbc>=4.0.16
@@ -23,6 +23,7 @@ fabric
 dateparser
 ipykernel
 jupyterlab
+notebook
 rich
 requests
 plotly
