### 2024-03-15 jepegit/cellpy#278 @ma-sadeghi
> Hi @juliawind, thanks for addressing the issues. I just found another minor revision, but I won't clutter your issue tracker with another issue, so I'll add it here:
>
> In the statement of need section, you've cited `galvani`, `neware_reader`, and `galv`, all of which are limited in their scope (can only process certain formats, etc.). Then, you've cited BEEP:
>
> "BEEP (Battery Evaluation and Early Prediction [@beep]) provides a structured interface for collecting and processing battery test data and exports to text format."
>
> but haven't provided a critiuqe of why BEEP is not good. In fact, the wording is a bit confusing because it implies that BEEP is a universal tool, as opposed to the others. It'd be ideal if you could provide a critique of BEEP so that `cellpy` is differentiated. Thanks!

This still seems to not have been addressed.
