### 2024-03-15 jepegit/cellpy#277 @ma-sadeghi
Let's keep this open since it's due soon, but I'll approve it for JOSS review.
