### 2024-03-15 #6236 @ma-sadeghi
- There's only one outstanding issue: https://github.com/jepegit/cellpy/issues/278.
- CellPy team needs some time to address https://github.com/jepegit/cellpy/issues/277, but it's a non-blocker for this review.
