### 2024-03-17 aefb4ec
create method 'collect' in collectors.py
