### 2024-03-17 a1adf37
seaborn plotting for summary collector v0
