+++ cellpy/utils/collectors.py
@@ -40,6 +40,13 @@ try:
 except ImportError:
     print("WARNING: seaborn not installed")
 
+try:
+    import matplotlib.pyplot as plt
+
+    supported_backends.append("matplotlib")
+except ImportError:
+    print("WARNING: matplotlib not installed")
+
 
 DEFAULT_CYCLES = [1, 10, 20]
 CELLPY_MINIMUM_VERSION = "1.0.0"
@@ -363,13 +370,16 @@ class BatchCollector:
         name = "_".join(names)
         return name
 
-    def render(self):
+    def render(self, **kwargs):
+        """Render the figure."""
+
+        kwargs = {**self.plotter_arguments, **kwargs}
         self.figure = self.plotter(
             self.data,
             backend=self.backend,
             journal=self.b.journal,
             units=self.units,
-            **self.plotter_arguments,
+            **kwargs,
         )
 
     def _parse_elevated_arguments(
@@ -532,7 +542,13 @@ class BatchCollector:
     def show(self, **kwargs):
         """Show the figure.
 
-        Notes:
+        Note:
+            If you are working in a Jupyter notebook and using a "matplotlib"-type backend,
+            the figure will most likely be shown automatically when the figure is rendered.
+            You should not need to use this method in that case, unless you want to
+            show the figure one extra time.
+
+        Note:
             The ``show`` method returns the ``figure`` object and  if the ``backend`` used
             does not provide automatic rendering in the editor / running environment you
             are using, you might have to issue the rendering yourself. For example, if you
@@ -549,6 +565,8 @@ class BatchCollector:
         if not self._figure_valid():
             return
 
+        skip_render_for_seaborn = kwargs.pop("skip_render_for_seaborn", False)
+
         print(f"figure name: {self.name}")
         if kwargs:
             logging.info(f"updating figure with {kwargs}")
@@ -556,7 +574,14 @@ class BatchCollector:
             self.render()
 
         if self.backend == "seaborn":
-            return self.figure.figure
+            if not skip_render_for_seaborn:
+                return self.figure.figure
+            print(
+                "WARNING: skipping rendering for seaborn, assuming it is already rendered during the plotter call"
+            )
+            print(
+                "WARNING: if you want to show the figure, provide `skip_render_for_seaborn=False` as keyword argument"
+            )
         else:
             return self.figure
 
@@ -1570,7 +1595,6 @@ def sequence_plotter(
 
     plotly_arguments = dict()
     seaborn_arguments = dict()
-    seaborn_post_plot_operations = dict()
 
     if method == "film":
         labels = {
@@ -1593,14 +1617,14 @@ def sequence_plotter(
         }
         plotly_arguments = dict(x=x, y=y, labels=labels, markers=markers)
         seaborn_arguments = dict(x=x, y=y, markers=markers)
-        seaborn_post_plot_operations["labels"] = labels
+        seaborn_arguments["labels"] = labels
 
         if g == "variable" and len(collected_curves[g].unique()) > 1:
             plotly_arguments["facet_row"] = g
             seaborn_arguments["row"] = g
         if standard_deviation:
             plotly_arguments["error_y"] = standard_deviation
-            seaborn_post_plot_operations["error_y"] = standard_deviation
+            seaborn_arguments["error_y"] = standard_deviation
 
     else:
         labels = {
@@ -1608,6 +1632,7 @@ def sequence_plotter(
             f"{y}": f"{y_label} ({y_unit})",
         }
         plotly_arguments = dict(x=x, y=y, labels=labels, facet_col_wrap=cols)
+        seaborn_arguments = dict(x=x, y=y, labels=labels, row=g, col=subgroup)
 
     if method in ["fig_pr_cell", "film"]:
         group_cells = False
[…]
