### 2024-03-18 84f7fa7
seaborn plotting for summary collector ok and added b.save method for updating the journal and the cellpy-files
