### 2024-03-19 2ac2a09
remove version fixing for github_actions_environment.yml
