### 2024-03-19 bc43315
just a quick test of reading a maccor-file for SIMBA
