### 2024-03-20 71b2612
add option for cycles plotting by seaborn and select specific (unit) in collectors.py
