+++ cellpy/utils/collectors.py
@@ -47,6 +47,7 @@ CELLPY_MINIMUM_VERSION = "1.0.0"
 PLOTLY_BASE_TEMPLATE = "seaborn"
 IMAGE_TO_FILE_TIMEOUT = 30
 HDF_KEY = "collected_data"
+MAX_POINTS_SEABORN_FACET_GRID = 60_000
 
 
 def load_data(filename):
@@ -512,11 +513,16 @@ class BatchCollector:
         cellpy_units = c_units[0]
         self.units = dict(raw_units=raw_units, cellpy_units=cellpy_units)
 
-    def collect(self, **kwargs):
+    def collect(self, *args, **kwargs):
         """Collect data."""
         self.data = self.data_collector(
             self.b, **self.data_collector_arguments, **kwargs
         )
+        self.post_collect(*args, **kwargs)
+
+    def post_collect(self, *args, **kwargs):
+        """Perform post-operation after collecting the data"""
+        ...
 
     def update(
         self,
@@ -1115,8 +1121,8 @@ class BatchCyclesCollector(BatchCollector):
                 - "forth" - discharge (or charge) continues along x-axis.
                 - "forth-and-forth" - discharge (or charge) also starts at 0.
 
-            data_collector_arguments (dict) - arguments transferred to the plotter
-            plotter_arguments (dict) - arguments transferred to the plotter
+            data_collector_arguments (dict) - arguments transferred to the to `CellpyCell.get_cap`.
+            plotter_arguments (dict) - arguments transferred to `cycles_plotter`.
             cycles (list): select these cycles (elevated data collector argument).
             max_cycle (int): drop all cycles above this value (elevated data collector argument).
             rate (float): filter on rate (C-rate) (elevated data collector argument).
@@ -1132,7 +1138,7 @@ class BatchCyclesCollector(BatchCollector):
                 The dictionary must have the cell labels as given in the `journal.pages` index and new label as values.
                 Similarly, if it is a function it takes the cell label as input and returns the new label.
                 Remark! No check are performed to ensure that the new cell labels are unique.
-            cycles_to_plot (int): plot points if True (elevated plotter argument).
+            cycles_to_plot (int): plot only these cycles (elevated plotter argument).
             width (float): width of plot (elevated plotter argument).
             legend_position (str): position of the legend (elevated plotter argument).
             show_legend (bool): set to False if you don't want to show legend (elevated plotter argument).
@@ -1179,6 +1185,26 @@ class BatchCyclesCollector(BatchCollector):
             *args,
             **kwargs,
         )
+
+    def post_collect(self, *args, **kwargs):
+        """Update the x-unit after collecting the data in case the mode has been set."""
+
+        logging.debug("updating the x-unit")
+        if m := self.data_collector_arguments.get("mode"):
+            if m == "gravimetric":
+                self.plotter_arguments["x_unit"] = (
+                    f'{self.units["cellpy_units"].charge}/{self.units["cellpy_units"].gravimetric}'
+                )
+            elif m == "areal":
+                self.plotter_arguments["x_unit"] = (
+                    f'{self.units["cellpy_units"].charge}/{self.units["cellpy_units"].areal}'
+                )
+            elif m == "volumetric":
+                self.plotter_arguments["x_unit"] = (
+                    f'{self.units["cellpy_units"].charge}/{self.units["cellpy_units"].volumetric}'
+                )
+            elif m == "absolute":
+                self.plotter_arguments["x_unit"] = self.units["cellpy_units"].charge
 
[…]
