### 2024-03-20 32decf8
automatically drop duplicates and NaN in serial number (pk) for excel db
