### 2024-03-22 jepegit/cellpy#278 @jepegit
Wrote:

"""In addition, the Battery Evaluation and Early Prediction (BEEP) library [@beep] provides an interface
for parsing text-based data from several instruments. However, the target audience for this library seems to be
machine learning researchers, and the library appears at its current state not to be ideal for in-depth data handling
and analysis by battery researchers."""

@juliawind: Sounds OK? Or can it be interpreted as if the library is not a good library? What I tried to convey is that it seems to be not very easy to use if you just want to look and play with single data-sets, but is more useful if you have a large amount of data that you want to extract features from and train using scikit-learn.
