### 2024-03-22 #6236 @TomTranter
Re-installed following updates and ran examples. Works for me on windows. In general I find the package a little non-intuitive to use but the classes and functions are documented well enough to figure it out. My review is finished. Thanks to the authors for making this useful package open source
