### 2024-03-22 jepegit/cellpy#278 @juliawind
@jepegit yes, sounds good! Could rephrase a bit to make it sound more positive:
"""... However, the target audience for BEEP seems to be
machine learning researchers, focusing on handling and preparing large datasets, while cellpy provides tools for more in-depth data handling and analysis for battery researchers. """
