### 2024-03-23 jepegit/cellpy#278 @jepegit
Great!
