### 2024-03-23 jepegit/cellpy#278 @ma-sadeghi
Sounds good! Thanks for clarifying!
