### 2024-04-15 248cd8a
minor edits on paper.md
