### 2024-04-19 691ba79
complete citations for github repos
