+++ paper/paper.bib
@@ -16,27 +16,30 @@ abstract = {Battery evaluation and early prediction software package (BEEP) prov
       title={Galv},
       url = {https://github.com/Battery-Intelligence-Lab/galv},
       author={Battery-Intelligence-Lab},
-      year = {2023},
+      year = {2020},
       publisher = {GitHub},
-      journal = {GitHub repository}
-
+      journal = {GitHub repository},
+      note = {Commit: 118a8a1, 02/2024}
 }
 
 @misc{galvani,
       title={galvani},
       url = {https://github.com/echemdata/galvani},
       author={Echemdata},
-      year = {2022},
+      year = {2007},
       publisher = {GitHub},
-      journal = {GitHub repository}
+      journal = {GitHub repository},
+      note = {Commit: 3c904db, 03/2024}
 }
 
 @misc{neware_reader,
       title={Neware reader},
       author={{Beyonder} and {Corvus Energy} and {IFE}},
       url = {https://github.com/FTHuld/neware_reader},
+      year = {2020},
       publisher = {GitHub},
-      journal = {GitHub repository}
+      journal = {GitHub repository},
+      note = {Commit: 7ad6d6c, 11/2022}
 }
 
 @article{2023spi,
