### 2024-04-22 f779353
update docstring for batch.load
