### 2024-04-22 8f75908
update printing for batch.load
