### 2024-04-23 #6236 @jepegit
## Latest release:

v.1.0.1post1

## On zenodo:

![image](https://github.com/openjournals/joss-reviews/assets/16334959/fb0c0d13-d232-4ab5-81ec-b4ba56dacd89)

DOI: 10.5281/zenodo.11047004
