### 2024-04-24 #6236 @mbarzegary
@jepegit It seems that my previous fix hasn't resolved the citation rendering issue, so maybe using semicolon instead of comma to cite multiple papers can help. Please merge my new PR and then we check if the problem is resolved. Meanwhile, I move on with the rest of the steps.
