### 2024-04-24 8299142
fix citation rendering problem (#309)
