+++ paper/paper.md
@@ -61,7 +61,7 @@ for parsing text-based data from several instruments. However, the target audien
 machine learning researchers, focusing on handling and preparing large datasets, while cellpy provides
 tools for more in-depth data handling and analysis for battery researchers.
 
-`cellpy` provides powerful and versatile tools for the simple and efficient handling of battery testing data originating from different battery cell testers, all the way from data collection to data analysis and visualisation, ensuring consistency, accuracy and comparability. `cellpy` can directly parse the data from most common commercial battery testers ([Arbin](http://www.arbin.com/), Maccor, [PEC](https://www.peccorp.com/battery-testing-solutions/), Neware, BioLogic), in addition to offering full flexibility by allowing the user to provide other file format specifications (in YAML format). The data is converted into and saved in a common format, accommodating not only data from diverse testers but also thoughtfully embedding battery-specific metadata (*e.g.*, step-types, type of cell, type of chemistry, electrode properties, etc.). This makes subsequent data handling considerably easier and proves invaluable in interpreting and comparing results across tests and conditions. In addition to translating data to a common format, `cellpy` has a range of utilities for studying and analysing the data. These include methods for the extraction of key characteristics from tests, cell comparison, plotting and statistical analysis, as well as advanced tools such as incremental-capacity analysis (ICA, dQ/dV), OCV relaxation analysis and batch processing of results from many battery test [@2019and, @2020ulv, @2023hul, @2023spi].
+`cellpy` provides powerful and versatile tools for the simple and efficient handling of battery testing data originating from different battery cell testers, all the way from data collection to data analysis and visualisation, ensuring consistency, accuracy and comparability. `cellpy` can directly parse the data from most common commercial battery testers ([Arbin](http://www.arbin.com/), Maccor, [PEC](https://www.peccorp.com/battery-testing-solutions/), Neware, BioLogic), in addition to offering full flexibility by allowing the user to provide other file format specifications (in YAML format). The data is converted into and saved in a common format, accommodating not only data from diverse testers but also thoughtfully embedding battery-specific metadata (*e.g.*, step-types, type of cell, type of chemistry, electrode properties, etc.). This makes subsequent data handling considerably easier and proves invaluable in interpreting and comparing results across tests and conditions. In addition to translating data to a common format, `cellpy` has a range of utilities for studying and analysing the data. These include methods for the extraction of key characteristics from tests, cell comparison, plotting and statistical analysis, as well as advanced tools such as incremental-capacity analysis (ICA, dQ/dV), OCV relaxation analysis and batch processing of results from many battery tests [@2019and;@2020ulv;@2023hul;@2023spi].
 
 The `cellpy` library provides a valuable toolset and has been in frequent use for both everyday and advanced tasks in battery research. The ability to effortlessly import and process the data through a simple but highly flexible API allows for quick and simple comparison of different cells. At the same time, `cellpy` serves as an excellent starting point for researchers leaning towards advanced analysis: `cellpy` can automatically convert data with different units, summarize and perform statistical evaluations all the way down to the individual cycle and step level, while giving the user fine-grained control of the behaviour through setting of parameters or directly by using a more advanced, deeper API. This eases further use of the data, *e.g.*, as features for machine learning algorithms, and promotes reproducibility and traceability throughout the entire process.
 
