### 2024-04-25 236b17e
skip trying to load missing cellpyfile in batch.update
