### 2024-04-28 321a21c
comment on warnings in post_processors.py and remove some folders from linquist
