+++ cellpy/readers/instruments/processors/post_processors.py
@@ -367,6 +367,15 @@ def _state_splitter(
             ]
 
             if not charge.empty:
+                # TODO: fix this warning
+                # FutureWarning: A value is trying to be set on a copy of a DataFrame or Series through
+                # chained assignment using an inplace method.
+                # The behavior will change in pandas 3.0. This inplace method will never work because the
+                # intermediate object on which we are setting values always behaves as a copy.
+                #  For example, when doing 'df[col].method(value, inplace=True)',
+                #  try using 'df.method({col: value}, inplace=True)' or
+                #  df[col] = df[col].method(value) instead,
+                #  to perform the operation inplace on the original object.
                 raw[temp_col_name_charge].update(n_charge * charge[base_col_name])
                 if propagate:
                     charge_last_index, charge_last_val = charge.iloc[-1]
@@ -377,6 +386,15 @@ def _state_splitter(
                     ] = charge_last_val
 
             if not discharge.empty:
+                # TODO: fix this warning
+                # FutureWarning: A value is trying to be set on a copy of a DataFrame or Series through
+                # chained assignment using an inplace method.
+                # The behavior will change in pandas 3.0. This inplace method will never work because the
+                # intermediate object on which we are setting values always behaves as a copy.
+                #  For example, when doing 'df[col].method(value, inplace=True)',
+                #  try using 'df.method({col: value}, inplace=True)' or
+                #  df[col] = df[col].method(value) instead,
+                #  to perform the operation inplace on the original object.
                 raw[temp_col_name_discharge].update(
                     n_discharge * discharge[base_col_name]
                 )
