### 2024-04-30 d4879cd
add simple option for filtering for batch.plot
